% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastsimcoal.R
\name{AIC}
\alias{AIC}
\title{Compute the AIC given the likelihood in log10 scale}
\usage{
AIC(dir.in, k = NULL)
}
\arguments{
\item{dir.in}{The directory where the analysis was conducted}

\item{k}{The number of parameters in the model. If \code{NULL} is read
automatically by the function}
}
\description{
Compute the AIC given the likelihood in log10 scale (default fsc
 output) and number of parameters, which is automatically read by the 
 function if \code{k=NULL}.
}
\details{
It is important to note that this function will probably fail if there are
more than one .est within each directory as it will read these to evaluate
the number of parameters in the model. If there is more than one file, there
is no way to know which one is correct.
}
\references{
Excoffier L., Dupanloup I., Huerta-Sánchez E., Sousa V. C. and
  Foll M. (2013) Robust demographic inference from genomic and SNP data. PLoS
  genetics 9(10)
}
\author{
Carlo Pacioni (Post to \url{https://groups.google.com/d/forum/dartr})
}
