% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{bandicoot.gl}
\alias{bandicoot.gl}
\title{A genlight object created via the read.dart functions}
\format{genlight object}
\usage{
bandicoot.gl
}
\description{
This a test data set to test the validity of functions within dartR and is based on a DArT SNp data set of simulated bandicoots across Australia. It contains 96 individuals and 1000 SNPs.
}
\author{
Bernd Gruber (bugs? Post to \url{https://groups.google.com/d/forum/dartr}
}
\keyword{datasets}
