% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dartRgenlight.r
\name{cbind.dartRgenlight}
\alias{cbind.dartRgenlight}
\title{adjust cbind for dartRgenlight}
\usage{
\method{cbind}{dartRgenlight}(...)
}
\arguments{
\item{...}{list of dartRgenlight objects}
}
\description{
cbind is a bit lazy and does not take care for the metadata (so data in the other slot is lost). You can get most of the loci metadata back using gl.compliance.check.
}
