% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastsimcoal.R
\docType{data}
\name{CLR}
\alias{CLR}
\title{Compute bootstrap confidence intervals for estimated parameters with fsc}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
CLR
}
\arguments{
\item{dir.in}{The directory where the analysis was conducted}

\item{nLoci}{The number of polymorphic loci to retain}

\item{nSim}{The number of datasets that need to be simulated}

\item{par.indLoci}{The two integers value that need to be used for the number
of independent loci in the .par that it is used to run the simulations.}

\item{par.nBlocks}{The number of linkage blocks in the .par that it is used to
run the simulations.}

\item{par.data}{The string to be used in the 'per Block: data type, num loci,
rec. rate and mut rate + optional parameters' line of the .par that it is
used to run the simulations. The default value is \code{par.data="DNA 100 0
2.5e-8 0.33"}}

\item{nBoot}{The number of bootstrap replicates}

\item{conf}{The confidence level to compute the confidence intervals}

\item{boot.type}{The method to be used to compute the confidence intervals.
See \code{?boot.ci} for details. By default, percentile intervals are
computed.}
}
\value{
A list with the following elements \itemize{ \item Bootstr.stats:
 Descriptive statistics from bootstraps (Median, lower and upper limit), an
 the initial estimated parameters  \item P.Rand.less.Obs: The probability that
 a random value from the null Composite Likelihood Ratio distribution is less
 than the observed CLR \item P.Rand.gt.Obs: The probability that a random
 value from the null Composite Likelihood Ratio distribution is greater than
 the observed CLR \item Sim: The estimates from the simulated data }
}
\description{
Once fsc estimated parameter values, this function uses the *_maxL.par to
simulate \code{nSim} datasets with the model parameters. It then uses these
data to estimate the parameter values and confidence intervals are finally
returned using the R package \code{boot}.
}
\details{
It also uses the analyses fromt he simulated data to build an empirical
cumulative density function of the Composite Likelihood Ratio (CLR), to build
a statistical test for the fit of the model (See Excoffier et al 2013 for
details). That is, from the simulated data, it is possible to estimate the
probability that a randpm value from the null distribution is smaller or
greater than the observed CLR. It is important to note that the probability
values (\code{P.Rand.less.Obs} and \code{P.Rand.gt.Obs}) are constructed from
the simulated datasets, so a large enough number of simulations needs to be
run for these to be reliable. Bootstrapped percentiles are also reported (see
first item of the list returned as results), if this approach is preferred.

It is important that enough sites are simulated to ensure that sufficient
polymorphic loci are present in the simulated data. It is better to simulate
an excess of sites and retained those needed using \code{nLoci}.

Initial values when estimating parameters from simulated datasets are passed
using the .pv so that a reduced number of replicates need to be run.

For some reason, which is a mystery to me, sometimes there is a need to
'print' to screen twice to get the first element of the list to actually be
visible on the screen.
}
\references{
Excoffier L., Dupanloup I., Huerta-Sánchez E., Sousa V. C. and
 Foll M. (2013) Robust demographic inference from genomic and SNP data. PLoS
 genetics 9(10)
}
\author{
Carlo Pacioni (Post to \url{https://groups.google.com/d/forum/dartr})
}
\keyword{datasets}
