% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart2fasta.r
\name{dart2fasta}
\alias{dart2fasta}
\title{Export DArT genlight object \{adegenet\} to fastA format}
\usage{
dart2fasta(gl, outfile = "genlight.fasta")
}
\arguments{
\item{gl}{-- genlight object}

\item{outfile}{-- filename of the output fastA file [default genlight.fasta]}
}
\description{
Export DArT genlight object \{adegenet\} to fastA format
}
\author{
Bernd Gruber (glbugs@aerg.canberra.edu.au)
}

