% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{drosophila.gl}
\alias{drosophila.gl}
\title{Experimental populations of Drosophila melanogaster}
\format{
genlight object
}
\description{
Populations 25, 26 and 27 at the end of the experiment
}
\author{
Luis Mijangos (bugs? Post to 
\url{https://groups.google.com/d/forum/dartr}
}
\keyword{datasets}
