% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{drosophila_T0.gl}
\alias{drosophila_T0.gl}
\title{Experimental populations of Drosophila melanogaster 
Populations 1, 13 and 27 at the beginning of the experiment (T0)}
\format{
genlight object
}
\description{
Experimental populations of Drosophila melanogaster 
Populations 1, 13 and 27 at the beginning of the experiment (T0)
}
\author{
Luis Mijangos (bugs? Post to 
\url{https://groups.google.com/d/forum/dartr}
}
\keyword{datasets}
