% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{drosophila_T2.gl}
\alias{drosophila_T2.gl}
\title{Experimental populations of Drosophila melanogaster 
Populations 25, 26 and 27 at the end of the experiment (T2)}
\format{
genlight object
}
\description{
Experimental populations of Drosophila melanogaster 
Populations 25, 26 and 27 at the end of the experiment (T2)
}
\author{
Luis Mijangos (bugs? Post to 
\url{https://groups.google.com/d/forum/dartr}
}
\keyword{datasets}
