% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.dart.r
\name{filter.dart}
\alias{filter.dart}
\title{Filter function to facilitate analysing of dart data [deprecated use explicit filter functions gl.filter...]}
\usage{
filter.dart(...)
}
\arguments{
\item{...}{vectors of true false created via comparisons between metrics (see examples)}
}
\value{
returns the combined index of filters. this function is unusual as it has no predefined parameters. The idea is here that you can provide any type of filter (and index of true false that is used to filter the loci of your genlight object afterwards. This functions combines your filter and provides a plot on their single and combined effect. see example
}
\description{
Filter function to facilitate analysing of dart data [deprecated use explicit filter functions gl.filter...]
}
\examples{
\dontrun{
#gl is a genlight object created with the read.dart and dart2genlight functions 
index.repro <- gl@other$loc.metrics[,"RepAvg"] > 0.98
index.callrate <-  gl@other$loc.metrics[,"CallRate"] > 0.90
index.highhet <- fox.gl.keep@other$loc.metrics[,"FreqHets"] <0.75
index.comb <- filter.dart(index.repro, index.callrate, index.coverage)
}
}

