% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.sex.linked.R
\name{filter.sex.linked}
\alias{filter.sex.linked}
\title{Filters loci that are sex linked}
\usage{
filter.sex.linked(gl = LBP, system = "xy", plots = TRUE, 
parallel = TRUE)
}
\arguments{
\item{gl}{Name of the genlight object containing the SNP data. This genlight
object needs to contain the sex of the individuals. See explanation in 
details [required].}

\item{system}{String that declares the sex-determination system of the 
species: 'zw' or 'xy' [required].}

\item{plots}{Creates four output plots. See explanation in details
[default TRUE].}

\item{parallel}{Runs the function in parallel. See explanation in
details [default FALSE].}
}
\value{
A list of 6 elements and 4 plots.
}
\description{
This function identifies sex-linked and autosomal loci present in a SNP 
dataset (genlight object) using individuals with known sex. It identifies
five types of loci: w-linked or y-linked, sex-biased, z-linked or
x-linked, gametologous and autosomal.

This function produces as output a list with 6 elements and 4 plots.
}
\details{
The sex of the individuals for which sex is known with certainty has to be 
provided in the genlight object via \code{gl@other$ind.metrics$sex}.
Coding is: 'M' for male, 'F' for female. The function ignores individuals 
that are assigned anything else or nothing at all.

The creation of plots can be turn-off (plots = FALSE) in order to save a 
little bit of running time for very large datasets (>50,000 SNPs). However, 
we strongly encourage you to always inspect the output plots at least once to
make sure everything is working properly.

\strong{ Function's output }

This function creates a list of 6 elements: \itemize{
\item {$results.table > Table with statistics (columns) for each loci (rows)}
\item {$w.linked or $y.linked > Genlight object with w-linked/y-linked loci}
\item {$sex.biased > Genlight object with sex-biased scoring rate loci} 
\item {$z.linked or $x.linked > Genlight object with z-linked/x-linked loci}
\item {$gametolog > Genlight object with gametologs} 
\item {$autosomal > Genlight object with autosomal loci}
}

And 4 plots:\itemize{
\item {A plot BEFORE filtering sex-linked loci by call rate}
\item {A plot AFTER filtering sex-linked loci by call rate} 
\item {A plot BEFORE filtering sex-linked loci by heterozygosity}
\item {A plot AFTER filtering sex-linked loci by heterozygosity}
}
}
\examples{
out <- filter.sex.linked(gl = LBP, system = "xy", plots = TRUE, 
parallel = TRUE)
out$results.table
out$x.linked


}
\author{
Diana Robledo-Ruiz & Floriaan Devloo-Delva (Write to
 \email{diana.robledoruiz1@monash.edu})
}
