% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{foxes.gl}
\alias{foxes.gl}
\title{A genlight object created via the read.dart functions}
\format{genlight object}
\usage{
foxes.gl
}
\description{
This a test data set to test the validity of functions within dartR and is based on a DArT SNp data set of foxes across Australia. It contains 100 individuals and 500 SNPs.
}
\author{
Bernd Gruber \email{(bernd@aerg.canberra.edu.au}
}
\keyword{datasets}
