% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastsimcoal.r
\name{fsc.estimate}
\alias{fsc.estimate}
\title{Run fastsimcoal to estimate parameters' values}
\usage{
fsc.estimate(
  dir.in,
  n = 5e+05,
  L = 100,
  maf = TRUE,
  ncpu = 0,
  nBatches = NULL,
  fsc.cmd = "fsc2702",
  fsc.path = "path"
)
}
\arguments{
\item{dir.in}{The path where to run fsc}

\item{n}{The number of coalescent simulations to approximate the expected SFS
(-n option). This should be larger than 100,000.}

\item{L}{The number of optimization cycles (-L option). It should be >50}

\item{maf}{Whether a MAF SFS (default) or a derived SFS is provided (if
\code{FALSE})}

\item{ncpu}{The number of CPU (threads) to use in the analysis. Automatically
handle if \code{ncpu=0} (default)}

\item{nBatches}{The number of batches (-B option)}

\item{fsc.cmd}{The command to use to call fsc (that may be different
depending on the version installed)}

\item{fsc.path}{The path where fsc is installed or \code{"path"} if it is in
the PATH (that means that it can be called regardless of the working
directory)}
}
\description{
Run fastsimcoal assuming that in \code{dir.in} there are the tpl, est and obs
files properly named and formatted. It writes a fsc_run.txt file and execute
fsc in \code{dir.in}
}
\details{
If \code{ncpu=0} and \code{nBatches=NULL}, \code{nBatches} is set to 12 (fsc
default value). If \code{ncpu}>0 and \code{nBatches=NULL}, \code{nBatches} is
set to twice ]code{nspu}.
}
\references{
Excoffier L., Dupanloup I., Huerta-Sánchez E., Sousa V. C. and
  Foll M. (2013) Robust demographic inference from genomic and SNP data. PLoS
  genetics 9(10)
}
\author{
Carlo Pacioni (Post to \url{https://groups.google.com/d/forum/dartr})
}
