% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastsimcoal.R
\name{fsc.multiple.estimate}
\alias{fsc.multiple.estimate}
\title{Run fsc over multiple models}
\usage{
fsc.multiple.estimate(
  dir.in,
  n = 5e+05,
  L = 100,
  maf = TRUE,
  ncpu = 0,
  nBatches = NULL,
  fsc.cmd = "fsc2702",
  fsc.path = "path"
)
}
\arguments{
\item{dir.in}{The path where to run fsc}

\item{n}{The number of coalescent simulations to approximate the expected SFS
(-n option). This should be larger than 100,000.}

\item{L}{The number of optimization cycles (-L option). It should be >50}

\item{maf}{Whether a MAF SFS (default) or a derived SFS is provided (if
\code{FALSE})}

\item{ncpu}{The number of CPU (threads) to use in the analysis. Automatically
handle if \code{ncpu=0} (default)}

\item{nBatches}{The number of batches (-B option)}

\item{fsc.cmd}{The command to use to call fsc (that may be different
depending on the version installed)}

\item{fsc.path}{The path where fsc is installed or \code{"path"} if it is in
the PATH (that means that it can be called regardless of the working
directory)}
}
\description{
This function expects one or more folders within \code{dir.in}, where in each
directory there are all the files needed to run \code{fsc.estimate}, which is
 run iteratively
}
