% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gi.hwe.pop.r
\name{gi.hwe.pop}
\alias{gi.hwe.pop}
\title{Filter function to facilitate analysing of dart data}
\usage{
gi.hwe.pop(gi, pvalue = 0.05, plot = TRUE)
}
\arguments{
\item{gi}{a genind object (often created using the gl2gi function)}

\item{pvalue}{the p-value for the HWE test.}

\item{plot}{a switch if a barplot is wanted.}
}
\value{
tests HWE for every loci in every population (using the setupSNP function in package SNPassoc
}
\description{
Filter function to facilitate analysing of dart data
}
\examples{
\dontrun{
HWEperpop(gi=geni, pvalue = 0.05, plot = TRUE)
}
}
