% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gi2gl.R
\name{gi2gl}
\alias{gi2gl}
\title{Genind to genlight conversion}
\usage{
gi2gl(gi, parallel = TRUE)
}
\arguments{
\item{gi}{a genind object}

\item{parallel}{Switch passed to genlight constructor.}
}
\value{
A genlight object, with all slots filled.
}
\description{
Converts a genind object to genlight object.
}
\details{
Note that \code{gl2gi(gi2gl(x))} may not return the original
object due to the ambiguity in choice of the reference alleles.
}
\seealso{
\code{\link{gi2gl}}, \code{\link[adegenet]{genlight}}
}
\author{
Bernd Gruber (bugs? Post to \url{https://groups.google.com/d/forum/dartr})
}
