% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.alf.r
\name{gl.alf}
\alias{gl.alf}
\title{Allelle Frequency}
\usage{
gl.alf(gl)
}
\arguments{
\item{gl}{a genlight object}
}
\value{
A dataframe with columns
\item{\code{alf1}}{frequency of the reference allele for each loci}
\item{\code{alf2}}{frequency of the reference allele for each loci}, alf2
}
\description{
Calculates allele frequencies of the first and second allele for
each loci.
}
\author{
Bernd Gruber (bugs? Post to \url{https://groups.google.com/d/forum/dartr})
}
