% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.amova.r
\name{gl.amova}
\alias{gl.amova}
\title{Analysis of Molecular Variance}
\usage{
gl.amova(x, nperm = 100)
}
\arguments{
\item{x}{a genlight object containing the SNP genotypes and
population grouping}

\item{nperm}{number of permuations to perform for hypothesis
testing.}
}
\value{
An object of class \code{pegas::amova}.
}
\description{
Performs an AMOVA using genlight data.
}
\details{
This script performs an AMOVA based on the genetic distance matrix
from stamppNeisD() [package StAMPP] using the amova() function
from the package PEGAS for exploring within and between population
variation. For detailed information see the help pages for amova
and stamppAmova
}
\seealso{
\code{\link[pegas]{amova}}, \code{\link[StAMPP]{stamppNeisD}}
}
\author{
Bernd Gruber (bugs? Post to \url{https://groups.google.com/d/forum/dartr})
}
