% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.assign.mahalanobis.r
\name{gl.assign.mahalanobis}
\alias{gl.assign.mahalanobis}
\title{Assign an individual of unknown provenance to population based on PCA}
\usage{
gl.assign.mahalanobis(x, plevel = 1 - 0.001, unknown, verbose = NULL)
}
\arguments{
\item{x}{Name of the input genlight object [required].}

\item{plevel}{Probability level for bounding ellipses in the PCoA plot
[default 1- 0.001].}

\item{unknown}{Identity label of the focal individual whose provenance is
unknown [required].}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2,
progress log; 3, progress and results summary; 5, full report
[default 2 or as specified using gl.set.verbosity].}
}
\value{
A data frame with the results of the assignment analysis.
}
\description{
This script assigns an individual of unknown provenance to one or more target
populations based on the unknown individual's proximity to population centroids;
proximity is estimated using Mahalanobis Distance. 

The following process is followed:
\enumerate{
\item An ordination is undertaken on the populations to again yield a
series of orthogonal (independent) axes.
\item A workable subset of dimensions is chosen, normally equal to the number
 of dimensions with substantive eigenvalues.
\item The Mahalobalis Distance is calculated for the unknown against each
population and probability of membership of each population is calculated.
The assignment probabilities are listed in support of a decision.
}
}
\details{
There are three considerations to assignment. First, consider only those
populations for which the unknown has no private alleles. Private alleles are
an indication that the unknown does not belong to a target population
(provided that the sample size is adequate, say >=10). This can be evaluated
 with gl.assign.pa().

A next step is to consider the PCoA plot for populations where no private
alleles have been detected. The position of the unknown in relation to the
confidence ellipses is plotted by this script as a basis for narrowing down
the list of putative source populations. 

Note, this plot is
considering only the top two dimensions of the ordination, and so an unknown
lying outside the confidence ellipse can be interpreted as it lying outside
the confidence envelope. However, if the unknown lies inside the confidence
ellipse in two dimensions, then it may still lie outside the confidence
envelope. Thus this second step is good for eliminating populations from
consideration, but does not provide confidence in assignment.

The third step (delivered by this script) is to consider the assignment probabilities based on 
the squared Generalised Linear Distance (Mahalanobis distance) of
the unknown from the centroid for each population, then to consider the
probability associated with its quantile using the Chisquare approximation. 
In effect, this index takes into account position of the unknown in
relation to the confidence envelope in all selected dimensions of the
ordination. The larger the assignment probability, the greater the confidence in the
assignment. 

If the unknown individual is an extreme outlier, say at less than 0.001 probability
of population membership, then the associated population can be eliminated from
further consideration.

Each of these above approaches provides evidence, none are 100% definitive. They
need to be interpreted cautiously.
}
\examples{
Test run with a focal individual from the Macleay River (EmmacMaclGeor) 
x <- gl.assign.pa(testset.gl, unknown='UC_01044', nmin=10, threshold=1,verbose=3) 
x <- gl.assign.pca(x, unknown='UC_01044', plevel=0.95, verbose=3)
df <- gl.assign.mahalanobis(x, unknown='UC_01044', verbose=3)
}
\author{
Custodian: Arthur Georges --
Post to \url{https://groups.google.com/d/forum/dartr}
}
