% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.basic.stats.r
\name{gl.basic.stats}
\alias{gl.basic.stats}
\title{Calculates basic statistics for each loci (Hs, Ho, Fis etc.)}
\usage{
gl.basic.stats(gl, digits = 4)
}
\arguments{
\item{gl}{-- a genlight object}

\item{digits}{-- number of digits that should be returned}
}
\value{
several tables and lists with all basic stats. Check \code{\link[hierfstat]{basic.stats}} for details.
}
\description{
Based on function \code{\link[hierfstat]{basic.stats}}. Check ?basic.stats for help.
}
\examples{
gl.basic.stats(possums.gl)
}
\author{
Bernd Gruber (bugs? Post to \url{https://groups.google.com/d/forum/dartr})
}
