% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.check.verbosity.r
\name{gl.check.verbosity}
\alias{gl.check.verbosity}
\title{A function to check the current global verbosity}
\usage{
gl.check.verbosity(verbose = NULL)
}
\arguments{
\item{verbose}{-- user requested level of verbosity}
}
\value{
The verbosity, in variable verbose
}
\description{
The verbosity can be set in one of two ways -- (a) explicitly by the user by passing a value using the parameter
verbose= in a function, or (b) by setting the verbosity globally as part of the r environment (gl.set.verbosity).
}
\author{
Bernd Gruber (Post to \url{https://groups.google.com/d/forum/dartr})
}
