% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.diagnostics.sim.r
\name{gl.diagnostics.sim}
\alias{gl.diagnostics.sim}
\title{Function to diagnose simulations against theoretical expectations}
\usage{
gl.diagnostics.sim(
  x,
  Ne,
  Ne_fst,
  iteration = 1,
  pop_he = 1,
  pops_fst = c(1, 2),
  plot_colors = two_colors_contrast,
  plot_theme = theme_dartR(),
  save2tmp = FALSE,
  verbose = NULL
)
}
\arguments{
\item{x}{[default ]}

\item{iteration}{[default ]}

\item{pop_he}{[default ]}

\item{pops_fst}{[default ]}

\item{plot_colors}{Vector with two color names for the significant and
not-significant loci [default two_colors_contrast].}

\item{plot_theme}{User specified theme [default theme_dartR()].}

\item{save2tmp}{If TRUE, saves any ggplots and listings to the session
temporary directory (tempdir) [default FALSE].}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2,
progress log ; 3, progress and results summary; 5, full report
[default NULL, unless specified using gl.set.verbosity].}
}
\description{
Function to diagnose simulations against theoretical expectations
}
