% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.dist.ind.r
\name{gl.dist.ind.new}
\alias{gl.dist.ind.new}
\title{Calculates a distance matrix for individuals defined in a genlight object}
\usage{
gl.dist.ind.new(
  x,
  method = NULL,
  scaled = TRUE,
  plot.out = TRUE,
  plot_theme = theme_dartR(),
  plot_colors = two_colors,
  save2tmp = FALSE,
  verbose = NULL
)
}
\arguments{
\item{x}{Name of the genlight containing the SNP genotypes or presence-absence data [required].}

\item{method}{Specify distance measure [SNP: Euclidean; P/A: SMD].}

\item{scaled}{If TRUE, the distances are scaled to fall in the range [0,1] [default TRUE]}

\item{plot.out}{If TRUE, display a histogram and a boxplot of the genetic distances [TRUE].}

\item{plot_theme}{User specified theme [default theme_dartR].}

\item{plot_colors}{Vector with two color names for the borders and fill [default two_colors].}

\item{save2tmp}{If TRUE, saves any ggplots to the session temporary directory [default FALSE].}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2,
progress log ; 3, progress and results summary; 5, full report
[default 2 or as specified using gl.set.verbosity].}
}
\value{
An object of class 'matrix' or dist' giving distances between individuals
}
\description{
This script calculates various distances between individuals based on allele
 frequencies or presence-absence data
}
\details{
The distance measure for SNP genotypes can be one of:
\itemize{
 \item Euclidean Distance
 \item Scaled Euclidean Distance
 \item Simple Mismatch Distance
 \item Absolute Mismatch Distance
 \item Czekanowski (Manhattan) Distance
 }

The distance measure for Sequence Tag Presence/Absence data (binary) can be one of:
\itemize{
 \item Euclidean Distance
 \item Scaled Euclidean Distance
 \item Simple Matching Distance
 \item Jaccard Distance
 \item Bray-Curtis Distance
 }

Refer to the dartR Technical Note on Distances in Genetics.
}
\examples{
D <- gl.dist.ind(testset.gl, method='euclidean')
D <- gl.dist.ind(testset.gs, method='euclidean')
}
\author{
Custodian: Arthur Georges -- Post to #' \url{https://groups.google.com/d/forum/dartr}
}
