% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.filter.callrate.loc.pop.r
\name{gl.filter.callrate.loc.pop}
\alias{gl.filter.callrate.loc.pop}
\title{Filter loci in a genlight \code{adegenet} object based on call rate and population level.}
\usage{
gl.filter.callrate.loc.pop(x, threshold)
}
\value{
The filtered genlight or genind object
}
\description{
This funciton is a convenience function that filters callrate for the set threshold for each population and then returns a genlight object where each loci of every population passes the threshold of the call rate filter. For more details on call rates see \code{gl.filter.callrate}.
}
\examples{
#Filter every loci in every population by callrate 0.9
gg <- gl.filter.callrate.loc.pop(testset.gl, 0.9)
}
