% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.filter.callrate.r
\name{gl.filter.callrate}
\alias{gl.filter.callrate}
\title{Filter loci or individuals in a genlight \{adegenet\} object based on call rate}
\usage{
gl.filter.callrate(
  x,
  method = "loc",
  threshold = 0.95,
  mono.rm = FALSE,
  recalc = FALSE,
  recursive = FALSE,
  plot = TRUE,
  bins = 50,
  plot_theme = theme_dartR(),
  plot_colours = two_colors,
  verbose = NULL
)
}
\arguments{
\item{x}{Name of the genlight object containing the SNP or presence/absence
(SilicoDArT) data [required].}

\item{method}{"loc" to specify that loci are to be filtered, "ind" to specify that specimens are to be filtered, "pop"
to remove loci that fail to meet the specified threshold in any one population [default "loc"].}

\item{threshold}{Threshold value below which loci will be removed [default 0.95].}

\item{mono.rm}{Remove monomorphic loci after analysis is complete [default FALSE].}

\item{recalc}{Recalculate the locus metadata statistics if any individuals are deleted in the filtering [default FALSE].}

\item{recursive}{Repeatedly filter individuals on call rate, each time removing monomorphic loci. Only applies if method="ind" and mono.rm=TRUE [default FALSE]}

\item{plot}{Specify if histograms of call rate, before and after, are to be produced [default TRUE].}

\item{bins}{Number of bins to display in histograms [default 50].}

\item{plot_theme}{Theme for the plot. See Details for options [default theme_dartR()].}

\item{plot_colours}{List of two color names for the borders and fill of the
plots [default two_colors].}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2, progress log ; 3, progress and results summary; 5, full report [default NULL, unless specified using gl.set.verbosity].}
}
\value{
The reduced genlight or genind object
}
\description{
SNP datasets generated by DArT have missing values primarily arising from failure
to call a SNP because of a mutation at one or both of the restriction enzyme 
recognition sites. Tag Presence/Absence datasets (SilicoDArT) have missing values 
where it is not possible to determine reliably if there the sequence tag can be 
called at a particular locus. This function will filtered out loci or individuals 
based on call rate.
}
\details{
The function \code{\link{gl.report.callrate}} will report summary of 
Call Rate for loci or individuals
 
\strong{ Function's output }

 The minimum, maximum, mean and a tabulation of call rate quantiles against
 thresholds are provided. Output also includes a boxplot and a
 histogram before and after the filtering.

 Plots and table are saved to the temporal directory (tempdir) and can be accessed
 with the function \code{\link{gl.print.reports}} and listed with the function
 \code{\link{gl.list.reports}}. Note that they can be accessed only in the 
 current R session because tempdir is cleared each time that the R session is closed.
  
 Examples of other themes that can be used can be consulted in \itemize{
 \item \url{https://ggplot2.tidyverse.org/reference/ggtheme.html} and \item
 \url{https://yutannihilation.github.io/allYourFigureAreBelongToUs/ggthemes/}
 }
}
\examples{
# SNP data
  result <- gl.filter.callrate(testset.gl, method="loc", threshold=0.95, verbose=3)
  result <- gl.filter.callrate(testset.gl, method="ind", threshold=0.8, verbose=3)
# Tag P/A data
  result <- gl.filter.callrate(testset.gs, method="loc", threshold=0.95, verbose=3)
  result <- gl.filter.callrate(testset.gs, method="ind", threshold=0.8, verbose=3)

}
\seealso{
\code{\link{gl.report.callrate}}, \code{\link{gl.list.reports}},
 \code{\link{gl.print.reports}}

Other filter functions: 
\code{\link{gl.filter.hamming}()},
\code{\link{gl.filter.parent.offspring}()},
\code{\link{gl.filter.sexlinked}()}
}
\author{
Arthur Georges and Bernd Gruber (Post to \url{https://groups.google.com/d/forum/dartr})
}
\concept{filter functions}
