% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.filter.cloneid.r
\name{gl.filter.cloneid}
\alias{gl.filter.cloneid}
\title{Filter Clones}
\usage{
gl.filter.cloneid(gl)
}
\arguments{
\item{gl}{a genlight object}
}
\value{
A genlight object containing only unique clones.
}
\description{
Filter to select only one instance of each clone.
}
\author{
Bernd Gruber (bugs? Post to \url{https://groups.google.com/d/forum/dartr})
}
