% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.filter.hwe.r
\name{gl.filter.hwe}
\alias{gl.filter.hwe}
\title{Filters loci that show significant departure from Hardy-Weinberg Equilibrium}
\usage{
gl.filter.hwe(
  x,
  subset = "each",
  multi_comp = TRUE,
  multi_comp_method = "bonferroni",
  method_sig = "Exact",
  alpha_val = 0.05,
  sample_size = 5,
  verbose = NULL
)
}
\arguments{
\item{x}{Name of the genlight object containing the SNP data [required].}

\item{subset}{Either a vector with population names, "each", "all" (see details)
[default "each"].}

\item{multi_comp}{Whether to adjust p-values for multiple comparisons [default TRUE].}

\item{multi_comp_method}{Method to adjust p-values for multiple comparisons: 
"holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr"
(see details) [default "bonferroni"]}

\item{method_sig}{Method for determining statistical significance: "ChiSquare" 
or "Exact" [default "Exact"].}

\item{alpha_val}{Level of significance for testing [default 0.05].}

\item{sample_size}{Minimum sample size per population required to test for 
Hardy-Weinberg proportions [default 5]}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2, 
progress log ; 3, progress and results summary; 5, full report 
[default NULL, unless specified using gl.set.verbosity]}
}
\value{
A genlight object with the loci departing significantly from HWE removed
}
\description{
Calculates the probabilities of agreement with H-W equilibrium based on observed
frequencies of reference homozygotes, heterozygotes and alternate homozygotes.
}
\details{
This function filters out loci departing from Hardy-Weinberg proportions in three ways:
\itemize{
\item across all populations together using: subset="all"
\item within each population separately using: subset="each"
\item within selected populations using for example: subset=c("pop1","pop2")
}

Loci are filtered out if they show HWE departure in any one population.

Tests for HWE are only valid if there is no population substructure (assuming
random mating), and the tests have sufficient power only when there is 
sufficient sample size (say, n individuals > 20). Note also that correction 
for multiple comparisons is probably required if you wish to place particular
importance on one or a few significant departures.


\code{\link[stats]{p.adjust}}

p.adjust.methods
 \code{\link[HardyWeinberg]{HWChisqStats}}
Uses the exact calculations contained in the function 
\code{\link[HardyWeinberg]{HWExactStats}} as developed by Wigginton 
et al. (2005).

The C++ code was generously shared by Christopher Chang, and the same code is 
used in the program PLINK (2.0).
}
\examples{
result <- gl.filter.hwe(x = bandicoot.gl)
}
\references{
\itemize{
\item Wigginton, J.E., Cutler, D.J., & Abecasis, G.R. (2005). A Note on Exact 
Tests of Hardy-Weinberg Equilibrium. American Journal of Human Genetics 76:887-893.
\item Graffelman, J. & Morales-Camarena, J. (2008). Graphical tests for 
Hardy-Weinberg equilibrium based on the ternary plot. Human Heredity 65:77-84.
\item Graffelman, J. (2015). Exploring Diallelic Genetic Markers: The Hardy 
Weinberg Package. Journal of Statistical Software 64:1-23.
}
}
\seealso{
\code{\link{gl.filter.hwe}}

Other filters/filter reports: 
\code{\link{gl.report.hwe}()}
}
\author{
Arthur Georges -- Post to \url{https://groups.google.com/d/forum/dartr}
}
\concept{filters/filter reports}
