% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.filter.ld.r
\name{gl.filter.ld}
\alias{gl.filter.ld}
\title{Filters loci in a genlight \{adegenet\} object based on linkage
 disequilibrium (LD)}
\usage{
gl.filter.ld(
  x,
  ld_report,
  threshold = 0.2,
  pop.limit = ceiling(nPop(x)/2),
  ind.limit = 10,
  verbose = NULL
)
}
\arguments{
\item{x}{Name of the genlight object containing the SNP data [required].}

\item{ld_report}{Output from function \code{\link{gl.report.ld.map}} [required].}

\item{threshold}{Threshold value below which loci will be removed
[default 0.2].}

\item{pop.limit}{Minimum number of populations in which LD should be more
than the threshold for a locus to be filtered out.
The default value is half of the populations [default ceiling(nPop(x)/2)].}

\item{ind.limit}{Minimum number of individuals that a population should
contain to take it in account to filter out loci in LD [default 10].}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2,
progress log; 3, progress and results summary; 5, full report
[default 2, unless specified using gl.set.verbosity].}
}
\value{
The reduced genlight object.
}
\description{
This function uses the statistic set in the parameter stat_keep from function
 \code{\link{gl.report.ld.map}} to choose the SNP to keep when two SNPs are 
 in LD. When a SNP is selected to be filtered out in each pairwise comparison,
  the function stores its  name in a list. In subsequent pairwise comparisons,
   if the SNP is already in the list, the other SNP will be kept.
}
\examples{
test <- bandicoot.gl
test <- gl.filter.callrate(test,threshold = 1)
res <- gl.report.ld.map(test)
res_2 <- gl.filter.ld(x=test,ld_report = res)
res_3 <- gl.report.ld.map(res_2)
}
\seealso{
\code{\link{gl.report.ld.map}}

Other filter functions: 
\code{\link{gl.filter.callrate}()},
\code{\link{gl.filter.parent.offspring}()},
\code{\link{gl.filter.sexlinked}()}
}
\author{
Custodian: Luis Mijangos -- Post to
 \url{https://groups.google.com/d/forum/dartr}
}
\concept{filter functions}
