% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.filter.monomorphs.r
\name{gl.filter.monomorphs}
\alias{gl.filter.monomorphs}
\title{Filter Monomorphs}
\usage{
gl.filter.monomorphs(gl)
}
\arguments{
\item{gl}{a genlight object}
}
\value{
A genlight object with monomorphic loci removed
}
\description{
Remove monomorphic loci, including those with all NAs
}
\details{
A DArT dataset will not have monomorphic loci, but they can arise
when populations are deleted by assignment or by using the delete
option in gl.pop.recode(). Retaining monomorphic loci
unnecessarily increases the size of the dataset.
}
\examples{
gl <- gl.filter.monomorphs(testset.gl)
}
\author{
Arthur Georges (Post to \url{https://groups.google.com/d/forum/dartr})
}
