% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.filter.rdepth.r
\name{gl.filter.rdepth}
\alias{gl.filter.rdepth}
\title{Filter Read Depth}
\usage{
gl.filter.rdepth(gl, lower = 5, upper = 50)
}
\arguments{
\item{gl}{a genlight object}

\item{lower}{lower read depth threshold}

\item{upper}{upper read depth threshold}
}
\value{
a genlight object
}
\description{
Filter loci in a genlight \{adegenet\} object based on counts of
sequence tags scored at a locus (read depth)
}
\details{
SNP datasets generated by DArT report AvgCountRef and AvgCountSnp
as counts of sequence tags for the reference and alternate alleles
respectively. These can be added for an index of Read
Depth. Filtering on Read Depth can be on the basis of loci with
exceptionally low counts, or loci with exceptionally high counts.
}
\examples{
gl.report.rdepth(testset.gl)
result <- gl.filter.rdepth(testset.gl, lower=8, upper=50, v=3)
}
\author{
Arthur Georges (Post to \url{https://groups.google.com/d/forum/dartr})
}
