% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.filter.repavg.r
\name{gl.filter.repavg}
\alias{gl.filter.repavg}
\title{Filter loci in a genlight \{adegenet\} object based on average reproducibility of alleles at a locus}
\usage{
gl.filter.repavg(gl, t = 1)
}
\arguments{
\item{gl}{-- name of the genlight object containing the SNP data [required]}

\item{t}{-- threshold value below which loci will be removed [default 0.95]}
}
\value{
Returns a genlight object retaining loci with a RepAvg less than the specified threshold deleted.
}
\description{
SNP datasets generated by DArT have in index, RepAvg, generated by reproducing the data independently for 30\% of loci.
RepAvg is the proportion of alleles that give a reproducible result, averaged over both alleles for each locus.
}
\examples{
\dontrun{
gl.report.repavg(gl)
result <- gl.filter.repavg(gl, t=0.95)
}
}
\author{
Arthur Georges (glbugs@aerg.canberra.edu.au)
}

