% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.filter.secondaries.r
\name{gl.filter.secondaries}
\alias{gl.filter.secondaries}
\title{Filter secondary SNPs from a genlight object}
\usage{
gl.filter.secondaries(gl, method = c("random", "best", "first"))
}
\arguments{
\item{gl}{a genlight object}

\item{method}{the method of selecting SNP locus to retain, best, random or first}
}
\value{
The reduced genlight object
}
\description{
SNP datasets generated by DArT include fragments with more than one
SNP and record them separately with the same CloneID. These
multiple SNP loci within a fragment (secondaries) are likely to be
linked, and are often removed.  This script filters out secondary
loci, retaining either the loci with greatest repeatability and
avgPIC (method="best"), a random loci (method="random") or the
first loci encountered (method="first").
}
\examples{
result <- gl.report.secondaries(testset.gl)
result2 <- gl.filter.secondaries(testset.gl)
}
\author{
Arthur Georges (Post to \url{https://groups.google.com/d/forum/dartr})
}
