% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.grm.r
\name{gl.grm}
\alias{gl.grm}
\title{Calculates the genomic relatedness matrix}
\usage{
gl.grm(x, plotheatmap = TRUE, return.imputed = FALSE, verbose = NULL,
  ...)
}
\arguments{
\item{x}{-- a genlight object}

\item{plotheatmap}{-- a switch if a heatmap should be shown [Default:TRUE]}

\item{return.imputed}{switch if loci with imputed data should be returned (see ?A.mat in package rrBLUP)}

\item{verbose}{-- verbosity: 0, silent or fatal errors; 1, begin and end; 2, progress log ; 3, progress and results summary; 5, full report [default 2 or as specified using gl.set.verbosity]}

\item{...}{parameters passed to function A.mat from package rrBLUP}
}
\value{
a genomic relatedness matrix
}
\description{
The G matrix is calculated by centering the allele frequency matrix  of the second allele by substracting 2 times the allefrequency
}
\examples{
gl.grm(bandicoot.gl[1:5,1:10],plotheatmap=TRUE)  
}
