% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.He.r
\name{gl.He}
\alias{gl.He}
\title{Estimates expected Heterozygosity}
\usage{
gl.He(gl)
}
\arguments{
\item{gl}{A genlight object [required]}
}
\value{
A simple vector whit Ho for each loci
}
\description{
Estimates expected Heterozygosity
}
\author{
Bernd Gruber (Post to \url{https://groups.google.com/d/forum/dartr})
}
