% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.Ho.r
\name{gl.Ho}
\alias{gl.Ho}
\title{Observed Heterozygosity.}
\usage{
gl.Ho(gl)
}
\arguments{
\item{gl}{a genlight object}
}
\value{
The observed heterozygosity for each loci as a vector.
}
\description{
Calculate the observed heterozygosity for each loci.
}
\author{
Bernd Gruber (bugs? Post to \url{https://groups.google.com/d/forum/dartr})
}
