% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.Hs.r
\name{gl.Hs}
\alias{gl.Hs}
\title{Expected Heterozygosity.}
\usage{
gl.Hs(gl)
}
\arguments{
\item{gl}{a genlight object}
}
\value{
The expected heterozygosity for each loci as a vector.
}
\description{
Calculate the expected heterozygosity for each loci.
}
\author{
Bernd Gruber (bugs? Post to \url{https://groups.google.com/d/forum/dartr})
}
