% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.hwe.diagnostics.r
\name{gl.hwe.diagnostics}
\alias{gl.hwe.diagnostics}
\title{Provides descriptive stats and plots to diagnose potential problems 
with Hardy-Weinberg proportions}
\usage{
gl.hwe.diagnostics(
  x,
  alpha_val = 0.05,
  bins = 20,
  colors_hist = two_colors,
  colors_barplot = two_colors_contrast,
  plot_theme = theme_dartR(),
  save2tmp = FALSE,
  verbose = NULL
)
}
\arguments{
\item{x}{Name of the genlight object containing the SNP data [required].}

\item{alpha_val}{Level of significance for testing [default 0.05].}

\item{bins}{Number of bins to display in histograms [default 20].}

\item{colors_hist}{List of two color names for the borders and fill of the
histogram [default two_colors].}

\item{colors_barplot}{Vector with two color names for the observed and
expected number of significant HWE tests [default two_colors_contrast].}

\item{plot_theme}{User specified theme [default theme_dartR()].}

\item{save2tmp}{If TRUE, saves any ggplots and listings to the session
temporary directory (tempdir) [default FALSE].}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2,
progress log ; 3, progress and results summary; 5, full report
[default NULL, unless specified using gl.set.verbosity].}
}
\value{
A list with plots and table.
}
\description{
Different causes may be responsible for lack of Hardy-Weinberg proportions. 
This function helps diagnose potential problems.
}
\details{

}
\examples{
res <- gl.hwe.diagnostics(x = gl.filter.allna(platypus.gl))
}
\references{
\itemize{
\item De Meeûs, T., Guégan, J.-F., Teriokhin, A.T., 2009. MultiTest V.1.2, a 
program to binomially combine independent tests and performance comparison 
with other related methods on proportional data. BMC Bioinformatics 10, 443-443.
\item Fisher, R., 1970. Statistical methods for research workers Edinburgh: 
Oliver and Boyd.
\item Waples, R. S. (2015). Testing for Hardy–Weinberg proportions: have we
lost the plot?. Journal of heredity, 106(1), 1-19.
}
}
\seealso{
\code{\link{gl.report.hwe}}

Other reporting functions: 
\code{\link{gl.report.bases}()},
\code{\link{gl.report.diversity}()},
\code{\link{gl.report.heterozygosity}()},
\code{\link{gl.report.parent.offspring}()},
\code{\link{gl.report.pa}()},
\code{\link{gl.report.sexlinked}()}
}
\author{
Custodian: Carlo Pacioni -- Post to
\url{https://groups.google.com/d/forum/dartr}
}
\concept{reporting functions}
