% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.ID.replicate.r
\name{gl.ID.replicate}
\alias{gl.ID.replicate}
\title{Identify replicate individuals}
\usage{
gl.ID.replicate(x, loc_threshold = 100, perc_geno = 0.99)
}
\arguments{
\item{x}{Name of the genlight object containing the SNP data [required].}

\item{loc_threshold}{Minimum number of loci required to asses that two 
individuals are replicates [default 100].}

\item{perc_geno}{Mimimum percentage of genotypes in which two individuals 
should be the same [default 0.99].}
}
\value{
Dataframe with pairwise results of percentage of same genotypes 
between two individuals and the number of loci used in the comparison.
}
\description{
Identify replicate individuals
}
\details{
This function uses an C++ implementation, so package Rcpp needs to be 
installed and it is therefore fast (once it has compiled the function after 
the first run).
}
\examples{
res_rep <- gl.ID.replicate(platypus.gl, loc_threshold = 500, perc_geno = 0.85)
}
\author{
Custodian: Luis Mijangos -- Post to
\url{https://groups.google.com/d/forum/dartr}
}
