% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.impute.r
\name{gl.impute}
\alias{gl.impute}
\title{Imputation of missing data}
\usage{
gl.impute(x, method = "HW", parallel = FALSE, verbose = NULL)
}
\arguments{
\item{x}{Name of the genlight object containing the SNP data [required].}

\item{method}{Imputation method, either "frequency" or "HW" [default "HW"].}

\item{parallel}{A logical indicating whether multiple cores -if available- 
should be used for the computations (TRUE), or not (FALSE); requires the 
package parallel to be installed [default FALSE].}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2, 
progress log ; 3, progress and results summary; 5, full report 
[default 2 or as specified using gl.set.verbosity].}
}
\value{
A genlight object with the missing data imputed.
}
\description{
This function imputes genotypes on a population-by-population basis, where 
populations can be considered panmictic.
}
\details{
We advice that imputation should be performed on sampling locations, before 
any aggregation. The imputation is achieved by replacing missing values using
either of two methods:
\itemize{
\item "frequency" where allele frequencies for the population to which 
the respective individual belongs are used as probabilities to sample 2 
alleles for each missing data in each individual. 
\item "HW" where the Hardy-Weinberg equation (p^2 + 2pq + q^2 = 1) within 
each population is used to sample genotypes for each missing data in each 
individual.
}

Note that loci that have all the values missing are not imputed. Consider 
using the function \code{\link{gl.filter.allna}}.
}
\examples{
gl <- gl.impute(platypus.gl)

}
\author{
Custodian: Luis Mijangos (Post to \url{https://groups.google.com/d/forum/dartr})
}
