% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.ld.haplotype.r
\name{gl.ld.haplotype}
\alias{gl.ld.haplotype}
\title{Provides descriptive stats and plots to diagnose linkage
disequilibrium patterns}
\usage{
gl.ld.haplotype(
  x,
  pop_name = NULL,
  chrom_name = NULL,
  ld_max_pairwise = 1e+07,
  maf = 0.05,
  ld_stat = "R.squared",
  ind.limit = 10,
  min_snps = 10,
  ld_threshold_haplo = 0.5,
  save2tmp = FALSE,
  verbose = NULL
)
}
\arguments{
\item{x}{Name of the genlight object containing the SNP data [required].}

\item{pop_name}{Name of the population to analyse. If NULL all the 
populations are analised [default NULL].}

\item{chrom_name}{Nme of the chromosome to analyse. If NULL all the 
chromosomes are analised [default NULL].}

\item{ld_max_pairwise}{Maximum distance in number of base pairs at which LD
should be calculated [default 10000000].}

\item{maf}{Minor allele frequency (by population) threshold to filter out 
loci. If a value > 1 is provided it will be interpreted as MAC (i.e. the
 minimum number of times an allele needs to be observed) [default 0.05].}

\item{ld_stat}{The LD measure to be calculated: "LLR", "OR", "Q", "Covar",
"D.prime", "R.squared", and "R". See \code{\link[snpStats]{ld}}
 (package snpStats) for details [default "R.squared"].}

\item{ind.limit}{Minimum number of individuals that a population should
contain to take it in account to report loci in LD [default 10].}

\item{min_snps}{Minimum number of SNPs that should have a haplotype to call 
it [default 10].}

\item{ld_threshold_haplo}{Minimum LD between adjacent SNPs to call a 
haplotype [default 0.5].}

\item{save2tmp}{If TRUE, saves any ggplots and listings to the session
temporary directory (tempdir) [default FALSE].}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2,
progress log; 3, progress and results summary; 5, full report
[default 2, unless specified using gl.set.verbosity].}
}
\value{
A dataframe with information for each SNP pair in LD.
}
\description{
Provides descriptive stats and plots to diagnose linkage
disequilibrium patterns
}
\details{
This function reports LD between SNP pairs by population. Reports include
the number of SNPs that would be filtered out given specific LD thresholds.
The function code{\link{gl.filter.ld}} filters out the SNPs in LD using as
input the results of gl.report.ld.map. However, the actual number of SNPs to
be filtered out depends on the parameters set in the function
code{\link{gl.filter.ld}}. Therefore, the number of SNPs to be filtered out
 reported by gl.report.ld.map should be used as a guide and not as a definite
  number.

\enumerate{
\item

}
A bar plot with observed
and expected (null expectation) number of significant HWE tests for the same
locus in multiple populations (that is, the x-axis shows whether a locus
results significant in 1, 2, ..., n populations. The y axis is the count of
these occurrences.

If SNPs are mapped to a reference genome, the function creates a plot showing
the pairwise LD measure against distance in number of base pairs pooled over
all the chromosomes and a red line representing the threshold (R.squared =
0.2) that is commonly used to imply that two loci are unlinked (Delourme et
al., 2013; Li et al., 2014). Additionally, boxplots of LD by population and
a histogram showing LD frequency are presented.

If SNPs are not mapped to a reference genome, only boxplots by population and
 a histogram showing LD frequency are presented.
}
\examples{
\dontrun{
x <- platypus.gl
x$position <- x$other$loc.metrics$ChromPos_Platypus_Chrom_NCBIv1
x$chromosome <- x$other$loc.metrics$Chrom_Platypus_Chrom_NCBIv1
ld_res <- gl.report.ld.map(x,ld_max_pairwise = 1000000)
}
}
\references{
\itemize{
\item Delourme, R., Falentin, C., Fomeju, B. F., Boillot, M., Lassalle, G.,
André, I., . . . Marty, A. (2013). High-density SNP-based genetic map
development and linkage disequilibrium assessment in Brassica napusL. BMC
genomics, 14(1), 120.
\item Li, X., Han, Y., Wei, Y., Acharya, A., Farmer, A. D., Ho, J., . . .
Brummer, E. C. (2014). Development of an alfalfa SNP array and its use to
evaluate patterns of population structure and linkage disequilibrium. PLoS
One, 9(1), e84329.
 }
}
\seealso{
\code{\link{gl.filter.ld}}
}
\author{
Custodian: Luis Mijangos -- Post to
 \url{https://groups.google.com/d/forum/dartr}
}
