% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.list.reports.r
\name{gl.list.reports}
\alias{gl.list.reports}
\title{Prints dartR reports saved in tempdir}
\usage{
gl.list.reports()
}
\value{
Prints a table with all reports saved in tempdir. Currently the style cannot be changed.
}
\description{
Prints dartR reports saved in tempdir
}
\examples{
gl.report.callrate(testset.gl)
gl.list.reports()
}
\author{
Bernd Gruber & Luis Mijangos (bugs? Post to \url{https://groups.google.com/d/forum/dartr})
}
