% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.pcoa.plot.r
\name{gl.pcoa.plot}
\alias{gl.pcoa.plot}
\title{PCoA ordination plot}
\usage{
gl.pcoa.plot(glPca, gl, scale = FALSE, ellipse = FALSE, level = 0.95,
  labels = c("pop", "none", "ind", "interactive", "legend"), xaxis = 1,
  yaxis = 2)
}
\arguments{
\item{glPca}{a glPca object}

\item{gl}{a genlight object}

\item{scale}{logical indicating whether the x and y axes are to be scaled in proportion to \% variation explained}

\item{ellipse}{logical indicating to display tolerance ellipses for ach population}

\item{level}{quantiles for tolerance ellipses}

\item{labels}{the type of labelling to be applied.}

\item{xaxis}{the component to plot as the x axis}

\item{yaxis}{the component to plot as the y axis}
}
\value{
A ggplot object.
}
\description{
Plot a PCoA ordination.
}
\details{
The factor scores are taken from the output of gl.pcoa() -- an
object of class glPca -- and the population assignments are taken
from from the original data file. The specimens are shown in a
bivariate plot optionally with adjacent labels and enclosing
ellipses. Population labels on the plot are shuffled so as not to
overlap (using package \{directlabels\}).  This can be a bit
clunky, as the labels may be some distance from the points to
which they refer, but it provides the opportunity for moving
labels around using graphics software (Adobe Illustrator).

Any pair of axes can be specified from the ordination, provided
they are within the range of the nfactors value provided to
gl.pcoa(). Axes can be scaled to represent the proportion of
variation explained. In any case, the proportion of variation
explained by each axis is provided in the axis label.

Points displayed in the ordination can be identified if the option
labels="interactive" is chosen, in which case the resultant plot
is ggplotly() friendly. Running ggplotyly() with no parameters
will replot the data and allow identification of points by moving
the mouse over them. Refer to the plotly package for further
information. Do not forget to load the library via
library(plotly).
}
\author{
Arthur Georges (bugs? Post to \url{https://groups.google.com/d/forum/dartr})
}
