% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.pcoa.r
\name{gl.pcoa}
\alias{gl.pcoa}
\title{PCoA ordination}
\usage{
gl.pcoa(gl, nfactors = 5)
}
\arguments{
\item{gl}{a genlight object}

\item{nfactors}{Number of dimensions to retain in the ordination.}
}
\value{
An object of class glPca containing the eigenvalues, factor scores and factor loadings.
}
\description{
Perform a PCoA of the genetic data in a genlight object.
}
\examples{
pcoa <- gl.pcoa(testset.gl, nfactors=3)
}
\author{
Arthur Georges (bugs? Post to \url{https://groups.google.com/d/forum/dartr})
}
