% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.pcoa.scree.r
\name{gl.pcoa.scree}
\alias{gl.pcoa.scree}
\title{PCoA Scree Plot}
\usage{
gl.pcoa.scree(glPca, n = length(glPca$eig))
}
\arguments{
\item{glPca}{a glPca object}

\item{n}{number of components to display}
}
\value{
a ggplot object.
}
\description{
Produce a scree plot from a PCoA analysis
}
\details{
Plot the contribution each principal component makes to the total
variance, expressed as a percentage.
}
\author{
Arthur Georges (bugs? Post to \url{https://groups.google.com/d/forum/dartr})
}
