% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.plot.r
\name{gl.plot}
\alias{gl.plot}
\title{Plotting genlight object as a smear plot (loci by individuals color 
coded for scores of 0, 1, 2 and NA)}
\usage{
gl.plot(
  x,
  group_pop = FALSE,
  labels_plot = FALSE,
  ind_labels = indNames(x),
  plot_colours = four_colors,
  posi = "bottom",
  save2tmp = FALSE,
  verbose = NULL
)
}
\arguments{
\item{x}{Name of the genlight object containing the SNP or presence/absence
(SilicoDArT) data [required].}

\item{group_pop}{Group by population [default TRUE]}

\item{labels_plot}{If TRUE, individual labels are added [default FALSE].}

\item{ind_labels}{Labels for individuals [default indNames(x)].}

\item{plot_colours}{Vector with four color names for homozygotes for the 
reference allele, heterozygotes, homozygotes for the alternative allele and 
for missing values (NA) [default four_colors].}

\item{posi}{Position of the legend: “left”, “top”, “right”, “bottom” or
"none" [default = "bottom"].}

\item{save2tmp}{If TRUE, saves plot to the session temporary directory 
(tempdir) [default FALSE].}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2, 
progress log ; 3, progress and results summary; 5, full report [default NULL].}
}
\value{
Returns unaltered genlight object
}
\description{
It adds the option to put labels on the individuals and grouping by populations. 
If there are too many individuals, it is best to use labels_plot=FALSE.
}
\examples{
gl.plot(bandicoot.gl[1:10,],labels=TRUE)

}
\seealso{
\code{\link{gl.filter.callrate}}
}
\author{
Arthur Georges -- Post to \url{https://groups.google.com/d/forum/dartr}
}
\concept{Exploration/visualisation functions}
