% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.print.reports.r
\name{gl.print.reports}
\alias{gl.print.reports}
\title{Prints dartR reports saved in tempdir}
\usage{
gl.print.reports(print_report)
}
\value{
Prints reports that were saved in tempdir.
}
\description{
Prints dartR reports saved in tempdir
}
\examples{
\dontrun{
reports <- gl.print.reports(1)
}
}
\author{
Bernd Gruber & Luis Mijangos (bugs? Post to \url{https://groups.google.com/d/forum/dartr})
}
