% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.propShared.r
\name{gl.propShared}
\alias{gl.propShared}
\title{Calculate a similarity(distance) matrix for individuals on the proportion of shared alleles}
\usage{
gl.propShared(x)
}
\arguments{
\item{x}{-- name of the genlight containing the SNP genotypes [required]}
}
\description{
This script calculates a individual based distance matrix. It uses an C++ implementation, so package Rcpp needs to be installed.
}
\examples{
res <- gl.propShared(testset.gl)
res[1:5,1:7] #show only a small part of the matrix
}
\author{
Bernd Gruber (Post to \url{https://groups.google.com/d/forum/dartr})
}
