% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.read.dart.r
\name{gl.read.dart}
\alias{gl.read.dart}
\title{Import DarT data as genlight object.}
\usage{
gl.read.dart(filename, ind.metafile = NULL, na.strings = "-",
  skip = NULL, last.metric = "RepAvg", nrows = NULL)
}
\arguments{
\item{filename}{DarT data in csv format.}

\item{ind.metafile}{name of csv file with covariates for each individual.}

\item{na.strings}{strings representing NAs.}

\item{skip}{number of lines before the row of column names. If \code{NULL} it is diagnosed from the file.}

\item{last.metric}{the name of the last column of preceeding the genetic data}

\item{nrows}{the DarT format.  If \code{NULL} it is diagnosed from the file.}
}
\value{
A genlight object.
}
\description{
Import DarT data into R and convert it to a genlight object.
}
\details{
This function is a wrapper for \code{\link{utils.read.dart}} and
\code{\link{utils.dart2genlight}} that converts a dart file into a
genlight object in one step.

The file of individual covariates is searched for the columns
\describe{
\item{\code{id}}{individual identifiers matching the genetic data in the dart object}
\item{\code{pop}}{population identifier for each individual}
\item{\code{lat}, \code{lon}}{spatial coordinates (perferably in decimal degrees WGS1984 format)}
}
Of these, only \code{id} is mandatory, but other functions in the
package may require metadata provided from DArT, Other covariates
contained in the file will be appended to individual covariate
dataframe in the genlight object.
}
\seealso{
\code{\link{utils.read.dart}}, \code{\link{utils.dart2genlight}}
}
\author{
Bernd Gruber (bugs? Post to \url{https://groups.google.com/d/forum/dartr})
}
