% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.read.fasta.r
\name{gl.read.fasta}
\alias{gl.read.fasta}
\title{Reads FASTA files and converts them to genlight object}
\usage{
gl.read.fasta(fasta_files, parallel = FALSE, n_cores = NULL, verbose = NULL)
}
\arguments{
\item{fasta_files}{Fasta files to read [required].}

\item{parallel}{A logical indicating whether multiple cores -if available-
should be used for the computations (TRUE), or not (FALSE); requires the
 package parallel to be installed [default FALSE].}

\item{n_cores}{If parallel is TRUE, the number of cores to be used in the
computations; if NULL, then the maximum number of cores available on the
 computer is used [default NULL].}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2,
progress log; 3, progress and results summary; 5, full report
[default 2, unless specified using gl.set.verbosity].}
}
\value{
A genlight object.
}
\description{
The following IUPAC Ambiguity Codes are taken as heterozygotes:
 W, S, M, K, R, Y. 
 
 Take into account that the function will take a little bit longer the first
  time you run it because C++ functions must be compiled.
}
\details{
Ambiguity characters are often used to code heterozygotes. However, using
 heterozygotes as ambiguity characters may bias many estimates. See more
  information in the link below:
\url{https://evodify.com/heterozygotes-ambiguity-characters/}
}
\examples{
 # Folder where the fasta files are located. 
 folder_samples <- system.file('extdata', package ='dartR')
 # listing the FASTA files, including their path. Files have an extension
 # that contains "fas".
 file_names <- list.files(path = folder_samples, pattern = "*.fas", 
                          full.names = TRUE)
 # reading fasta files
  obj <- gl.read.fasta(file_names)
}
\author{
Custodian: Luis Mijangos -- Post to
\url{https://groups.google.com/d/forum/dartr}
}
\concept{reading functions}
