% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.recalc.metrics.r
\name{gl.recalc.metrics}
\alias{gl.recalc.metrics}
\title{Recalculate Locus Metrics}
\usage{
gl.recalc.metrics(gl)
}
\arguments{
\item{gl}{a genlight object}
}
\value{
Returns a genlight object with the recalculated locus metadata.
}
\description{
Recalculates the locus metrics of a genlight object
}
\details{
If individuals are deleted from a genlight object, the cached
locus metrics no longer valid. This function recalculates the locus
metrics avgPIC, CallRate, freqHets, freqHomRef, freqHomSnp,
OneRatioRef, OneRatioSnp, PICRef and PICSnp.
}
\author{
Arthur Georges (bugs? Post to \url{https://groups.google.com/d/forum/dartr})
}
