% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.report.callrate.r
\name{gl.report.callrate}
\alias{gl.report.callrate}
\title{Report summary of Call Rate for loci or individuals}
\usage{
gl.report.callrate(x, method = "loc", plot = FALSE,
  smearplot = FALSE)
}
\arguments{
\item{x}{-- name of the genlight or genind object containing the SNP data [required]}

\item{method}{specify the type of report by locus (method="loc") or individual (method="ind") [default method="loc"]}

\item{plot}{specify if a histogram of call rate is to be produced [default FALSE]}

\item{smearplot}{if TRUE, will produce a smearplot of individuals against loci [default FALSE]}
}
\value{
Tabulation of CallRate against Threshold
}
\description{
SNP datasets generated by DArT have missing values primarily arising from failure to call a SNP because of a mutation
at one or both of the the restriction enzyme recognition sites. This script reports the number of missing values for each
of several percentiles. The script gl.filter.callrate() will filter out the loci with call rates below a specified threshold.
}
\details{
A histogram and or a smearplot can be requested. Note that the smearplot is computationally intensive, and will take time to 
execute on large datasets.
}
\examples{
gl.report.callrate(testset.gl)
}
\author{
Arthur Georges (Post to \url{https://groups.google.com/d/forum/dartr})
}
