% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.report.dups.r
\name{gl.report.dups}
\alias{gl.report.dups}
\title{Report duplicated loci in a genlight \{adegenet\} object}
\usage{
gl.report.dups(gl)
}
\arguments{
\item{gl}{-- name of the genlight object containing the SNP data, or the genind object containing the SilocoDArT data [required]}
}
\value{
1
}
\description{
SNP datasets generated by DArT include fragments with more than one SNP and record them separately with the same CloneID (=AlleleID).
These multiple SNP loci within a fragment are likely to be linked, and so you may wish to remove duplicates.
This script reports duplicate loci.
}
\examples{
gl.report.dups(testset.gl)
}
\author{
Arthur Georges (glbugs@aerg.canberra.edu.au)
}

