% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.report.hamming.r
\name{gl.report.hamming}
\alias{gl.report.hamming}
\title{Calculates the pairwise Hamming distance between DArT trimmed DNA sequences}
\usage{
gl.report.hamming(gl)
}
\arguments{
\item{gl}{-- genlight object [required]}
}
\value{
Histogram of Hamming distance for the gl object
}
\description{
Hamming distance can be computed 
by exploiting the fact that the dot product of two binary vectors x and (1 – y) 
counts the corresponding elements that are different between x and y.
This approach can also be used for vectors that contain more than two possible 
values at each position (e.g. A, C, T or G).
}
\details{
If a pair of DNA sequences are of differing length, the longer is truncated.

The algorithm is that of Johann de Jong 
(https://johanndejong.wordpress.com/2015/10/02/faster-hamming-distance-in-r-2/)
as implimented in utils.hamming.r
}
\examples{
gl.report.hamming(testset.gl)
}
\author{
Arthur Georges (glbugs@aerg.canberra.edu.au)
}
