% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.report.hwe.r
\name{gl.report.hwe}
\alias{gl.report.hwe}
\title{Reports departure from Hardy-Weinberg proportions}
\usage{
gl.report.hwe(
  x,
  subset = "each",
  method_sig = "Exact",
  multi_comp = FALSE,
  multi_comp_method = "fdr",
  alpha_val = 0.05,
  pvalue_type = "midp",
  cc_val = 0.5,
  min_sample_size = 5,
  plot.out = TRUE,
  plot_colours = two_colors_contrast,
  max_plots = 4,
  verbose = NULL
)
}
\arguments{
\item{x}{Name of the genlight object containing the SNP data [required].}

\item{subset}{Way to group individuals to perform H-W tests. Either a vector 
with population names, "each", "all" (see details) [default "each"].}

\item{method_sig}{Method for determining statistical significance: "ChiSquare" 
or "Exact" [default "Exact"].}

\item{multi_comp}{Whether to adjust p-values for multiple comparisons [default FALSE].}

\item{multi_comp_method}{Method to adjust p-values for multiple comparisons: 
"holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr"
(see details) [default "fdr"].}

\item{alpha_val}{Level of significance for testing [default 0.05].}

\item{pvalue_type}{Type of p-value to be used in the Exact method. 
Either "dost","selome","midp" (see details) [default "midp"].}

\item{cc_val}{[default 0.5]}

\item{min_sample_size}{Minimum number of individuals per population in which 
perform H-W tests [default 5]}

\item{plot.out}{If TRUE, will produce Ternary Plot(s) [default TRUE].}

\item{plot_colours}{Vector with two colour names for the significant and 
not-significant loci [default two_colors_contrast].}

\item{max_plots}{Maximum number of plots to print per page [default 4].}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2, 
progress log ; 3, progress and results summary; 5, full report 
[default NULL, unless specified using gl.set.verbosity].}
}
\value{
returns a dataframe containing loci, counts of reference SNP homozygotes,
heterozygotes and alternate SNP homozygotes; probability of departure from 
H-W equilibrium, and per locus significance with and without Bonferroni correction.
}
\description{
Calculates the probabilities of agreement with H-W proportions based on observed
frequencies of reference homozygotes, heterozygotes and alternate homozygotes.
}
\details{
Testing for Hardy-Weinberg proportions should be a process that involves
 a careful evaluation of the results. There are several factors that can cause deviations from Hardy-Weinberg proportions
including: mutation, finite population size, selection, population structure, age 
structure, assortative mating, sex linkage, nonrandom sampling and genotyping
 errors.  A good place to start is Waples (2015). 
 
\strong{Grouping individuals}
Tests for HWE are only valid if there is no population substructure (assuming random mating).
This function tests for deviations from Hardy Weinberg proportions in each 
locus within populations Populations can be defined in three ways:
\itemize{
\item Merging all populations in the dataset using subset = "all".
\item Within each population separately using: subset = "each".
\item Within selected populations using for example: subset = c("pop1","pop2").
}

\strong{Statistical methods}
This function tests for deviations from Hardy Weinberg proportions using two 
different methods:
\itemize{
\item "ChiSquare"   \code{\link[HardyWeinberg]{HWChisqStats}}

\item "Exact"  Uses the exact calculations contained in the function 
\code{\link[HardyWeinberg]{HWExactStats}} as developed by Wigginton 
et al. (2005). The C++ code was generously shared by Christopher Chang. The same is 
used in the program PLINK (2.0).
pvaluetype	if pvaluetype is set to dost then the p-value of a two-sided test 
is computed as twice the tail area of a one-sided test. When set to selome, 
the p-value is computed as the sum of the probabilities of all samples less 
or equally likely as the current sample. When set to midp, the p-value is 
computed as half the probability of the current sample + the probabilities of 
all samples that are more extreme.
}

\strong{Correction for multiple comparisons}
This function uses correction for multiple tests using the following methods
\code{\link[stats]{p.adjust}}:
\itemize{
\item "holm" is also known as the sequential Bonferroni technique (Rice, 1989). 
This method has a greater statistical power than the standard Bonferroni test, 
however this method becomes very stringent when many tests are performed and 
many real deviations from the null hypothesis can go undetected (Waples, 2015).
\item "hochberg" (Hochberg, 1988) 
\item "hommel" (Hommel, 1988). This method is more powerful than Hochberg's,  
but the difference is usually small .
\item "bonferroni" in which p-values are multiplied by the number of tests 
(i.e. the number of loci). This method is very stringent and therefore has 
reduced power to detect multiple departures from the null hypothesis.
\item "BH" (Benjamini & Hochberg, 1995).
\item "BY" (Benjamini & Yekutieli, 2001). 
}

The first four methods are designed to give strong control of the family-wise 
error rate. The last two methods control the false discovery rate (FDR), 
the expected proportion of false discoveries among the rejected hypotheses.
The false discovery rate is a less stringent condition than the family-wise
error rate, so these methods are more powerful than the others, especially 
when number of tests is large.
The number of tests on which the adjustment for multiple comparisons is 
number of populations times number of loci 

\strong{Sample size}
Tests have sufficient power only when there is 
sufficient sample size (say, n individuals > 15). Note also that correction 
for multiple comparisons is probably required if you wish to place particular
importance on one or a few significant departures.

\strong{Ternary plots}
Ternary plots can be used to visualise patterns of hardy weinerg proportions  We design a graphical test for Hardy-Weinberg equilibrium. This can circumvent the calculation of p values and the statistical (non)significance of a large number of bi- allelic markers can be inferred from their position in a graph.
it is necessary to install ggtern
A Ternary Plot is optionally produced (plot.out = TRUE) -- see Graffelman & 
Morales-Camarena (2008) for further details. Ternary plots are based on the package {HardyWeinberg} 
(Graffelman (2015). The plot labels loci that depart significantly from HWE 
as red, and those not showing significant departure as green. 

colour the marker points automatically according to the result of a 
significance test (green markers non-significant, red markers significant).
#' Uses the exact calculations contained in function \code{\link{utils.prob.hwe}}
as developed by Wigginton et al. (2005).
}
\examples{
gl.report.hwe(x = bandicoot.gl)
}
\references{
\itemize{
 \item Benjamini, Y., and Yekutieli, D. (2001). The control of the false 
 discovery rate in multiple testing under dependency. Annals of Statistics, 
 29, 1165–1188.
\item Graffelman, J. (2015). Exploring Diallelic Genetic Markers: The Hardy 
Weinberg Package. Journal of Statistical Software 64:1-23.
\item Graffelman, J. & Morales-Camarena, J. (2008). Graphical tests for 
Hardy-Weinberg equilibrium based on the ternary plot. Human Heredity 65:77-84.
\item Hochberg, Y. (1988). A sharper Bonferroni procedure for multiple tests
 of significance. Biometrika, 75, 800–803.
\item Hommel, G. (1988). A stagewise rejective multiple test procedure based 
 on a modified Bonferroni test. Biometrika, 75, 383–386.
\item Rice, W. R. (1989). Analyzing tables of statistical tests. Evolution,
 43(1), 223-225.
\item Waples, R. S. (2015). Testing for Hardy–Weinberg proportions: have we 
lost the plot?. Journal of heredity, 106(1), 1-19.
\item Wigginton, J.E., Cutler, D.J., & Abecasis, G.R. (2005). A Note on Exact 
Tests of Hardy-Weinberg Equilibrium. American Journal of Human Genetics 76:887-893.
}
}
\seealso{
\code{\link{gl.filter.hwe}}

Other filters/filter reports: 
\code{\link{gl.filter.hwe}()}
}
\author{
Arthur Georges -- Post to \url{https://groups.google.com/d/forum/dartr}
}
\concept{filters/filter reports}
