% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.report.maf.r
\name{gl.report.maf}
\alias{gl.report.maf}
\title{Minor Allele Frequency Report}
\usage{
gl.report.maf(gl, maf.limit = 0.5, ind.limit = 5)
}
\arguments{
\item{gl}{a genelight object}

\item{maf.limit}{the maximum MAF to show for each population}

\item{ind.limit}{histograms are not produced for populations with
fewer individuals than this}
}
\description{
Report minor allele frequency (MAF) distributions.
}
\details{
Displays histograms showing the distribution of minor allele
frequency both overall, and for each population.
}
\author{
Arthur Georges (Post to \url{https://groups.google.com/d/forum/dartr})
}
