% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.report.pa.r
\name{gl.report.pa}
\alias{gl.report.pa}
\title{Report private alleles (and fixed allelic differences) per pair of populations}
\usage{
gl.report.pa(x = gl, nmin = 10, verbose = 3)
}
\arguments{
\item{x}{-- name of the genlight object containing the SNP data broken down by population [default gl]}

\item{nmin}{-- minimum sample size for a population to be included in the analysis [default 10]}

\item{verbose}{-- verbosity: 0, silent or fatal errors; 1, begin and end; 2, progress log ; 3, progress and results summary; 5, full report [default 3]}
}
\value{
A dataframe giving population names, sample sizes, number of fixed allelic differences, number of private alleles held by population 1,
number of private alleles held by population 2, total private alleles, and mean allelic difference between the two populations.
}
\description{
This function reports separates the genlight object by populations and reports fixed allelic differences, private alleles 
and the mean absolute allele frequency differences between pairs of populations.
}
\details{
A fixed difference at a locus occurs when the two focal populations share no alleles. This is a symmetric measure. A private allele is an allele
that is found in one population but not in the other. The distance defined by a count of private alleles between two populations is not symmetric.

The table below shows a cross tablulation of possible cases of allele frequencies between two populations 
(0=homozygote for Allele 1,x= both Alleles are present, 1=homozygote for Allele 2)

p: cases where there is a private allele in pop1 compared to pop2 (but not vice versa)

f: cases where there is a fixed allelic difference between pop1 and pop2.

\tabular{ccccc}{ 
\tab\tab \tab \emph{pop1}\tab\cr
\tab\tab \strong{0} \tab   \strong{x}  \tab  \strong{1}\cr
\tab     \strong{0}\tab -  \tab  p \tab  p,f\cr
 \emph{pop2} \tab \strong{x}\tab -  \tab- \tab -\cr
\tab \strong{1} \tab p,f\tab p \tab   -\cr
}
}
\examples{
gl.report.pa(testset.gl,nmin=11)

}
\author{
Bernd Gruber (Post to \url{https://groups.google.com/d/forum/dartr})
}
