% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.report.rdepth.r
\name{gl.report.rdepth}
\alias{gl.report.rdepth}
\title{Report summary of Read Depth calculated as the sum of AvgCountRef and AvgCountSnp for each locus in a genlight {adegenet} object}
\usage{
gl.report.rdepth(x, plot = FALSE, smearplot = FALSE)
}
\arguments{
\item{x}{-- name of the genlight object containing the SNP data [required]}

\item{plot}{if TRUE, will produce a histogram of call rate [default FALSE]}

\item{smearplot}{if TRUE, will produce a smearplot of individuals against loci [default FALSE]}
}
\value{
-- Tabulation of REad Depth against Threshold
}
\description{
SNP datasets generated by DArT report AvgCountRef and AvgCountSnp as counts of sequence tags for the reference and alternate alleles respectively.
These can be added for an index of Read Depth. Filtering on Read Depth can be on the basis of loci with exceptionally low counts, or
loci with exceptionally high counts.
}
\details{
A histogram and or a smearplot can be requested. Note that the smearplot is computationally intensive, and will take time to 
execute for large datasets.
}
\examples{
gl.report.rdepth(testset.gl)
}
\author{
Arthur Georges (Post to \url{https://groups.google.com/d/forum/dartr})
}
