% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.report.repavg.r
\name{gl.report.repavg}
\alias{gl.report.repavg}
\title{Report summary of RepAvg, repeatability averaged over both alleles for each locus in a genlight {adegenet} object}
\usage{
gl.report.repavg(x, verbose = 3)
}
\arguments{
\item{x}{-- name of the genlight object containing the SNP data [required]}

\item{verbose}{-- verbosity: 0, silent or fatal errors; 1, begin and end; 2, progress log ; 3, progress and results summary; 5, full report [default 3]}
}
\value{
-- Tabulation of repAvg against prospective Thresholds
}
\description{
SNP datasets generated by DArT have in index, RepAvg, generated by reproducing the data independently for 30% of loci.
RepAvg is the proportion of alleles that give a repeatable result, averaged over both alleles for each locus.
}
\details{
A histogram is produced to aid in selecting a threshold.
}
\examples{
gl.report.repavg(testset.gl)
}
\author{
Arthur Georges (Post to \url{https://groups.google.com/d/forum/dartr})
}
