% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.report.secondaries.r
\name{gl.report.secondaries}
\alias{gl.report.secondaries}
\title{Report loci containing secondary SNPs in a genlight \{adegenet\} object}
\usage{
gl.report.secondaries(x, plot = FALSE, smearplot = FALSE,
  verbose = 0)
}
\arguments{
\item{x}{-- name of the genlight object containing the SNP data [required]}

\item{plot}{-- if TRUE, will produce a frequency plot the number of SNPs per sequence tag [default = FALSE]}

\item{smearplot}{if TRUE, will produce a smearplot of individuals against loci [default FALSE]}

\item{verbose}{level of verbosity. verbose=0 is silent, verbose=1 returns more detailed output during conversion.}
}
\value{
A genlight object of loci with multiple SNP calls
}
\description{
SNP datasets generated by DArT include fragments with more than one SNP (that is, with secondaries) and record them separately with the same CloneID (=AlleleID).
These multiple SNP loci within a fragment are likely to be linked, and so you may wish to remove secondaries.
}
\details{
A histogram and or a smearplot can be requested. Note that the smearplot is computationally intensive, and will take time to 
execute on large datasets.
}
\examples{
gl.report.secondaries(testset.gl)
}
\author{
Arthur Georges (Post to \url{https://groups.google.com/d/forum/dartr})
}
