% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.run.EMIBD9.R
\name{gl.run.EMIBD9}
\alias{gl.run.EMIBD9}
\title{run program EMIBD9}
\usage{
gl.run.EMIBD9(
  x,
  Inbreed = 1,
  GtypeFile = "EMIBD9_Gen.dat",
  OutFileName_par = "MyData.par",
  OutFileName = "EMIBD9_Res.ibd9",
  ISeed = 52
)
}
\arguments{
\item{x}{Name of the genlight object containing the SNP data [required].}

\item{Inbreed}{A Boolean, taking values 0 or 1 to indicate inbreeding is not
and is allowed in estimating IBD coefficients [default 1].}

\item{OutFileName_par}{A string, giving the path and name of the parameter
file [default "MyData.par"].}

\item{OutFileName}{A string, giving the path and name of the output file 
[default "EMIBD9_Res.ibd9"].}

\item{GtypeFile:}{A string, giving the path and name of the genotype file 
[default "EMIBD9_Gen.dat"].}

\item{ISeed:}{An integer used to seed the random number generator [default 52].}
}
\value{
Dataframe with pairwise results of percentage of same genotypes 
between two individuals and the number of loci used in the comparison.
}
\description{
run program EMIBD9
}
\details{
Download the program from here:

https://www.zsl.org/about-zsl/resources/software/emibd9

For Windows, install the program and then move the following files to your
 working directory: "EM_IBD_P.exe", "impi.dll" and "libiomp5md.dll".

For Mac move the file "EM_IBD_P" to the working directory.
}
\examples{
\dontrun{
t1 <- gl.filter.allna(platypus.gl)
res_rel <- gl.run.EMIBD9(t1)
}
}
\author{
Custodian: Luis Mijangos -- Post to
\url{https://groups.google.com/d/forum/dartr}
}
