% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.select.colors.r
\name{gl.select.colors}
\alias{gl.select.colors}
\title{Select colors from one of several palettes and output as a vector}
\usage{
gl.select.colors(
  x = NULL,
  library = NULL,
  palette = NULL,
  ncolors = NULL,
  select = NULL,
  verbose = NULL
)
}
\arguments{
\item{x}{Optionally, provide a gl object from which to determine the number of populations [default NULL]}

\item{library}{Name of the colour library to be used [default scales::hue_pl]}

\item{palette}{Name of the colour palette to be pulled from the specified library [default is library specific]}

\item{ncolors}{number of colours to be displayed and returned [default 9]}

\item{select}{select the colours to retain in the output vector [default NULL]}

\item{verbose}{-- verbosity: 0, silent or fatal errors; 1, begin and end; 2, progress log ; 3, progress and results summary; 5, full report [default 2 or as specified using gl.set.verbosity]}
}
\value{
A vector with the requried number of colours
}
\description{
This script draws upon a number of specified colour libraries to extract a vector of colours for plotting, where
the script that follows has a colour parameter expecting a vector of colours.
}
\details{
The available colour libraries include base R, RColorBrewer, gr.palette (grDevice palette) and gr.hcl (grDevice hcl).
Each of these libraries have a set of palettes, which will be listed if the nominated palette is not specified. A
default palette will then be chosen.

The colour palette will be displayed in the graphics window for the requested number of colours (or 9 if not specified),
and the vector of colours returned for later use.

The select parameter can be used to select colours from the specified ncolors. For example, select=c(1,1,3) will
select colour 1, 1 again and 3 to retain in the final vector. This can be useful for fine-tuning colour selection, and
matching colours and shapes.
}
\examples{
# SET UP DATASET 
gl <- testset.gl
levels(pop(gl))<-c(rep("Coast",5),rep("Cooper",3),rep("Coast",5),
rep("MDB",8),rep("Coast",7),"Em.subglobosa","Em.victoriae")
# EXAMPLES -- SIMPLE
colors <- gl.select.colors()
colors <- gl.select.colors(library="brewer",palette="Spectral",ncolors=6)
colors <- gl.select.colors(library="baseR",palette="terrain.colors",ncolors=6)
colors <- gl.select.colors(library="baseR",palette="rainbow",ncolors=12)
colors <- gl.select.colors(library="gr.hcl",palette="RdBu",ncolors=12)
colors <- gl.select.colors(library="gr.palette",palette="Pastel 1",ncolors=6)
# EXAMPLES -- SELECTING COLOURS
colors <- gl.select.colors(library="baseR",palette="rainbow",ncolors=12,select=c(1,1,1,5,8))
# EXAMPLES -- CROSS-CHECKING WITH A GENLIGHT OBJECT
colors <- gl.select.colors(x=gl,library="baseR",palette="rainbow",ncolors=12,select=c(1,1,1,5,8))

}
\seealso{
\code{\link{gl.select.shapes}}

Other Exploration/visualisation functions: 
\code{\link{gl.pcoa.plot.beta}()},
\code{\link{gl.pcoa.plot}()},
\code{\link{gl.plot}()},
\code{\link{gl.select.shapes}()}
}
\author{
Custodian: Arthur Georges -- Post to \url{https://groups.google.com/d/forum/dartr}
}
\concept{Exploration/visualisation functions}
