% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.sim.emigration.r
\name{gl.sim.emigration}
\alias{gl.sim.emigration}
\title{Simulate emigration between populations}
\usage{
gl.sim.emigration(x, perc.mig = NULL, emi.m = NULL, emi.table = NULL)
}
\description{
A simple functions that allows to exchange individuals of genlight object to be swapped between populations
}
