% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.sim.WF.run.r
\name{gl.sim.WF.run}
\alias{gl.sim.WF.run}
\title{Wright-Fisher simulations}
\usage{
gl.sim.WF.run(
  file_var,
  ref_table,
  x = NULL,
  number_iterations = 1,
  every_gen = 5,
  store_phase1 = FALSE,
  interactive_vars = TRUE,
  seed = NULL,
  parallel = FALSE,
  n.cores = NULL,
  verbose = NULL
)
}
\arguments{
\item{file_var}{[required if interactive_vars = TRUE].}

\item{ref_table}{[required].}

\item{x}{Name of the genlight object containing the SNP data [default NULL].}

\item{number_iterations}{[default 1].}

\item{every_gen}{[default 5].}

\item{store_phase1}{[default FALSE].}

\item{interactive_vars}{[default TRUE].}

\item{seed}{[default NULL].}

\item{parallel}{[default FALSE].}

\item{n.cores}{[default NULL].}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2,
progress log; 3, progress and results summary; 5, full report
[default 2, unless specified using gl.set.verbosity].}
}
\value{
Returns unaltered genlight object
}
\description{
Forward simulattions
}
\details{
system.file('extdata', 'sim_variables.csv', package ='dartR')
}
\examples{

}
\seealso{
\code{\link{gl.sim.WF.table}}

Other simulation functions: 
\code{\link{gl.sim.WF.table}()}
}
\author{
Custodian: Luis Mijangos -- Post to
\url{https://groups.google.com/d/forum/dartr}
}
\concept{simulation functions}
