% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.sim.WF.table.r
\name{gl.sim.WF.table}
\alias{gl.sim.WF.table}
\title{Create reference table for running gl.sim.WF.run}
\usage{
gl.sim.WF.table(
  file_var,
  x = NULL,
  file_targets_sel = NULL,
  file_r_map = NULL,
  interactive_vars = TRUE,
  verbose = NULL
)
}
\arguments{
\item{file_var}{[required if interactive_vars = TRUE].}

\item{x}{Name of the genlight object containing the SNP data [default NULL].}

\item{file_targets_sel}{[default NULL].}

\item{file_r_map}{[default NULL].}

\item{interactive_vars}{[default TRUE].}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2,
progress log; 3, progress and results summary; 5, full report
[default 2, unless specified using gl.set.verbosity].}
}
\value{
Returns unaltered genlight object
}
\description{
Reference table for gl.sim.WF.run
}
\details{
system.file('extdata', 'ref_variables.csv', package = 'dartR')
}
\examples{

}
\seealso{
\code{\link{gl.sim.WF.run}}

Other simulation functions: 
\code{\link{gl.sim.WF.run}()}
}
\author{
Custodian: Luis Mijangos -- Post to
\url{https://groups.google.com/d/forum/dartr}
}
\concept{simulation functions}
