% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.spatial.ac.multiPlot.r
\name{gl.spatial.ac.multiPlot}
\alias{gl.spatial.ac.multiPlot}
\title{plot of mulitple Spatial autocorrelation analyses}
\usage{
gl.spatial.ac.multiPlot(
  l.spa,
  grp.var.name = "Pop",
  plot_theme = theme_classic(),
  outpath = tempdir(),
  out_file_rootname = "ac.multiPlot",
  verbose = NULL
)
}
\arguments{
\item{l.spa}{The list of multiple gl.spatial.autoCorr outputs}

\item{grp.var.name}{The name to be used for the grouping variable}

\item{outpath}{The path where to save the output}

\item{out_file_rootname}{The name of the root of the output file}
}
\value{
Returns the ggplot object
}
\description{
This function takes multiple  outputs of gl.spatial.autoCorr 
and plots them to enable their graphical comparisons. If upper and lower estimates
are present in the outputs, these are plotted as error bars around the r estimates.
}
\details{
The input is a list. If the list has names, these are used 
in the legend. If names are absent, these are internally created assuming that each 
element of the list represents different populations.
}
\examples{
 # Select one pop only
plat_Tent <- gl.keep.ind(platypus.gl, 
ind.list = platypus.gl@ind.names[pop(platypus.gl) == "TENTERFIELD"], 
mono.rm = TRUE)
# Compute a simple distance matrix and reverse it so that correlated values 
# indicated more similar individuals as we are used to see plots in GenAleEx
gd <- 1 - as.matrix(gl.dist.ind(plat_Tent, method = "Simple", plot.out = FALSE))
# Replace the diagonal with zeros
diag(gd) <- 0
# Compute the geographical distance matrix
ggd <- as.matrix(dist(plat_Tent@other$latlon))

# Spatial autocorrelation
spa <- gl.spatial.autoCorr(gd, ggd, bins = 3, reps = 100, 
permutation = TRUE, bootstrap = TRUE)

# Alternatively, use Smouse distance from PopGneReport package
smouse.plat<-as.matrix(PopGenReport::gd.smouse(gl2gi(plat_Tent), verbose=T))
spa.sm <- gl.spatial.autoCorr(smouse.plat, ggd, bins = 3, reps = 100, 
permutation = TRUE, bootstrap = TRUE)
}
\author{
Carlo Pacioni
}
