% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.spatial.autoCorr.r
\name{gl.spatial.autoCorr}
\alias{gl.spatial.autoCorr}
\title{Spatial autocorrelation following Smouse and Pekall 1999}
\usage{
gl.spatial.autoCorr(
  GD,
  GGD,
  bins = 1,
  reps = 100,
  permutation = FALSE,
  bootstrap = FALSE,
  plot_theme = theme_classic(),
  save2tmp = FALSE,
  verbose = NULL
)
}
\arguments{
\item{GD}{a matrix of individual pairwise genetic distances. In principle 
any other squared distance matrix can be used. see example}

\item{GGD}{A geographical distance matrix, based on the coordinates of
individuals. This is typically an Euclidean distance but it can be 
any meaningful (geographical) distance metrics.}

\item{permutation}{Whether permutation calculations for the null hypothesis of 
no spatial structure should be carried out}

\item{bootstrap}{Whether bootstrap calculations to compute the 95% confidence intervals 
around r should be carried out}
}
\value{
Returns a data frame with the following columns:
\enumerate{
\item Bin  The distance classes
\item N The number of pairwise comparisons within each distance class
\item r.uc The uncorrected autocorrelation coefficient
\item Correction the correction 
\item r The corrected autocorrelation coefficient
\item L.r The corrected autocorrelation coefficient lower limit 
(if \code{bootstap = TRUE})
\item U.r The corrected autocorrelation coefficient upper limit 
(if \code{bootstap = TRUE})
\item L.r.null.uc The uncorrectedlower limit for the null hypothesis of no spatial 
autocorrelation (if \code{permutation = TRUE}) 
\item U.r.null.uc  The uncorrected upper limit for the null hypothesis of no spatial 
autocorrelation (if \code{permutation = TRUE})
\item L.r.null The corrected lower limit for the null hypothesis of no spatial 
autocorrelation (if \code{permutation = TRUE})
\item U.r.null The corrected upper limit for the null hypothesis of no spatial 
autocorrelation (if \code{permutation = TRUE})
\item p.one.tail The p value of the one tail statistical test
}
}
\description{
Global spatial autocorrelation is a multivariate approach combining all loci
into a single analysis. The autocorrelation coefficient r is calculated for
each pairwise genetic distance pairs for all specified distance classes. For
more information see Smouse and Peakall 1999, Peakall et a. 2003 and Smouse
et al. 2008.
}
\details{
The code of this function execute behind the scene a modified 
version of \code{spautocorr} from the package 
\code{PopGenReport}, and it also includes 
bootstraps to compute the 95% confidence intervals around the r 
estimates, the one-tail test, and the correction factor described by 
Peakall et al 2003.

If \code{bins} is of length = 1 it is interpreted as the number of (even) 
bins to use. In this case the starting point is always the minimum 
value in the distance matrix, and the last is the maximum. If it is a numeric
vector, it is interpreted as the braking points. In this case, the first has 
to be the lowest value, and the last has to be the highest. The number of 
bins will be  \code{length(bins) - 1}.

The permutation constructs the 95% confidence intervals around the null 
hypothesis of no spatial structure (this is a two-tail test). The same data
are also used to calculate the probability of the one-tail test (See reference 
below for details).
}
\references{

}
\author{
Carlo Pacioni & Bernd Gruber
}
