% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.stockR.r
\name{gl.stockR}
\alias{gl.stockR}
\title{Generate a SNP matrix from a genlight \{adegenet\} object for subsequent use in R package
stockR.}
\usage{
gl.stockR(x, verbose = 2)
}
\arguments{
\item{x}{-- name of the genlight object containing SNP genotypes [required]}

\item{verbose}{-- verbosity: 0, silent or fatal errors; 1, begin and end; 2, progress log ; 3, progress and results summary; 5, full report [default 2]}
}
\value{
A matrix with the SNP scores in the form expected by stockR
}
\description{
The script extracts the SNP data from the gl object as a matrix and transposes it
to comply with the format expected by stockR.
}
\details{
The script returns the transposed SNP matrix
}
\examples{
   gl <- gl.stockR(testset.gl)
}
\author{
Arthur Georges (Post to \url{https://groups.google.com/d/forum/dartr})
}
