% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.tree.nj.r
\name{gl.tree.nj}
\alias{gl.tree.nj}
\title{Output an nj tree to summarize genetic similarity among populations}
\usage{
gl.tree.nj(gl, type = "phylogram", outgroup = NULL, labelsize = 0.7)
}
\arguments{
\item{gl}{-- Name of the genlight object containing the SNP data or a genind object containing presence absence data [required]}

\item{type}{-- Type of dendrogram phylogram|cladogram|fan|unrooted [Default Phylogram]}

\item{outgroup}{-- Vector containing the population names that are the outgroups [Default NULL]}

\item{labelsize}{-- Size of the labels as a proportion of the graphics default [Default 0.7]}
}
\value{
A tree file of type phylo
}
\description{
This function is a wrapper for the nj\{ape\} function applied to Euclidian
distances calculated from the genlight object.
}
\examples{
gl.tree.nj(testset.gl,type="fan")
}
\author{
Arthur Georges (glbugs@aerg.canberra.edu.au)
}
