% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl2bayescan.r
\name{gl2bayescan}
\alias{gl2bayescan}
\title{Convert a genlight object to format suitable for input to Bayescan}
\usage{
gl2bayescan(x, outfile = "bayescan.txt", outpath = tempdir(),
  verbose = 2)
}
\arguments{
\item{x}{-- name of the genlight object containing the SNP data [required]}

\item{outfile}{-- file name of the output file (including extension) [default bayescan.txt]}

\item{outpath}{-- path where to save the output file [default tempdir(), mandated by CRAN]. Use outpath=getwd() or outpath="." when calling this function to direct output files to your working directory.}

\item{verbose}{-- specify the level of verbosity: 0, silent, fatal errors only; 1, flag function begin and end; 2, progress log ; 3, progress and results summary; 5, full report [default 2]}
}
\description{
The output text file contains the snp data and relevant BAyescan command lines to guide input.
}
\examples{
gl2bayescan(testset.gl)
}
\references{
Foll M and OE Gaggiotti (2008) A genome scan method to identify selected loci appropriate for both dominant and codominant markers: A Bayesian perspective. Genetics 180: 977-993.
}
\author{
Arthur Georges (Post to \url{https://groups.google.com/d/forum/dartr})
}
