% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl2dartR.R
\name{gl2dartR}
\alias{gl2dartR}
\title{Convert a genlight object to a dartR object}
\usage{
gl2dartR(x, filename = NULL, file.path = getwd())
}
\arguments{
\item{x}{An object of class `genlight` to be converted.}

\item{save}{Logical. If `TRUE`, the converted object is saved as an file.}
}
\value{
The input object with class changed to `"dartR"` and its package attribute set to `"dartR.base"`.
}
\description{
This function converts a `genlight` object into a `dartR` object by 
changing its class attribute.
It is used to convert legacy data sets to the new dartR format.
}
\examples{
gl <- glSim(10, 100, ploidy = 2, ind.names=1:10, loc.names=1:100)  # Simulating a genlight object
gl <- gl2dartR(gl)
pop(gl)<- rep("A",10)
}
