% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl2demerelate.r
\name{gl2demerelate}
\alias{gl2demerelate}
\title{Create a dataframe suitable for input to package \{Demerelate\} from a genlight \{adegenet\} object}
\usage{
gl2demerelate(gl)
}
\arguments{
\item{gl}{-- name of the genlight object containing the SNP data [required]}
}
\value{
A dataframe suitable as input to package \{Demerelate\}
}
\description{
Create a dataframe suitable for input to package \{Demerelate\} from a genlight \{adegenet\} object
}
\examples{
df <- gl2demerelate(testset.gl)
}
\author{
Arthur Georges (Post to \url{https://groups.google.com/d/forum/dartr})
}
