% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl2fasta.nohets.r
\name{gl2fasta.nohets}
\alias{gl2fasta.nohets}
\title{Concatenates DArT trimmed sequences after removing heterozygous loci by random allele assignment}
\usage{
gl2fasta.nohets(gl, outfile = "output.fasta")
}
\arguments{
\item{gl}{-- name of the DArT genlight object [required]}

\item{outfile}{-- name of the output file (fasta format) [output.fasta]}
}
\value{
A new gl object with all loci rendered homozygous
}
\description{
This script randomly allocates haplotypes in a DArT genlight object in preparation for analyses
that expect only a single character state for each individual at each locus. For example, many 
phylogenetic analyses have difficulty accommodating heterozygosity. Under this approach, such an
analysis becomes one of determining the phylogeny of composite haplotypes.
}
\details{
The haplotypes are generated by randomly assigning one or the other SNP variant to the individual 
in instances of homozygosity, then concatenating the haplotypes across loci to generate 
a single haplotype to be used in the analysis. If repeated runs yield the same topology and 
roughly the same branch lengths, then the resultant haplotype tree can be considered to also 
be the species tree.

The script writes out the composite haplotypes for each individual as a fastA file. Requires
trimmed sequences to be among the locus metrics.
}
\examples{
\dontrun{
gl2fasta.nohets(gl, outfile="test.fasta")
}
}
\author{
Arthur Georges (glbugs@aerg.canberra.edu.au)
}

