% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl2fasta_old.r
\name{gl2fasta_old}
\alias{gl2fasta_old}
\title{Export DArT genlight object \{adegenet\} to fastA format}
\usage{
gl2fasta_old(gl, outfile = "genlight.fasta")
}
\arguments{
\item{gl}{-- genlight object}

\item{outfile}{-- filename of the output fastA file [default genlight.fasta]}
}
\description{
Concatenates DArT trimmed sequences after replacing allelic states at heterozygous loci with standard ambiguity codes.
}
\details{
This script generates haplotypes from SNP genotypes in a DArT genlight object in preparation for analyses
that expect only a single character state for each individual at each locus. For example, many 
phylogenetic analyses have difficulty accommodating heterozygosity. Under this approach, heterozytotes are represented
by ambiguity codes then the sequence tags are concatenated across loci to generate
a single haplotype to be used in the analysis.

The script writes out the composite haplotypes for each individual as a fastA file. Requires
trimmed sequences to be among the locus metrics.
}
\author{
Bernd Gruber (glbugs@aerg.canberra.edu.au)
}

