% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl2gi.r
\name{gl2gi}
\alias{gl2gi}
\title{Genlight to genind conversion}
\usage{
gl2gi(gl)
}
\arguments{
\item{gl}{a genlight object}
}
\value{
A genind object, with all slots filled.
}
\description{
Convert a genlight object to a genind object.
}
\details{
Note that \code{gl2gi(gi2gl(x))} may not return the original
object due to ordering of major alleles.
}
\seealso{
\code{\link{gi2gl}}, \code{\link[adegenet]{df2genind}}
}
\author{
Bernd Gruber (bugs? Post to \url{https://groups.google.com/d/forum/dartr})
}
