% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl2migraten.r
\name{gl2migraten}
\alias{gl2migraten}
\title{Converts a genlight object into a migrate-n SNP input file}
\usage{
gl2migraten(
  x,
  plink.cmd = "plink",
  plink.path = "path",
  plink.flags = NULL,
  outfile = "migrateHapMap",
  outpath = tempdir(),
  verbose = NULL
)
}
\arguments{
\item{x}{Name of the genlight object containing the SNP data [required].}

\item{outfile}{File name of the output file [default 'migrateHapMap'].}

\item{outpath}{Path where to save the output file
[default tempdir(), mandated by CRAN]. Use outpath=getwd() or outpath='.'
 when calling this function to direct output files to your working directory.
 @param plink.flags Character vector with additioanl flags to pass to plink 
 during the conversion (e.g "--dog") [default NULL]
 @inheritParams utils.plink.run}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2,
progress log; 3, progress and results summary; 5, full report
[default 2 or as specified using gl.set.verbosity].}
}
\description{
This function exports a genlight object into a migrate-n SNP input file
}
\details{
The input file generated is compatible with migrate-n (Beerli ) 
v3.7.2 onward. For the (limited) testing I have done with migrate-n and SNP data, 
migrate-n seems to be able to handle up to 20k SNPs (v3.7.2) and 3-5k SNPs in v4
and later. That seems to depend on some memory issues that Peter Beerli may be able
to resolve (or may have resolved by the time you use this function).

This function needs PLINK installed, which is used for intermediate data manipulation.

migrate-n and its manual can be downloaded from:
\url{https://peterbeerli.com/migrate-html5/index.html}
}
\examples{
\donttest{
}
}
\references{
Beerli, P. (2004). Effect of unsampled populations on the estimation of 
population sizes and migration rates between sampled populations. 
Molecular Ecology 13, 827-836.

Beerli, P. and Felsenstein, J. (2001). Maximum likelihood estimation of a 
migration matrix and effective population sizes in n subpopulations by using 
a coalescent approach. Proceedings of the National Academy of Sciences 98, 4563-4568. 
doi: 10.1073/pnas.081068098.

Beerli, P. and Palczewski, M. (2010). Unified framework to evaluate panmixia 
and migration direction among multiple sampling locations. Genetics 185, 313-326.
}
\author{
Custodian: Carlo Pacioni (Post to
 \url{https://groups.google.com/d/forum/dartr})
}
