% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl2nexusCompact.r
\name{gl2nexusCompact}
\alias{gl2nexusCompact}
\title{Concatenates DArT trimmed sequences and outputs a nexus file (for BEAST).}
\usage{
gl2nexusCompact(
  gl,
  method = 1,
  outfile = "output.nex",
  outpath = tempdir(),
  min_nSNPs = 3,
  probar = TRUE
)
}
\arguments{
\item{gl}{-- name of the DArT genlight object [required]}

\item{method}{-- 1 | 2. Type method=0 for a list of options. Only mathod=1 implemented at the moment [method=1]}

\item{outfile}{-- name of the output file (fasta format) [output.fasta]}

\item{outpath}{-- path where to save the output file (set to tempdir by
default)}

\item{min_nSNPs}{Minimum number of SNPs to retain a locus [min_nSNPs=3]}

\item{probar}{-- if TRUE, a progress bar will be displayed for long loops
[default = TRUE]}
}
\value{
A new gl object with all loci rendered homozygous
}
\description{
Concatenated sequence tags are useful for phylogenetic methods where
information on base frequencies and transition and transversion ratios are
required (for example, Maximum Liklihood methods). Where relevant,
heterozygous loci are resolved before concatenation by random allele
assignment.
}
\details{
This function generates and aligment for use in BEAST or similar software
that accept a nexus file with sequence data. In these analyses, it is often
preferible to limit the analysis to reads that have several SNPs. For this
reason, the argument \code{min_nSNPs} can be used to retain only loci with a
minimum number of SNPs. Loci with multiple SNPs (i.e. secondaries) are
resolved in one sequence only.

Currently only method 1 is employed

Method 1 -- the heterozyous state is resolved by randomly assigning one or
the other SNP variant to the individual. The resultant sequence fragments are
concatenated across loci to generate a single composite haplotype to be used
in subsequent phylogenetic analyses.

Trimmed sequences for which the SNP has been trimmed out, rarely, by adaptor
mis-identity are deleted.

A charset block is appended at the end of the nexus file to facilitate
partitioning of the sequences.

The script writes out the composite haplotypes for each individual as a nexus
file. Requires 'TrimmedSequence' to be among the locus metrics
(\code{@other$loc.metrics}) and information of the type of alleles (slot
loc.all e.g. "G/A") and the position of the SNP in slot position of the
```genlight``` object (see testset.gl@position and testset.gl@loc.all for how
to format these slots.)
}
\examples{
gl <- gl.filter.repavg(testset.gl,t=1)
gl <- gl.filter.callrate(testset.gl,t=.98)
glBEAST <- gl2nexusBEAST(gl)
}
\author{
Carlo Pacioni (Post to \url{https://groups.google.com/d/forum/dartr})
}
