% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl2plink.r
\name{gl2plink}
\alias{gl2plink}
\title{Convert Genlight to Plink}
\usage{
gl2plink(gl, file = "plink.csv")
}
\arguments{
\item{gl}{a genlight object.}

\item{file}{file or connection to write to.}
}
\value{
Returns an object of class \code{snp} to be used with \pkg{SNPassoc}.
}
\description{
Converts a genlight object to plink format and writes it to a file
or connection.
}
\author{
Bernd Gruber (bugs? Post to \url{https://groups.google.com/d/forum/dartr})
}
