% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl2related.r
\name{gl2related}
\alias{gl2related}
\title{Create a dataframe suitable for input to package \{related\}, from a genlight (SNP) \{adegenet\} object}
\usage{
gl2related(gl)
}
\arguments{
\item{gl}{-- name of the genlight object containing the SNP data or a genind object containing presence absence data [required]}
}
\value{
A dataframe suitable as input to package \{related\}
}
\description{
Create a dataframe suitable for input to package \{related\}, from a genlight (SNP) \{adegenet\} object
}
\examples{
\dontrun{
#install package via
#install.packages("related", repos="http://R-Forge.R-project.org")
df <- gl2related(gl)
}
}
\author{
Arthur Georges (glbugs@aerg.canberra.edu.au)
}

