% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl2sa.r
\name{gl2sa}
\alias{gl2sa}
\title{Convert genlight objects to the format used in the SNPassoc package}
\usage{
gl2sa(x)
}
\arguments{
\item{x}{-- genlight object}
}
\value{
Returns an object of class 'snp' to be used with \pkg{SNPassoc}
}
\description{
This function exports a genlight object into a SNPassoc object. See \link[SNPassoc]{setupSNP}
}
\examples{
gl2plink(testset.gl)
}
\author{
Bernd Guber (glbugs@aerg.canberra.edu.au)
}
