% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl2sfs.R
\name{gl2sfs}
\alias{gl2sfs}
\title{Converts a genlight object into a sfs input file}
\usage{
gl2sfs(x, n.invariant = 0, outfile_root = "gl2sfs", outpath = tempdir())
}
\arguments{
\item{x}{Name of the genlight object containing the SNP [required].}

\item{n.invariant}{An estimate of the number of invariant sequence tags used
to adjust the heterozygosity rate [default 0].}

\item{outfile_root}{The root of the name of the output file}

\item{outpath}{Path where to save the output file
[default tempdir(), mandated by CRAN]. Use outpath=getwd() or outpath='.'
 when calling this function to direct output files to your working directory.}
}
\value{
A list with two elements: the DAF and MAF.
}
\description{
The output of this function is suitable for analysis in fastsimcoal2 or dada.
}
\details{
It saves a derived sfs, assuming that the reference allele is the ancestral,
and a MAF sfs.

At this stage this function caters only for diploid organisms, for samples
from one population only, and for genotypes without missing data.

If no invariant site estimates is provided (with \code{n.invariant}), it will
estimate the number of invariant site from the sequenced tags (but this will
be a biased estimates as it doesn't take into account the invariant tags).
Note that the invariant sites can be estimated with
\code{gl.report.secondaries}.

It expects a dartR formatted genlight object, but it should also  work with
other genlight objects. If \code{n.invariant == 0} the genlight object needs
to have a \code{data.frame} in \code{other} called loc.metrics with a column
named \code{TrimmedSequence} that contains a character string whose length
corresponds to the number of sites of the allele. This information is used to
obtained the number of non-polymorphic sites.
}
\references{
Excoffier L., Dupanloup I., Huerta-Sánchez E., Sousa V. C. and
 Foll M. (2013) Robust demographic inference from genomic and SNP data. PLoS
 genetics 9(10)
}
\seealso{
\code{\link{gl.report.heterozygosity}},
 \code{\link{gl.report.secondaries}}, \code{\link{utils.n.var.invariant}}
}
\author{
Custodian: Carlo Pacioni (Post to
 \url{https://groups.google.com/d/forum/dartr})
}
