% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl2shp.R
\name{gl2shp}
\alias{gl2shp}
\title{Convert genlight objects to ESRI shapefiles or kml files}
\usage{
gl2shp(x, type = "shp",
  proj4 = "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs",
  outfile = "gl", outpath = tempdir(), verbose = NULL)
}
\arguments{
\item{x}{-- name of the genlight object containing the SNP data and location data, lat longs  [required]}

\item{type}{-- type of output "kml" or "shp" [default 'shp']}

\item{proj4}{-- proj4string of data set (see spatialreference.org for projections) [default WGS84]}

\item{outfile}{-- name (path) of the output shape file [default 'gl']. shp extension is added automatically.}

\item{outpath}{-- path where to save the output file [default tempdir(), mandated by CRAN]. Use outpath=getwd() or outpath="." when calling this function to direct output files to your working directory.}

\item{verbose}{-- specify the level of verbosity: 0, silent, fatal errors only; 1, flag function begin and end; 2, progress log ; 3, progress and results summary; 5, full report [default 2]}
}
\description{
This function exports cordinates in a genlight object to a point shape file (including also individual meta data if available).
Coordinates are provided under x@other$latlong and assumed to be in WGS84 coordinates, if not proj4 string is provided.
}
\examples{
gl2shp(testset.gl)
}
\author{
Bernd Guber (Post to \url{https://groups.google.com/d/forum/dartr})
}
