% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl2structure.r
\name{gl2structure}
\alias{gl2structure}
\title{Converts genlight objects to STRUCTURE formated files}
\usage{
gl2structure(gl, pops = FALSE, outfile = "gl.str", outpath = getwd(),
  v = 1)
}
\arguments{
\item{gl}{-- genlight containing lat longs  [required]}

\item{pops}{-- switch if population column should be added}

\item{outfile}{-- name (path) of the output shape file}

\item{outpath}{-- path of the output file. Default is to tempdir(). If to be saved in the current working directory change to "."}

\item{v}{-- verbosity: if v=0 no output, v=1 reports name and path of output file. default 1}
}
\description{
This function exports genlight objects to STRUCTURE formatted files (be aware there is a gl2faststruture version as well). It is based on the code provided by Lindsay Clark (originally for genind objects: [Lindsay happy to link to your github repo if you want]) and this function is basically a wrapper around his genind2structure function.
}
\examples{
\dontrun{
gl2structure(testset.gl)
}
}
\author{
Lindsay V. Clark [your email], wrapper by Bernd Gruber
}
