% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.sims.r
\name{interactive_sim_run}
\alias{interactive_sim_run}
\title{Shiny app for the input of the simulations variables}
\usage{
interactive_sim_run(env_fun)
}
\arguments{
\item{env_fun}{Environment of the calling function}
}
\description{
Shiny app for the input of the simulations variables
}
\author{
Custodian: Luis Mijangos -- Post to
\url{https://groups.google.com/d/forum/dartr}
}
