% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{landscape.sim}
\alias{landscape.sim}
\title{A simulated rasteer object created to run a landscape genetic example}
\format{raster object}
\usage{
landscape.sim
}
\description{
This a test data set to run a landscape genetics example. The landscapes is a simulated resistance layer.
}
\author{
Bernd Gruber \email{(bernd@aerg.canberra.edu.au}
}
\keyword{datasets}
