% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{platypus.gl}
\alias{platypus.gl}
\title{A genlight object created via the gl.read.dart function}
\format{
genlight object
}
\usage{
platypus.gl
}
\description{
This is a test data set on platypus with 81 individuals, 3 populations and 
1,000 binary SNPs.
}
\author{
Luis Mijangos (bugs? Post to 
\url{https://groups.google.com/d/forum/dartr}
}
\keyword{datasets}
