% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dartR.r
\name{rbind.dartR}
\alias{rbind.dartR}
\title{adjust rbind for dartR}
\usage{
\method{rbind}{dartR}(...)
}
\arguments{
\item{...}{list of dartR objects}
}
\description{
rbind is a bit lazy and does not take care for the metadata (so data in the 
other slot is lost). You can get most of the loci metadata back using
 gl.compliance.check.
}
