% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dartRgenlight.r
\name{rbind.dartRgenlight}
\alias{rbind.dartRgenlight}
\title{adjust rbind for dartRgenlight}
\usage{
\method{rbind}{dartRgenlight}(...)
}
\arguments{
\item{...}{list of dartRgenlight objects}
}
\description{
rbind is a bit lazy and does not take care for the metadata (so data in the other slot is lost). You can get most of the loci metadata back using gl.compliance.check.
}
