% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{set_y_coords}
\alias{set_y_coords}
\title{TreeMix Plotting Functions}
\usage{
set_y_coords(d)
}
\description{
This suite of scripts is provided by the originators of TreeMix and are presented here unaltered, for convenience. The pipeline is
1. Run gl2treemix(gl) to generate a treemix input file.
2. Transfer the file to a unix machine or VM
3. gzip the file
4. Input to treemix following the instructions in the treemix manual
5. Transfer the generated files back to your PC
6. Use the commands in this script to visualize the results (e.g. plottree)
}
\details{
Reference: Pickrell and Pritchard (2012). Inference of population splits and mixtures from genome-wide allele frequency data. PLoS Genetics https://doi.org/10.1371/journal.pgen.1002967
}
