% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.sexlinkage.r
\name{gl.sexlinkage}
\alias{gl.sexlinkage}
\alias{is.sexlinkage}
\alias{print.sexlinkage}
\alias{summary.sexlinkage}
\title{A sexlinkage object}
\usage{
gl.sexlinkage(x, t.het = 0, t.hom = 0, t.abs = 0, verbose = FALSE,
  na.rm = FALSE)

is.sexlinkage(object)

print.sexlinkage(x, ..., system = NULL)

summary.sexlinkage(object, ..., system = NULL, include = c("locus", "count",
  "sequence", "metrics"))
}
\arguments{
\item{x}{-- name of the genlight object containing the SNP data}

\item{t.het}{-- tolerance, that is \code{t.het = 0.05} means that 5% of individuals in the sex expected to be homozygous
can be heterozygous and still be regarded as consistent with a sex specific marker (default 0)}

\item{t.hom}{-- tolerance, that is \code{t.hom = 0.05} means that 5% of individuals in the sex expected to be heterozygous
can be homozygous and still be regarded as consistent with a sex specific marker (default 0)}

\item{t.abs}{-- tolerance to errors when identifying absent individuals. \code{t.abs = 1} treats a single observation for
a given sex as an absence}

\item{verbose}{-- verbosity: logical indicating whether outputs should be printed to the console (default: FALSE)}

\item{na.rm}{-- logical: should NAs in sex assignments be ignored?}

\item{object}{-- a sexlinkage object created with \code{gl.sexlinkage}}

\item{system}{-- a string indicating whether the study organism comes from a \code{'xy'} or \code{'zw'} system}

\item{include}{-- a character vector denoting information to print in \code{summary(sexlinkage)}, any of \code{c('locus', 'count', 'sequence', 'metrics')}}

\item{x}{-- a sexlinkage object created with \code{gl.sexlinkage}}
}
\value{
The list of sex specific loci
}
\description{
Identify loci that are sex linked in specimens in a genlight \{adegenet\} object

Alleles unique to the Y or W chromosome and monomorphic on the X chromosomes will appear in the SNP dataset as 
genotypes that are heterozygotic in all individuals of the heterogametic sex and homozygous in all individuals 
of the homogametic sex.

This script will identify loci with alleles that behave in this way, as putative sex specific SNP markers.

Sex of the individuals for which sex is known with certainty is to be held in the variable x@other$ind.metrics$sex, 
as M for male, F for female, NA otherwise. The script abbreviates the entries here to the first character. So coding of "Female" and "Male" works as well. Character are also converted to upper cases.
}
\examples{
result <- gl.sexlinkage(testset.gl)

# check if an object is a sexlinkage object
  
\dontrun{
is.sexlinkage(object)
}

# Print information about a 'sexlinkage' object

\dontrun{
print(object)
}

# Summarise a 'sexlinkage' object

\dontrun{
summary(object)
}
}
\author{
Arthur Georges (glbugs@aerg.canberra.edu.au)
}
