% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dartR.r
\name{[,dartR,ANY,ANY,ANY-method}
\alias{[,dartR,ANY,ANY,ANY-method}
\title{indexing dartR objects correctly...}
\usage{
\S4method{[}{dartR,ANY,ANY,ANY}(x, i, j, ..., pop = NULL, treatOther = TRUE, quiet = TRUE, drop = FALSE)
}
\arguments{
\item{x}{dartR object}

\item{i}{index for individuals}

\item{j}{index for loci}

\item{...}{other parameters}

\item{pop}{list of populations to be kept}

\item{treatOther}{elements in other (and ind.metrics & loci.metrics) as indexed as well. default: TRUE}

\item{quiet}{warnings are surpessed. default: TRUE}

\item{drop}{reduced to a vector if a single individual/loci is selected. default: FALSE [should never set to TRUE]}
}
\description{
indexing dartR objects correctly...
}
