% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{testset.gl}
\alias{testset.gl}
\title{A genlight object created via the read.dart functions}
\format{genlight object}
\usage{
testset.gl
}
\description{
This is a test data set on turtles. 250 individuals on >30 populations.
}
\author{
Arthur Georges (bugs? Post to \url{https://groups.google.com/d/forum/dartr}
}
\keyword{datasets}
