% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{testset.gs}
\alias{testset.gs}
\title{A genlight object created via the gl.read.silicodart function}
\format{
genlight object
}
\usage{
testset.gs
}
\description{
This is a test data set on turtles. 218 individuals, 255 loci in >30 populations.
}
\author{
Arthur Georges (bugs? Post to \url{https://groups.google.com/d/forum/dartr}
}
\keyword{datasets}
