% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{testset_pop_recode}
\alias{testset_pop_recode}
\title{Recode file to be used with the function.}
\format{csv}
\description{
This test data set is provided to show a typical recode file format.
}
\author{
Arthur Georges \email{(glbugs@aerg.canberra.edu.au}
}
\keyword{datasets}

