% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\name{theme_dartR}
\alias{theme_dartR}
\title{dartR theme}
\usage{
theme_dartR(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\description{
This is the theme used as default for dartR plots.
This function controls all non-data display elements in the plots.
}
\examples{

ggplot(data.frame(dummy=rnorm(1000)),aes(dummy)) +
 geom_histogram() +
 theme_dartR()

}
