% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.assigment.r, R/utils.assignment.r
\name{utils.assignment}
\alias{utils.assignment}
\alias{gl2alleles}
\title{Population assignment probabilities}
\usage{
gl2alleles(gl)

gl2alleles(gl)
}
\arguments{
\item{ind}{Name of the individual to be assigned to a population [required].}

\item{x}{Name of the genlight object containing the SNP data [required].}

\item{inbreeding_par}{The inbreeding parameter (default=0)}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2,
progress log; 3, progress and results summary; 5, full report
[default 2, unless specified using gl.set.verbosity].}
}
\value{
A \code{data.frame} consisting of assignment probabilities for each
 population.

A \code{data.frame} consisting of assignment probabilities.
}
\description{
This function takes one individual and estimates
their probability of coming from individual populations
from multilocus genotype frequencies.

This function takes one individual and estimates
their probability of coming from individual populations
from multilocus genotype frequencies.
}
\details{
This function is a re-implementation of the function multilocus_assignment
 from package gstudio. 
 Description of the method used in this function can be found at:
https://dyerlab.github.io/applied_population_genetics/population-assignment.html

https://dyerlab.github.io/applied_population_genetics/population-assignment.html
}
\examples{
res <- utils.assignment(platypus.gl,ind="T27")
res <- utils.assignment(platypus.gl,ind="T27")
}
\author{
Custodian: Luis Mijangos -- Post to
\url{https://groups.google.com/d/forum/dartr}
}
