% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.check.gl.r
\name{utils.check.gl}
\alias{utils.check.gl}
\title{Utility function to check that the object passed to the function is a genlight object.}
\usage{
utils.check.gl(x, env, verbose = options()$dartR_verbose)
}
\arguments{
\item{x}{Name of the genlight object containing the SNP data or tag presence/absence data (SilicoDArT) [required]}

\item{env}{Name of the local environment of the main function [required].}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2, progress log ; 3, progress and results summary; 5, full report [default 2 or as specified using gl.set.verbosity]}
}
\value{
datatype, "SNP" for SNP data or  "SilicoDArT" for P/A data
}
\description{
Most functions require access to a genlight object, and this function checks that a genlight object has been passed,
whether it is a SNP dataset or a SilicoDArT object, and reports back if verbosity is >=2
}
\details{
This function takes as parameters the genlight object (x) and the name of the local environment of the main function. This function then does the above mentioned checks and writes the checks outputs to the local environment of the main function. In the future, other checks could be added to this function, such as checking for potential source of errors or checking flags.
}
\examples{
utils.check.gl(testset.gl,env = environment())

}
