% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.clumpp.r
\name{utils.clumpp}
\alias{utils.clumpp}
\title{Functions from package starmie for merging Q matrices from Structure runs
using the CLUMPP algorithms.}
\usage{
utils.clumpp(Q_list, method, iter)
}
\arguments{
\item{Q_list}{A list of of Q matrices.}

\item{method}{The algorithm to use to infer the correct permutations. One of
'greedy' or 'greedyLargeK' or 'stephens'}

\item{iter}{The number of iterations to use if running either 'greedy' or
'greedyLargeK'}
}
\description{
Functions from package starmie for merging Q matrices from Structure runs
using the CLUMPP algorithms.
}
