% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.dart2genlight.r
\name{utils.dart2genlight}
\alias{utils.dart2genlight}
\title{Convert DarT to Genlight}
\usage{
utils.dart2genlight(dart, ind.metafile = NULL)
}
\arguments{
\item{dart}{a list created by \code{read.dart}}

\item{ind.metafile}{name of csv file with metadata for each individual.}
}
\value{
A genlight object with all slots filled.
}
\description{
Converts dart data to an genlight object \code{\link{adegenet}}.
}
\details{
This is an internal function called by \code{gl.read.dart} to
convert the data read by \code{read.dart} and return a genlight
object.

The file of individual covariates is searched for the columns
\describe{
\item{\code{id}}{individual identifiers matching the genetic data in the dart object}
\item{\code{pop}}{population identifier for each individual}
\item{\code{lat}, \code{lon}}{spatial coordinates (perferably in decimal degrees WGS1984 format)}
}
Of these, only \code{id} is mandatory.  Covariates other than
these contained in the file will be appended to individual
metadata dataframe in the genlight object.
}
\author{
Bernd Gruber (bugs? Post to \url{https://groups.google.com/d/forum/dartr})
}
