% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.flag.start.r
\name{utils.flag.start}
\alias{utils.flag.start}
\title{A utility script to flag the start of a script}
\usage{
utils.flag.start(func = NULL, build = NULL, verbose = NULL)
}
\arguments{
\item{func}{Name of the function that is starting [required].}

\item{build}{Name of the build [default NULL].}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2,
progress log; 3, progress and results summary; 5, full report [default 2].}
}
\value{
calling function name
}
\description{
A utility script to flag the start of a script
}
\author{
Custodian: Arthur Georges -- Post to
 \url{https://groups.google.com/d/forum/dartr}
 @export
}
