% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.hamming.r
\name{utils.hamming}
\alias{utils.hamming}
\title{Calculates the Hamming distance between two DArT trimmed DNA sequences}
\usage{
utils.hamming(str1, str2, r = 4)
}
\arguments{
\item{str1}{string containing the first sequence}

\item{str2}{string containing the second sequence}

\item{r}{number of bases in the restriction enzyme recognition sequence}
}
\value{
Hamming distance between the two strings
}
\description{
Hamming distance is calculated as the number of base differences between two
sequences which can be expressed as a count or a proportion. Typically, it is
calculated between two sequences of equal length. In the context of DArT
trimmed sequences, which differ in length but which are anchored to the left
by the restriction enzyme recognition sequence, it is sensible to compare the
two trimmed sequences starting from immediately after the common recognition
sequence and terminating at the last base of the shorter sequence.
}
\details{
So if the two DNA sequences are of differing length, the longer is truncated. The
initial common restriction enzyme recognition sequence is ignored.
}
\author{
Arthur Georges (bugs? Post to \url{https://groups.google.com/d/forum/dartr})
}
