% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.plot.save.R
\name{utils.plot.save}
\alias{utils.plot.save}
\title{An internal function to save a ggplot object to disk in RDS binary format}
\usage{
utils.plot.save(x, dir = NULL, file = NULL, verbose = NULL, ...)
}
\arguments{
\item{x}{Name of the ggplot object.}

\item{dir}{Name of the directory to save the file.}

\item{file}{Name of the file to save the plot to (omit file extension)}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2,
progress log; 3, progress and results summary; 5, full report
 [default NULL, unless specified using gl.set.verbosity]}

\item{...}{Parameters passed to function \link[ggplot2]{ggsave}, 
such as width and height, when the ggplot is to be saved.}
}
\description{
WARNING: UTILITY SCRIPTS ARE FOR INTERNAL USE ONLY AND SHOULD NOT BE USED BY END USERS AS THEIR USE OUT OF CONTEXT COULD LEAD TO UNPREDICTABLE OUTCOMES.
}
\details{
An internal function to save a ggplot object to disk in RDS binary format.
Uses saveRDS() to save the file with an .RDS extension; can be reloaded with gl.load().
}
\author{
Custodian: Arthur Georges (Post to \url{https://groups.google.com/d/forum/dartr})
}
\concept{utilities}
