% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.read.dart.r
\name{utils.read.dart}
\alias{utils.read.dart}
\title{Import DarT data as a list.}
\usage{
utils.read.dart(filename, na.strings = "-", skip = NULL,
  last.metric = "RepAvg", nrows = NULL)
}
\arguments{
\item{filename}{DarT data in csv format.}

\item{na.strings}{strings representing NAs.}

\item{skip}{number of lines before the row of column names. If \code{NULL} it is diagnosed from the file.}

\item{last.metric}{the name of the last column of preceeding the genetic data}

\item{nrows}{the DarT format.  If \code{NULL} it is diagnosed from the file.}
}
\value{
a list with elements
\item{nrows}{number of rows per loci in DarT genetic data}
\item{nind}{number of individuals}
\item{nsnp}{number of snps}
\item{covmetrics}{non-genetic metrics}
\item{gendata}{genetic data}
}
\description{
Read DarT data in csv format.
}
\details{
This is an internal function called by \code{gl.read.dart}.
}
\seealso{
\code{\link{gl.read.dart}}
}
\author{
Bernd Gruber (bugs? Post to \url{https://groups.google.com/d/forum/dartr})
}
