% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.recalc.avgpic.r
\name{utils.recalc.avgpic}
\alias{utils.recalc.avgpic}
\title{Calculate AvgPIC}
\usage{
utils.recalc.avgpic(gl)
}
\arguments{
\item{gl}{a genlight object}
}
\value{
The modified genlight object
}
\description{
Calculate the OneRatioRef, OneRatioSnp, PICRef, PICSnp, and AvgPIC by locus.
}
\details{
The locus metadata supplied by DArT has OneRatioRef, OneRatioSnp,
PICRef, PICSnp, and AvgPIC included, but the allelec composition
will change when some individuals are removed from the dataset and
so the initial statistics will no longer apply. This script
recalculates these statistics and places the recalculated values
in the appropriate place in the genlight object.
}
\author{
Arthur Georges (Post to \url{https://groups.google.com/d/forum/dartr})
}
