% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.recalc.callrate.r
\name{utils.recalc.callrate}
\alias{utils.recalc.callrate}
\title{Calculate Callrate}
\usage{
utils.recalc.callrate(gl)
}
\arguments{
\item{gl}{a genlight object}
}
\value{
the modified genlight object
}
\description{
Calculate the callrate by locus
}
\details{
SNP datasets generated by DArT have missing values primarily
arising from failure to call a SNP because of a mutation at one or
both of the the restriction enzyme recognition sites. The locus
metadata supplied by DArT has callrate included, but the call rate
will change when some individuals are removed from the
dataset. This function recalculates the callrate and places these
recalculated values in the appropriate place in the genlight
object.
}
\author{
Arthur Georges (Post to \url{https://groups.google.com/d/forum/dartr})
}
