% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.recalc.freq.r
\name{utils.recalc.freq}
\alias{utils.recalc.freq}
\title{Calculate Allele Frequencies}
\usage{
utils.recalc.freq(gl)
}
\arguments{
\item{gl}{a genlight object}
}
\value{
the modified genlight object
}
\description{
Calculate the the various allele frequency by locus.
}
\details{
Recalculates frequencies of the heterozygous SNP and the
homozygous reference and alternate SNP. The locus metadata
supplied by DArT does not have MAF included, so it is calculated
and added to the metadata.
}
\examples{
#f <- utils.recalc.freq(testset.gl)
}
\author{
Arthur Georges (Post to \url{https://groups.google.com/d/forum/dartr})
}
