% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.recalc.freqhets.r
\name{utils.recalc.freqhets}
\alias{utils.recalc.freqhets}
\title{A utility script to recalculate the the frequency of the heterozygous SNPs by locus after some populations have been deleted}
\usage{
utils.recalc.freqhets(gl, v = 1)
}
\arguments{
\item{gl}{-- name of the genlight object containing the SNP data [required]}

\item{v}{-- v=0, silent; v=1, low verbosity; v=2, high verbosity [default 1]}
}
\value{
The modified genlight object
}
\description{
The locus metadata supplied by DArT has FreqHets included,
but the frequency of the heterozygotes will change when some individuals are removed from the dataset. 
This script recalculates the FreqHets and places these recalculated values in the appropriate place in the genlight object.
Note that the frequency of the homozygote reference SNPS is calculated from the individuals that could be scored.
}
\examples{
result <- dartR:::utils.recalc.freqhets(testset.gl)
}
\author{
Arthur Georges (glbugs@aerg.canberra.edu.au)
}
