% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.recalc.rdepth.r
\name{utils.recalc.rdepth}
\alias{utils.recalc.rdepth}
\title{Calculate the read depth by locus}
\usage{
utils.recalc.rdepth(gl)
}
\arguments{
\item{gl}{a genlight object}
}
\value{
The modified genlight object
}
\description{
The locus metadata supplied by DArT does not have read depth
included, so it is calculated and added to the locus metrics by
this script. It calculates the read depth as the sum of the average
counts for the reference and alterate alleles.
}
\examples{
#f <- utils.recalc.rdepth(testset.gl)
}
\author{
Arthur Georges (Post to \url{https://groups.google.com/d/forum/dartr})
}
