% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.spautocor.R
\name{utils.spautocorr}
\alias{utils.spautocorr}
\alias{utils.spautocor}
\title{Spatial autocorrelation coefficient calculations}
\usage{
utils.spautocor(GD, GGD, permutation = FALSE, bootstrap = FALSE, bins = 10)
}
\arguments{
\item{GD}{a matrix of individual pairwise genetic distances. In principle 
any other squared distance matrix can be used. see example}

\item{GGD}{A geographical distance matrix, based on the coordinates of
individuals. This is typically an Euclidean distance but it can be 
any meaningful (geographical) distance metrics.}

\item{permutation}{Whether permutation calculations for the null hypothesis of 
no spatial structure should be carried out}

\item{bootstrap}{Whether bootstrap calculations to compute the 95% confidence intervals 
around r should be carried out}
}
\value{
Returns a data frame with the following columns:
\enumerate{
\item Bin  The distance classes
\item N The number of pairwise comparisons within each distance class
\item r.uc The uncorrected autocorrelation coefficient
}
if both \code{bootstap} and \code{permutation} are \code{FALSE} otherwise only
\code{r} estimates are returned
}
\description{
Carries out calculation for spatial autocorrelation coefficient 
starting from a genetic and geogaphic distance matrix.
}
\details{
The code of this function is based one \code{spautocorr} from the package 
\code{PopGenReport}, which has been modified to fix a few bugs (as of 
\code{PopGenReport v 3.0.4} and allow calculations of bootstraps estimates. 

See details from \code{gl.spatial.autoCorr} for a detailed explanation.
}
\references{

}
\author{
Carlo Pacioni & Bernd Gruber
}
