% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.structure.evanno.r
\name{utils.structure.evanno}
\alias{utils.structure.evanno}
\title{Util function for evanno plots}
\usage{
utils.structure.evanno(sr, plot = TRUE)
}
\arguments{
\item{sr}{structure run object}

\item{plot}{should the plots be returned}
}
\description{
These functions were copied from package strataG, which is no longer on CRAN 
(maintained by Eric Archer)
}
\author{
Bernd Gruber (bugs? Post to
 \url{https://groups.google.com/d/forum/dartr}); original implementation of
  Eric Archer \url{https://github.com/EricArcher/strataG}
}
