% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.structure.genind2gtypes.r
\name{utils.structure.genind2gtypes}
\alias{utils.structure.genind2gtypes}
\title{structure util functions}
\usage{
utils.structure.genind2gtypes(x)
}
\arguments{
\item{x}{a genind object}
}
\value{
a gtypes object
}
\description{
These functions were copied from package strataG, which is no longer on CRAN (maintained by Eric Archer)
}
\author{
Bernd Gruber (bugs? Post to
 \url{https://groups.google.com/d/forum/dartr}); original implementation of
  Eric Archer \url{https://github.com/EricArcher/strataG}
}
