% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.structure.run.r
\name{utils.structure.structureRun}
\alias{utils.structure.structureRun}
\title{Utility function to run strcture}
\usage{
utils.structure.structureRun(
  g,
  k.range = NULL,
  num.k.rep = 1,
  label = NULL,
  delete.files = TRUE,
  exec = "structure",
  ...
)
}
\arguments{
\item{g}{a \linkS4class{gtypes} object.}

\item{k.range}{vector of values to for \code{maxpop} in multiple runs. If set
to \code{NULL}, a single STRUCTURE run is conducted with \code{maxpops}
groups. If specified, do not also specify \code{maxpops}.}

\item{num.k.rep}{number of replicates for each value in \code{k.range}.}

\item{label}{label to use for input and output files}

\item{delete.files}{logical. Delete all files when STRUCTURE is finished?}

\item{exec}{name of executable for STRUCTURE. Defaults to "structure".}

\item{...}{arguments to be passed to \code{structureWrite}.}

\item{maxpops}{number of groups.}

\item{burnin}{number of iterations for MCMC burnin.}

\item{numreps}{number of MCMC replicates.}

\item{noadmix}{logical. No admixture?}

\item{freqscorr}{logical. Correlated frequencies?}

\item{randomize}{randomize.}

\item{seed}{set random seed.}

\item{pop.prior}{a character specifying which population prior model to use:
"locprior" or "usepopinfo".}

\item{locpriorinit}{parameterizes locprior parameter \emph{r} - how
informative the populations are. Only used when \code{pop.prior} =
"locprior".}

\item{maxlocprior}{specifies range of locprior parameter \emph{r}. Only used
when \code{pop.prior} = "locprior".}

\item{gensback}{integer defining the number of generations back to test for
immigrant ancestry. Only used when \code{pop.prior} = "usepopinfo".}

\item{migrprior}{numeric between 0 and 1 listing migration prior. Only used
when \code{pop.prior} = "usepopinfo".}

\item{pfrompopflagonly}{logical. update allele frequencies from individuals
specified by \code{popflag}. Only used when \code{pop.prior} =
"usepopinfo".}

\item{popflag}{a vector of integers (0, 1) or logicals identifiying whether
or not to use strata information. Only used when \code{pop.prior} =
"usepopinfo".}

\item{inferalpha}{logical. Infer the value of the model parameter # from the
data; otherwise is fixed at the value \code{alpha} which is chosen by the
user. This option is ignored under the NOADMIX model. Small \code{alpha}
implies that most individuals are essentially from one population or
another, while \code{alpha} > 1 implies that most individuals are admixed.)}

\item{alpha}{Dirichlet parameter for degree of admixture. This is the initial
value if \code{inferalpha = TRUE}.}

\item{unifprioralpha}{logical. Assume a uniform prior for \code{alpha} which
runs between 0 and \code{alphamax}. This model seems to work fine; the
alternative model (when \code{unfprioralpha = 0}) is to take \code{alpha}
as having a Gamma prior, with mean \code{alphapriora × alphapriorb}, and
variance \code{alphapriora × alphapriorb^2}.}

\item{alphamax}{maximum for uniform prior on \code{alpha} when
\code{unifprioralpha = TRUE}.}

\item{alphapriora, alphapriorb}{parameters of Gamma prior on \code{alpha} when
\code{unifprioralpha = FALSE}.}

\item{file}{name of the output file from STRUCTURE.}

\item{pops}{vector of population labels to be used in place of numbers in
STRUCTURE file.}
}
\value{
\describe{ \item{\code{structureRun}}{a list where each element is a
list with results from \code{structureRead} and a vector of the filenames
used} \item{\code{structureWrite}}{a vector of the filenames used by
STRUCTURE} \item{\code{structureRead}}{a list containing: \describe{
\item{\code{summary}}{new locus name, which is a combination of loci in
group} \item{\code{q.mat}}{data.frame of assignment probabilities for each
id} \item{\code{prior.anc}}{list of prior ancestry estimates for each
individual where population priors were used} \item{\code{files}}{vector of
input and output files used by STRUCTURE} \item{\code{label}}{label for the
run} } } }
}
\description{
These functions were copied from package strataG, which is no longer on CRAN (maintained by Eric Archer)
}
\author{
Bernd Gruber (bugs? Post to
 \url{https://groups.google.com/d/forum/dartr}); original implementation of
 Eric Archer \url{https://github.com/EricArcher/strataG}
}
