% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dartR.r
\name{cbind.dartR}
\alias{cbind.dartR}
\title{adjust cbind for dartR}
\usage{
\method{cbind}{dartR}(...)
}
\arguments{
\item{...}{list of dartR objects}
}
\description{
cbind is a bit lazy and does not take care for the metadata (so data in the 
other slot is lost). You can get most of the loci metadata back using 
gl.compliance.check.
}
