% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gi2gl.R
\name{gi2gl}
\alias{gi2gl}
\title{Converts a genind object to genlight object}
\usage{
gi2gl(gi)
}
\arguments{
\item{gi}{-- a genind object}
}
\value{
A genlight object, with all slots filled.
}
\description{
Converts a genind object to genlight object
}
\details{
Be aware due to ambiguity which one is the reference allele a combination of gi2gl(gl2gi(gl)) does not return an identical object (but in terms of analysis this conversions are equivalent)
}
\author{
Bernd Gruber (glbugs@aerg.canberra.edu.au)
}

