% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.alf.r
\name{gl.alf}
\alias{gl.alf}
\title{Calculates allele frequency of the first and second allele for each loci #' 
A very simple function to report allele frequencies}
\usage{
gl.alf(gl)
}
\arguments{
\item{gl}{-- a genlight object}
}
\value{
a simple data.frame with alf1, alf2
}
\description{
Calculates allele frequency of the first and second allele for each loci #' 
A very simple function to report allele frequencies
}
\examples{
#for the first 10 loci only
gl.alf(possums.gl[,1:10])
barplot(t(as.matrix(gl.alf(possums.gl[,1:10]))))
}
\author{
Bernd Gruber (bugs? Post to \url{https://groups.google.com/d/forum/dartr})
}
