% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.assign.pca.r, R/gl.assign.pca2.r
\name{gl.assign.pca}
\alias{gl.assign.pca}
\alias{gl.assign.pca2}
\title{Assign an individual of unknown provenance to population based on PCA}
\usage{
gl.assign.pca(x, unknown, plevel = 0.95, verbose = NULL)

gl.assign.pca2(x, unknown, plevel = 0.95, verbose = NULL)
}
\arguments{
\item{x}{-- name of the input genlight object [required]}

\item{unknown}{-- identity label of the focal individual whose provenance is unknown [required]}

\item{plevel}{-- probability level for bounding ellipses in the PCoA plot [default 0.95]}

\item{verbose}{-- verbosity: 0, silent or fatal errors; 1, begin and end; 2, progress log ; 3, progress and results summary; 5, full report [default 2 or as specified using gl.set.verbosity]}
}
\value{
A genlight object containing the focal individual (assigned to population "unknown") and #' populations for which the focal individual is not distinctive (number of loci with private alleles less than or equal to thresold t.

A genlight object containing the focal individual (assigned to population "unknown") and #' populations for which the focal individual is not distinctive (number of loci with private alleles less than or equal to thresold t.
}
\description{
This script assigns an individual of unknown provenance to one or more target populations based on its proximity to
each population defined by a confidence ellipse.

The following process is followed.
(a) The space defined by the loci is ordinated to yield a series of orthogonal axes (independent), and the top two
dimensions are considered. Populations for which the unknown lies outside the specified confidence limits are no longer
considered.
(b) The ordination is repeated on the remaining populations to again yield a series of orthogonal (independent) axes, 
a necessary condition for combining likelihoods calculated from each axis.
(b) A workable subset of dimensions is chosen, normally equal to the number of dimensions
with substantive eigenvalues. 
(c) The log-likelihood of the value for the unknown on each axis is calculated, weighted by the eigenvalue for that axis, 
and summed over all dimensions as an assignment weighting. The assignment weighting is calculated for a point on the boundary of
the 95% (or as specified) confidence envelope, and an index is calculated as the ratio of the two. A value less than 1
indicates that the unknown is within the confidence envelope, a value greater than 1 indicates that the unknown lies
outside the confidence envelope. A smaller value of the index indicates greater conformity of the unknown genotype with
the target population.

This script assigns an individual of unknown provenance to one or more target populations based on its proximity to
each population defined by a confidence ellipse.

The following process is followed.
(a) The space defined by the loci is ordinated to yield a series of orthogonal axes (independent), and the top two
dimensions are considered. Populations for which the unknown lies outside the specified confidence limits are no longer
considered.
(b) The ordination is repeated on the remaining populations to again yield a series of orthogonal (independent) axes, 
a necessary condition for combining likelihoods calculated from each axis.
(b) A workable subset of dimensions is chosen, normally equal to the number of dimensions
with substantive eigenvalues. 
(c) The log-likelihood of the value for the unknown on each axis is calculated, weighted by the eigenvalue for that axis, 
and summed over all dimensions as an assignment weighting. The assignment weighting is calculated for a point on the boundary of
the 95% (or as specified) confidence envelope, and an index is calculated as the ratio of the two. A value less than 1
indicates that the unknown is within the confidence envelope, a value greater than 1 indicates that the unknown lies
outside the confidence envelope. A smaller value of the index indicates greater conformity of the unknown genotype with
the target population.
}
\details{
There are three considerations to the assignment. First, consider only those populations for which the unknown has no
private alleles. Private alleles are an indication that the unknown does not belong to a target population (provided that
the sample size is adequate, say >=10). This can be evaluated with gl.assign.pa().

A next step is to consider the PCoA plot for populations where no private alleles have been
detected and the position of the unknown in relation to the confidence ellipses as is plotted by this script. Note, this 
plot is considering only the top two dimensions of the ordination, and so an unknown lying outside the confidence 
ellipse can be interpreted as it lying outside the confidence envelope. However, if the unknown lies inside the 
confidence ellipse in two dimensions, then it may still lie outside the confidence envelope. This second step is good for 
eliminating populations from consideration, but does not provide confidence in assignment. 

The third step is to consider the assignment probabilities. This approach calculates the squared Generalised Linear 
Distance (Mahalanobis distance) of the unknown from the centroid for each population, and calculates the probability 
associated with its quantile under the zero truncated normal distribution. This index takes into account position of 
the unknown in relation to the confidence envelope in all selected dimensions of the ordination. An index is calculated
based on the distance of the unknown from the population centroid, scaled for the distance of the standardized
confidence ellipse. An index of less than 1 indicates that the unknown lies within the confidence ellipse for the focal
population; an index of greater than 1 indicates that the unknown lies outside the confidence ellipse for the focal
population. The smaller the index, the greater the confidence in the assignment.

  Wt is a weighted log-likelihood
  CE is the value of the Index on the boundary of the 95 % confidence envelope
  Index is the position of the unknown relative to the boundary of the 95 % confidence envelope
     An index value less than 1 indicates the unknown resides inside the confidence ellipse for the focal population
     An index value greater than 1 indicates the unknown resides outside the confidence ellipse for the focal population

Each of these approaches provides evidence, none are 100% definitive. They need to be interpreted cautiously.

There are three considerations to the assignment. First, consider only those populations for which the unknown has no
private alleles. Private alleles are an indication that the unknown does not belong to a target population (provided that
the sample size is adequate, say >=10). This can be evaluated with gl.assign.pa().

A next step is to consider the PCoA plot for populations where no private alleles have been
detected and the position of the unknown in relation to the confidence ellipses as is plotted by this script. Note, this 
plot is considering only the top two dimensions of the ordination, and so an unknown lying outside the confidence 
ellipse can be interpreted as it lying outside the confidence envelope. However, if the unknown lies inside the 
confidence ellipse in two dimensions, then it may still lie outside the confidence envelope. This second step is good for 
eliminating populations from consideration, but does not provide confidence in assignment. 

The third step is to consider the assignment probabilities. This approach calculates the squared Generalised Linear 
Distance (Mahalanobis distance) of the unknown from the centroid for each population, and calculates the probability 
associated with its quantile under the zero truncated normal distribution. This index takes into account position of 
the unknown in relation to the confidence envelope in all selected dimensions of the ordination. An index is calculated
based on the distance of the unknown from the population centroid, scaled for the distance of the standardized
confidence ellipse. An index of less than 1 indicates that the unknown lies within the confidence ellipse for the focal
population; an index of greater than 1 indicates that the unknown lies outside the confidence ellipse for the focal
population. The smaller the index, the greater the confidence in the assignment.

  Wt is a weighted log-likelihood
  CE is the value of the Index on the boundary of the 95 % confidence envelope
  Index is the position of the unknown relative to the boundary of the 95 % confidence envelope
     An index value less than 1 indicates the unknown resides inside the confidence ellipse for the focal population
     An index value greater than 1 indicates the unknown resides outside the confidence ellipse for the focal population

Each of these approaches provides evidence, none are 100% definitive. They need to be interpreted cautiously.
}
\examples{
# Test run with a focal individual from the Macleay River (EmmacMaclGeor)
  x <- gl.assign.pa(testset.gl, unknown="UC_00146", nmin=10, threshold=1,verbose=3)
  x <- gl.assign.pca(x, unknown="UC_00146", plevel=0.95, verbose=3)
# Test run with a focal individual from the Macleay River (EmmacMaclGeor)
  x <- gl.assign.pa(testset.gl, unknown="UC_00146", nmin=10, threshold=1,verbose=3)
  x <- gl.assign.pca(x, unknown="UC_00146", alpha=0.05,verbose=3)
}
\author{
Custodian: Arthur Georges -- Post to \url{https://groups.google.com/d/forum/dartr}

Custodian: Arthur Georges -- Post to \url{https://groups.google.com/d/forum/dartr}
}
