% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.collapse.r
\name{gl.collapse}
\alias{gl.collapse}
\title{Collapse a distance matrix by amalgamating populations with pairwise fixed difference count less that an threshold}
\usage{
gl.collapse(
  fd,
  recode.table = "tmp.csv",
  outpath = tempdir(),
  tpop = 0,
  tloc = 0,
  test = FALSE,
  delta = 0.02,
  reps = 1000,
  plot = FALSE,
  pb = FALSE,
  verbose = NULL
)
}
\arguments{
\item{fd}{-- name of the list of matricies produced by gl.fixed.diff() [required]}

\item{recode.table}{-- name of the new recode.table to receive the new population reassignments 
arising from the amalgamation of populations [tmp.csv]}

\item{outpath}{-- path where to save the output file [default tempdir(), mandated by CRAN]. Use outpath=getwd() or outpath="." when calling this function to direct output files to your working directory.}

\item{tpop}{-- threshold number of fixed differences above which populatons will not be amalgamated [0]}

\item{tloc}{-- threshold defining a fixed difference (e.g. 0.05 implies 95:5 vs 5:95 is fixed) [0]}

\item{test}{-- should a test be performed, default is [FALSE]}

\item{delta}{-- needs to be described}

\item{reps}{-- number of repeats - needs to be described}

\item{plot}{-- if TRUE, plot a PCoA with the new groupings [FALSE]}

\item{pb}{-- if TRUE, show a progress bar on time consuming loops [FALSE]}

\item{verbose}{-- verbosity: 0, silent or fatal errors; 1, begin and end; 2, progress log ; 3, progress and results summary; 5, full report [default 2 or as specified using gl.set.verbosity]}
}
\value{
A list containing the gl object x and the following square matricies
        [[1]] $gl -- the new genlight object with populations collapsed;
        [[2]] $fd -- raw fixed differences;
        [[3]] $pcfd -- percent fixed differences;
        [[4]] $nobs -- mean no. of individuals used in each comparison;
        [[5]] $nloc -- total number of loci used in each comparison;
        [[6]] $expobs -- if test=TRUE, the expected count of false positives for each comparison [by simulation];
        [[7]] $prob -- if test=TRUE, the significance of the count of fixed differences [by simulation])
}
\description{
This script takes a the file generated by gl.fixed.diff and generates a population recode
table to amalgamate populations with distance less than or equal to a specified threshold
The distance matrix is generated by gl.fixed.diff()
}
\details{
The script then applies the recode to the genlight object and recalculates the distance and associated matricies.
}
\examples{
\donttest{
fd <- gl.fixed.diff(testset.gl, tloc=0.05)
gl <- gl.collapse(fd, recode.table="testset_recode.csv",tpop=1)
}
}
\author{
Arthur Georges (Post to \url{https://groups.google.com/d/forum/dartr})
}
