% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.dist.ind.r
\name{gl.dist.ind}
\alias{gl.dist.ind}
\title{Calculates a distance matrix for individuals defined in a genlight object}
\usage{
gl.dist.ind(
  x,
  method = NULL,
  scale = FALSE,
  swap = FALSE,
  output = "dist",
  plot.out = TRUE,
  plot_theme = theme_dartR(),
  plot_colors = two_colors,
  save2tmp = FALSE,
  verbose = NULL
)
}
\arguments{
\item{x}{Name of the genlight containing the SNP genotypes or presence-absence data [required].}

\item{method}{Specify distance measure [SNP: Euclidean; P/A: SMD].}

\item{scale}{If TRUE, the distances are scaled to fall in the range [0,1] [default TRUE]}

\item{swap}{If TRUE and working with presence-absence data, then presence 
(no disrupting mutation) is scored as 0 and absence (presence of a disrupting 
mutation) is scored as 1 [default FALSE].}

\item{output}{Specify the format and class of the object to be returned, 
'dist' for a object of class dist, 'matrix' for an object of class matrix [default "dist"].}

\item{plot.out}{If TRUE, display a histogram and a boxplot of the genetic distances [TRUE].}

\item{plot_theme}{User specified theme [default theme_dartR].}

\item{plot_colors}{Vector with two color names for the borders and fill [default two_colors].}

\item{save2tmp}{If TRUE, saves any ggplots to the session temporary directory [default FALSE].}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2,
progress log ; 3, progress and results summary; 5, full report
[default 2 or as specified using gl.set.verbosity].}
}
\value{
An object of class 'matrix' or dist' giving distances between individuals
}
\description{
This script calculates various distances between individuals based on allele
 frequencies or presence-absence data
}
\details{
The distance measure for SNP genotypes can be one of:
\itemize{
 \item Euclidean Distance
 \item Scaled Euclidean Distance
 \item Simple Mismatch Distance
 \item Absolute Mismatch Distance
 \item Czekanowski (Manhattan) Distance
 }

The distance measure for Sequence Tag Presence/Absence data (binary) can be one of:
\itemize{
 \item Euclidean Distance
 \item Scaled Euclidean Distance
 \item Simple Matching Distance
 \item Jaccard Distance
 \item Bray-Curtis Distance
 }

Refer to the dartR Technical Note on Distances in Genetics.
}
\examples{
D <- gl.dist.ind(testset.gl, method='euclidean',scale=TRUE)
D <- gl.dist.ind(testset.gl, method='manhattan')
D <- gl.dist.ind(testset.gs, method='Jaccard',swap=TRUE)
}
\author{
Author(s): Arthur Georges. Custodian: Arthur Georges -- Post to #' \url{https://groups.google.com/d/forum/dartr}
}
