% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.drop.loc.r
\name{gl.drop.loc}
\alias{gl.drop.loc}
\title{Remove specified loci from a genelight \{adegenet\} object}
\usage{
gl.drop.loc(x, loc.list, v = 2)
}
\arguments{
\item{x}{-- name of the genlight object containing SNP genotypes or a genind object containing presence/absence data [required]}

\item{v}{-- verbosity: 0, silent or fatal errors; 1, begin and end; 2, progress log ; 3, progress and results summary; 5, full report [default 2]}
}
\value{
A genlight object with the reduced data
}
\description{
This script deletes selected loci from the nominated dataset.
}
\examples{
   gl <- gl.drop.loc(testset.gl, loc.list=c("100049687|12-A/G","100050106|50-G/A"))

}
\author{
Arthur Georges (glbugs@aerg.canberra.edu.au)
}
