% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.drop.pop.r
\name{gl.drop.pop}
\alias{gl.drop.pop}
\title{Remove specified populations from a genelight \{adegenet\} object}
\usage{
gl.drop.pop(gl, pop.list, recalc = TRUE, mono.rm = TRUE, v = 1)
}
\arguments{
\item{gl}{-- name of the genlight object containing SNP genotypes or a genind object containing presence/absence data [required]}

\item{pop.list}{-- a list of populations to be removed [required]}

\item{recalc}{-- Recalculate the locus metadata statistics [default TRUE]}

\item{mono.rm}{-- Remove monomorphic loci [default TRUE]}

\item{v}{-- verbosity: 0, silent; 1, brief; 2, verbose [default 1]}
}
\value{
A genlight object with the reduced data
}
\description{
Individuals are assigned to populations based on the specimen metadata data file (csv) used with gl.read.dart().
}
\details{
The script, having deleted populations, optionally identifies resultant monomorphic loci or loci
with all values missing and deletes them (using gl.filter.monomorphs.r). The script also optionally
recalculates statistics made redundant by the deletion of individuals from the dataset.

The script returns a genlight object with the new population assignments and the recalculated locus metadata.
}
\examples{
\dontrun{
   gl <- gl.drop.pop(testset.gl, pop.list=c("EmsubRopeMata","EmvicVictJasp"))
}
}
\seealso{
\code{\link{gl.filter.monomorphs}}
}
\author{
Arthur Georges (glbugs@aerg.canberra.edu.au)
}
