% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.filter.callrate.loc.pop.r
\name{gl.filter.callrate.loc.pop}
\alias{gl.filter.callrate.loc.pop}
\title{Filter loci in a genlight adegenet object based on call rate and population level.}
\usage{
gl.filter.callrate.loc.pop(x, threshold)
}
\arguments{
\item{x}{- genlight object to be filtered}

\item{threshold}{threshold on call rate for loci for each population (see gl.filter.callrate for details.)}
}
\value{
The filtered genlight or genind object
}
\description{
This funciton is a convenience function that filters callrate for the set threshold for each population and then returns a genlight object where each loci of every population passes the threshold of the call rate filter. For more details on call rates see gl.filter.callrate.
}
\examples{
gg <- gl.filter.callrate.loc.pop(testset.gl, 0.9)
}
