% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.filter.cloneid.r
\name{gl.filter.cloneid}
\alias{gl.filter.cloneid}
\title{Filter for CloneID to select only unique SNPs}
\usage{
gl.filter.cloneid(gl)
}
\arguments{
\item{gl}{a genlight object created via read.dart (needs to have a cloneID as provided by dart)}
}
\value{
filtered genlight object, with unique cloneIDs
}
\description{
Filter for CloneID to select only unique SNPs
}
\examples{
{
}
}
