% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.filter.heterozygosity.r
\name{gl.filter.heterozygosity}
\alias{gl.filter.heterozygosity}
\title{Filters individuals with average heterozygosity greater than a specified threshold}
\usage{
gl.filter.heterozygosity(x, threshold = 0.7, verbose = 2)
}
\arguments{
\item{x}{-- a genlight object containing the SNP genotypes [Required]}

\item{threshold}{-- filter individuals above the threshold [default 0.5]}

\item{verbose}{-- verbosity: 0, silent or fatal errors; 1, begin and end; 2, progress log ; 3, progress and results summary; 5, full report [default 2]}
}
\value{
the filtered genlight object
}
\description{
Calculates the observed and expected heterozygisities for each population (method="pop") 
or the observed heterozyosity for each individual (method="ind") in a genlight object.
}
\examples{
tmp <- gl.filter.heterozygosity(testset.gl,threshold=0.06,verbose=3)
}
\author{
Arthur Georges (Post to \url{https://groups.google.com/d/forum/dartr})
}
