% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.filter.heterozygosity.r
\name{gl.filter.heterozygosity}
\alias{gl.filter.heterozygosity}
\title{Filters individuals with average heterozygosity greater than a specified upper threshold or less than a specified lower threshold.}
\usage{
gl.filter.heterozygosity(x, t.upper = 0.7, t.lower = 0, verbose = 2)
}
\arguments{
\item{x}{-- a genlight object containing the SNP genotypes [Required]}

\item{t.upper}{-- filter individuals > the threshold [default 0.7]}

\item{t.lower}{-- filter individuals < the threshold [default 0]}

\item{verbose}{-- verbosity: 0, silent or fatal errors; 1, begin and end; 2, progress log ; 3, progress and results summary; 5, full report [default 2]}
}
\value{
the filtered genlight object
}
\description{
Calculates the observed heterozyosity for each individual in a genlight object and filters individuals based on specified threshold values.
Use gl.report.heterozygosity to determine the appropriate thresholds.
}
\examples{
tmp <- gl.filter.heterozygosity(testset.gl,t.upper=0.06,verbose=3)
gl.report.heterozygosity(tmp,method="ind")
}
\author{
Arthur Georges (Post to \url{https://groups.google.com/d/forum/dartr})
}
