% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.filter.pa.r
\name{gl.filter.pa}
\alias{gl.filter.pa}
\title{Calculates the number of private alleles possessed by an individual of unknown provenance.}
\usage{
gl.filter.pa(x, id, nmin = 10, threshold = 0, verbose = 2)
}
\arguments{
\item{x}{-- name of the input genlight object [required]}

\item{id}{-- identity label of the focal individual whose provenance is unknown [required]}

\item{nmin}{-- minimum sample size for a target population to be included in the analysis [default 10]}

\item{threshold}{-- retain those populations for which the focal individual has private alleles less or equal in number than the threshold [default 0]}

\item{verbose}{-- verbosity: 0, silent or fatal errors; 1, begin and end; 2, progress log ; 3, progress and results summary; 5, full report [default 2]}
}
\value{
A genlight object containing the focal individual (assigned to population "unknown") and 
populations for which the focal individual is not distinctive (number of loci with private alleles less than or equal to thresold t).
}
\description{
Calculates the number of private alleles possessed by an individual of unknown provenance in comparison with a set of populations, and
returns those populations for which the calculated private alleles are less than a threshold.
}
\details{
A private allele is an allele possessed by the focal individual, but absent from the target population.
}
\examples{
# Test run with a focal individual from the Macleay River (EmmacMaclGeor)
x <- gl.report.pa(testset.gl, id="UC_00146", nmin=10, threshold=1, verbose=2)
}
\author{
Arthur Georges (Post to \url{https://groups.google.com/d/forum/dartr})
}
