% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.grm.r
\name{gl.grm}
\alias{gl.grm}
\title{Calculates the genomic relatedness matrix}
\usage{
gl.grm(gl, plotheatmap = TRUE, return.imputed = FALSE, ...)
}
\arguments{
\item{gl}{-- a genlight object}

\item{plotheatmap}{-- a switch if a heatmap should be shown [Default:TRUE]}

\item{return.imputed}{switch if loci with imputed data should be returned (see ?A.mat in package rrBLUP)}

\item{...}{paramters passed to function A.mat from package rrBLUP}
}
\value{
a genomic relatedness matrix
}
\description{
The G matrix is calculated by centering the allele frequency matrix  of the second allele by substracting 2 times the allefrequency
}
\examples{
gl.grm(foxes.gl[1:5,1:10])  
}
