% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.He.r
\name{gl.He}
\alias{gl.He}
\title{A very simple function to report expected Heterozygosity}
\usage{
gl.He(gl)
}
\arguments{
\item{gl}{-- a genlight object}
}
\value{
a simple vector wiht Ho for each loci
}
\description{
A very simple function to report expected Heterozygosity
}
\author{
Bernd Gruber (Post to \url{https://groups.google.com/d/forum/dartr})
}
