% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.Ho.r
\name{gl.Ho}
\alias{gl.Ho}
\title{A very simple function to report observed Heterozygosity}
\usage{
gl.Ho(gl)
}
\arguments{
\item{gl}{A genlight object [required]}
}
\value{
A simple vector whit Ho for each loci
}
\description{
A very simple function to report observed Heterozygosity
}
\author{
Bernd Gruber (bugs? Post to \url{https://groups.google.com/d/forum/dartr})
}
