% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.hwe.diagnostics.R
\name{gl.hwe.diagnostics}
\alias{gl.hwe.diagnostics}
\title{Provides descriptive stats and plots to diagnose potential problems with H-W proportions}
\usage{
gl.hwe.diagnostics(x, alpha_val = 0.05, brk = seq(0, 1, 1/20), verbose = NULL)
}
\arguments{
\item{x}{Name of the genlight object containing the SNP data [required].}

\item{alpha_val}{Level of significance for testing [default 0.05].}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2,
progress log ; 3, progress and results summary; 5, full report
[default NULL, unless specified using gl.set.verbosity].}
}
\value{
A list with the plots and table.
}
\description{
Different causes may be responsible for lack of H-W proportions. This function helps diagnose potential problems
}
\details{

}
\examples{
gl.hwe.diagnostics(x = platypus.gl)
}
\references{
\itemize{
\item De Meeûs, T., Guégan, J.-F., Teriokhin, A.T., 2009. MultiTest V.1.2, a 
program to binomially combine independent tests and performance comparison 
with other related methods on proportional data. BMC Bioinformatics 10, 443-443.
\item Fisher, R., 1970. Statistical methods for research workers Edinburgh: 
Oliver and Boyd.
\item Waples, R. S. (2015). Testing for Hardy–Weinberg proportions: have we
lost the plot?. Journal of heredity, 106(1), 1-19.
}
}
\seealso{
\code{\link{gl.report.hwe}}
}
\author{
Custodian: Carlo Pacioni -- Post to
\url{https://groups.google.com/d/forum/dartr}
}
\concept{}
