% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.ibd.r
\name{gl.ibd}
\alias{gl.ibd}
\title{Isolation by distance}
\usage{
gl.ibd(
  x = NULL,
  distance = "Fst",
  coordinates = "latlon",
  logdist = TRUE,
  logoffset = 0,
  Dgen = NULL,
  Dgeo = NULL,
  permutations = 999,
  plot = TRUE,
  paircols = NULL,
  plot_theme = theme_dartR(),
  verbose = NULL
)
}
\arguments{
\item{x}{genlight object. If provided a standard analysis on Fst/1-Fst and log(distance) is performed}

\item{distance}{type of distance that is calculated and used for the analysis. Can be either population based "Fst" [\link[StAMPP]{stamppFst}], "D" [\link[StAMPP]{stamppNeisD}] or individual based "propShared", [gl.propShared], "euclidean" [gl.dist.ind, method="Euclidean"].}

\item{coordinates}{Can be either "latlon", "xy" or a two column data.frame with column names "lat","lon", "x", "y")  Coordinates are provided via \code{gl@other$latlon} ['latlon'] or via \code{gl@other$xy} ['xy']. If latlon data will be projected to meters using Mercator system [google maps] or if xy then distance is directly calculated on the coordinates.}

\item{logdist}{TRUE/FALSE switch if log of distance should be used [default is set to TRUE].}

\item{logoffset}{0. If you have individuals/populations with zero distances between each other log transfromation results in zero distances, hence you need to an offset to the distances.}

\item{Dgen}{genetic distance matrix if no genlight object is provided}

\item{Dgeo}{Euclidean distance matrix if no genlight object is provided}

\item{permutations}{number of permutations in the mantel test}

\item{plot}{should an isolation by distance plot be returned. Default is plot=TRUE}

\item{paircols}{should pairwise dots colored by "pop"ulation/"ind"ividual pairs [default: pop]. You can color pairwise individuals by pairwise population colors.}

\item{plot_theme}{Theme for the plot. See details for options [default theme_dartR()].}

\item{verbose}{-- verbosity: 0, silent or fatal errors; 1, begin and end; 2, progress log ; 3, progress and results summary; 5, full report [default 2 or as specified using gl.set.verbosity]}
}
\value{
returns a list of the following components: Dgen (the genetic distance matrix), Dgeo (the Euclidean distance matrix), mantel (the statistics of the mantel test)
}
\description{
This functions performs an isolation by distance analysis based on a mantel test and also produces an isolation by distance plot. If a genlight object with coordinates is provided, then a Euclidean and genetic distance matrix are calculated. Currently pairwise Fst and D between population and 1-propShared and Euclidean distance are implemented between individuals are implemented. Coordinates are expected as lat long and converted to Google Earth Mercator projection. If coordinates are already projected, provide them at the x@other$xy slot. 
You can provide also your own genetic and Euclidean distance matrix. The function is based on the code provided by the adegenet tutorial (\url{http://adegenet.r-forge.r-project.org/files/tutorial-basics.pdf}), using the functions  \link[vegan]{mantel} (package vegan), \link[StAMPP]{stamppFst}, \link[StAMPP]{stamppNeisD} (package StAMPP) and gl.propShared or gl.dist.ind. For transformation you need to have the dismo package installed. As a new feature you can plot pairwise relationship using double coloured points (paircols=TRUE). Pairwise relationship can be visualised via populations or individuals, depending which distance is calculated.
}
\examples{
\donttest{
ibd <- gl.ibd(bandicoot.gl)
ibd <- gl.ibd(bandicoot.gl,plot = FALSE)
}
}
\references{
Rousset (1997) Genetic Differentiation and Estimation of Gene Flow from F-Statistics Under Isolation by Distancenetics 145(4), 1219-1228.
}
\seealso{
\link[vegan]{mantel}, \link[StAMPP]{stamppFst}
}
\author{
Bernd Gruber (bugs? Post to \url{https://groups.google.com/d/forum/dartr})
}
