% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.install.vanilla.dartR.r
\name{gl.install.vanilla.dartR}
\alias{gl.install.vanilla.dartR}
\title{This functions installs all required packages for using all functions
available in dartR}
\usage{
gl.install.vanilla.dartR(flavour = NULL, verbose = TRUE)
}
\arguments{
\item{flavour}{If and which version of R you want to install. If NULL
then only needed packages for the current version will be installed. If
'CRAN' current CRAN version will be installed. 'master' installs the GitHub
master branch and 'dev' installs the experimental development branch from
GitHub [default NULL].}

\item{verbose}{Returns information on packages and dartR versions
[default TRUE].}
}
\value{
Returns a message if the installation was successful/required.
}
\description{
The function compares the installed packages with the the currently available
ones on cran. Be aware this function only works if a version of dartR is
already installed on your system. You can choose if you also want to have a
specific version of dartR installed ('CRAN', 'master' or 'dev' ). 'master'
and 'dev' are installed from Github. Be aware the dev version from github is
 not fully tested and most certainly will contain untested functions.
}
\author{
Bernd Gruber (bugs? Post to
\url{https://groups.google.com/d/forum/dartr})
}
