% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.keep.loc.r
\name{gl.keep.loc}
\alias{gl.keep.loc}
\title{Removes all but the specified loci from a genlight \{adegenet\} object}
\usage{
gl.keep.loc(x, loc.list = NULL, first = NULL, last = NULL, verbose = NULL)
}
\arguments{
\item{x}{Name of the genlight object containing SNP genotypes or
presence/absence data [required].}

\item{loc.list}{A list of loci to be kept
[required, if loc.range not specified].}

\item{first}{First of a range of loci to be kept
[required, if loc.list not specified].}

\item{last}{Last of a range of loci to be kept
[if not specified, last locus in the dataset].}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2,
progress log; 3, progress and results summary; 5, full report
[default 2 or as specified using gl.set.verbosity]}
}
\value{
A genlight object with the reduced data
}
\description{
The script returns a genlight object with the all but the specified loci
deleted.
}
\examples{
# SNP data
  gl2 <- gl.keep.loc(testset.gl, loc.list=c('100051468|42-A/T', '100049816-51-A/G'))
# Tag P/A data
  gs2 <- gl.keep.loc(testset.gs, loc.list=c('20134188','19249144'))

}
\seealso{
\code{\link{gl.drop.loc}} to drop rather than keep specified loci
}
\author{
Custodian: Arthur Georges -- Post to
\url{https://groups.google.com/d/forum/dartr}
}
