% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.map.interactive.r
\name{gl.map.interactive}
\alias{gl.map.interactive}
\title{Creates an interactive map (based on latlong) from a genlight object}
\usage{
gl.map.interactive(x, provider = "Esri.NatGeoWorldMap")
}
\arguments{
\item{x}{-- a genlight object [including coordinates within the latlong slot]}

\item{provider}{-- passed to leaflet}
}
\value{
plots a map
}
\description{
Creates an interactive map (based on latlong) from a genlight object
}
\details{
A wrapper around the \pkg{leaflet} package. For possible background maps check as specified via the provider: \url{http://leaflet-extras.github.io/leaflet-providers/preview/index.html}
}
\examples{
#gl.map(bandicoot.gl)
}
\author{
Bernd Gruber (glbugs@aerg.canberra.edu.au)
}
