% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.pcoa.plot.r
\name{gl.pcoa.plot}
\alias{gl.pcoa.plot}
\title{Plots of the results of an ordination generated using gl.pcoa()}
\usage{
gl.pcoa.plot(
  glPca,
  x,
  scale = FALSE,
  ellipse = FALSE,
  p = 0.95,
  labels = "pop",
  as.pop = NULL,
  overlaps = 3,
  interactive_plot = FALSE,
  three_D_plot = FALSE,
  xaxis = 1,
  yaxis = 2,
  zaxis = 3,
  radius = 3,
  plot.out = TRUE,
  plot_theme = theme_dartR(),
  palette_discrete = discrete_palette,
  verbose = NULL
)
}
\arguments{
\item{glPca}{Name of the PCA or PCoA object containing the factor scores and 
eigenvalues [required].}

\item{x}{Name of the genlight object or fd object containing the SNP genotypes 
or a genlight object containing the Tag P/A (SilicoDArT) genotypes or the 
Distance Matrix used to generate the ordination [required].}

\item{scale}{Flag indicating whether or not to scale the x and y axes in 
proportion to \% variation explained [default FALSE].}

\item{ellipse}{Flag to indicate whether or not to display ellipses to 
encapsulate points for each population [default FALSE].}

\item{p}{Value of the percentile for the ellipse to encapsulate points for 
each population [default 0.95].}

\item{labels}{Flag to specify the labels to be added to the plot 
["none"|"ind"|"pop", default = "pop"].}

\item{as.pop}{Assign another metric to represent populations for the plot 
from the slot $other$ind.metrics [default NULL].}

\item{overlaps}{Factor to exclude overlapping labels, higher values allow to 
overlap more labels [default 3].}

\item{interactive_plot}{Interactive plot of the results of a PCoA ordination 
using the package plotly [default FALSE].}

\item{three_D_plot}{3D interactive plot of the results of a PCoA ordination
using the package plotly [default FALSE].}

\item{xaxis}{Identify the x axis from those available in the ordination
[default 1].}

\item{yaxis}{Identify the y axis from those available in the ordination
[default 2].}

\item{zaxis}{Identify the z axis from those available in the ordination, 
only used for three_D_plot = TRUE [default 3].}

\item{radius}{Size of the points [default 3].}

\item{plot.out}{If TRUE, returns a plot object compatible with ggplot, 
otherwise returns a dataframe [default TRUE].}

\item{plot_theme}{User specified theme [default theme_dartR()].}

\item{palette_discrete}{A discrete palette for the color of populations or a 
list with as many colors as there are populations in the dataset
 [default discrete_palette].}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2, 
progress log ; 3, progress and results summary; 5, full report 
[default NULL, unless specified using gl.set.verbosity].}
}
\value{
A plot of the ordination [plot.out = TRUE] or a dataframe [plot.out = FALSE]
}
\description{
This script takes the output from the ordination generated by \code{\link{gl.pcoa}}
and plots the individuals classified by population.
}
\details{
The factor scores are taken from the output of \code{\link{gl.pcoa}} and the population 
assignments are taken from from the original data file. The specimens are 
shown in a bivariate plot optionally with adjacent labels and enclosing 
ellipses.

Any pair of axes can be specified from the ordination, provided they are 
within the range of the nfactors value provided to \code{\link{gl.pcoa}}, 
which can be defined with the parameters xaxis, yaxis and zaxis. 

Axes can be scaled to represent the proportion of variation explained (scale=TRUE).
In any case, the proportion of variation explained by each axis is provided 
in the axis label.

Points displayed in the ordination can be identified if interactive_plot = TRUE,
in which case the resultant plot is a plotly plot. Identification of points 
is by moving the mouse over them. Refer to the plotly package for further information. 

If plot.out=TRUE, returns an object of class ggplot so that layers can 
subsequently be added; if plot.out=FALSE, returns a dataframe with the 
individual labels, population labels and PCOA scores for subsequent plotting 
by the user with ggplot or other plotting software.

Plot themes can be obtained from \itemize{
 \item \url{https://ggplot2.tidyverse.org/reference/ggtheme.html} and \item
 \url{https://yutannihilation.github.io/allYourFigureAreBelongToUs/ggthemes/}
 }

Resultant ggplots and the tabulation are saved to the session's temporary 
directory.
}
\examples{
gl_test <- bandicoot.gl
# five populations in gl_test
nPop(gl_test)
# color list for population colors 
pop_colours <- c("blue","green","black","orange","deeppink")
# Ordination applied to genotypes
pca <- gl.pcoa(gl_test)
# interactive plot to examine labels
pca_plot_interactive <- gl.pcoa.plot(pca, gl_test, ellipse = TRUE, labels = "ind", interactive_plot = T,palette_discrete = pop_colours)
# 3D interactive plot
pca_plot_3D_interactive <- gl.pcoa.plot(pca, gl_test, three_D_plot = TRUE, palette_discrete=pop_colours)
# using sex as population 
names(gl$other$ind.metrics)
pca_plot_sex <- gl.pcoa.plot(glPca=pca, gl_test, as.pop ="assigned.sex", palette_discrete=pop_colours)

}
\seealso{
\code{\link{gl.pcoa}}
}
\author{
Arthur Georges -- Post to \url{https://groups.google.com/d/forum/dartr}
}
\concept{Exploration and visualisation functions}
