% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.pcoa.plot.r
\name{gl.pcoa.plot}
\alias{gl.pcoa.plot}
\title{Bivariate or trivariate plot of the results of an ordination generated using gl.pcoa()}
\usage{
gl.pcoa.plot(
  glPca,
  x,
  scale = FALSE,
  ellipse = FALSE,
  plevel = 0.95,
  labels = "pop",
  max.overlaps = 20,
  interactive = FALSE,
  as.pop = NULL,
  hadjust = 1.5,
  vadjust = 1,
  xaxis = 1,
  yaxis = 2,
  zaxis = NULL,
  pt.size = 2,
  label.size = 20,
  save2tmp = FALSE,
  verbose = NULL
)
}
\arguments{
\item{glPca}{Name of the PCA or PCoA object containing the factor scores and eigenvalues [required]}

\item{x}{Name of the genlight object or fd object containing the SNP genotypes or Tag P/A (SilicoDArT) genotypes or 
the Distance Matrix used to generate the ordination [required]}

\item{scale}{If TRUE, scale the x and y axes in proportion to \% variation explained [default FALSE]}

\item{ellipse}{If TRUE, display ellipses to encapsulate points for each population [default FALSE]}

\item{plevel}{Value of the percentile for the ellipse to encapsulate points for each population [default 0.95]}

\item{labels}{If TRUE, labels will be added to the plot. ["none"|"ind"|"pop"|"legend", default = "pop"]}

\item{max.overlaps}{Maximum number of labels to attempt to rearrange to minimize overlaps. Use Inf to include all labels regardless. [default=20]}

\item{interactive}{If TRUE then the populations are plotted without labels, mouse-over to identify points [default FALSE]}

\item{as.pop}{-- assign another metric to represent populations for the plot [default NULL]}

\item{hadjust}{Horizontal adjustment of label position in 2D plots [default 1.5]}

\item{vadjust}{Vertical adjustment of label position in 2D plots [default 1]}

\item{xaxis}{Identify the x axis from those available in the ordination (xaxis <= nfactors) [default 1]}

\item{yaxis}{Identify the y axis from those available in the ordination (yaxis <= nfactors) [default 2]}

\item{zaxis}{Identify the z axis from those available in the ordination for a 3D plot (zaxis <= nfactors) [default NULL]}

\item{pt.size}{Specify the size of the dispayed points [default 2]}

\item{label.size}{Specify the size of the dispayed axis labels [default 20]}

\item{save2tmp}{If TRUE, saves any ggplots and listings to the session temporary directory (tempdir) [default FALSE]}

\item{verbose}{-- verbosity: 0, silent or fatal errors; 1, begin and end; 2, progress log ; 3, progress and results summary; 5, full report [default 2 or as specified using gl.set.verbosity]}
}
\description{
This script takes output from the ordination generated by gl.pcoa() and plots the individuals classified by population.
}
\details{
The factor scores are taken from the output of gl.pcoa() and the population assignments are taken from
from the original data file. In the bivariate plots, the specimens are shown optionally with adjacent labels
and enclosing ellipses. Population labels on the plot are shuffled so as not to overlap (using package \{ggrepel\}).
This can be a bit clunky, as the labels may be some distance from the points to which they refer, but it provides the
opportunity for moving labels around using graphics software (e.g. Adobe Illustrator). A maximum number of labels to
shuffle is given by parameter max.overlaps.

3D plotting is activated by specifying a zaxis.

Any pair or trio of axes can be specified from the ordination, provided they are within the range of the nfactors value provided to gl.pcoa(). 
In the 2D plots, axes can be scaled to represent the proportion of variation explained. In any case, the proportion of variation explained by 
each axis is provided in the axis label.

Points displayed in the ordination can be identified if the option interactive=TRUE is chosen, in which case the resultant plot is
ggplotly() friendly. Identification of points is by moving the mouse over them. Refer to the plotly package for further information. 
The interactive option is automatically enabled for 3D plotting.
}
\examples{
gl <- testset.gl
levels(pop(gl))<-c(rep("Coast",5),rep("Cooper",3),rep("Coast",5),
rep("MDB",8),rep("Coast",7),"Em.subglobosa","Em.victoriae")

pca<-gl.pcoa(gl,nfactors=5)

gl.pcoa.plot(pca, gl, ellipse=TRUE, plevel=0.99, labels="pop",hadjust=1.5,vadjust=1)
gl.pcoa.plot(pca, gl, labels="none")
gl.pcoa.plot(pca, gl, ellipse=TRUE, plevel=0.99, labels="legend")
gl.pcoa.plot(pca, gl, ellipse=TRUE, plevel=0.99, labels="pop", xaxis=1, yaxis=3, zaxis=2)
gl.pcoa.plot(pca, gl, ellipse=TRUE, plevel=0.99, labels="none", xaxis=1, yaxis=3, zaxis=2)
gl.pcoa.plot(pca, gl, interactive=TRUE)  

}
\seealso{
\code{\link{gl.pcoa}}

Other Exploration/visualisation functions: 
\code{\link{gl.pcoa.plot.beta}()},
\code{\link{gl.plot}()}
}
\author{
Custodian: Arthur Georges -- Post to \url{https://groups.google.com/d/forum/dartr}
}
\concept{Exploration/visualisation functions}
