% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.pcoa.r
\name{gl.pcoa}
\alias{gl.pcoa}
\title{PCoA ordination (glPca)}
\usage{
gl.pcoa(x, nfactors = 5, parallel = FALSE, n.cores = 16,
  verbose = 2)
}
\arguments{
\item{x}{-- name of the genlight object containing the SNP data [required]}

\item{nfactors}{Number of dimensions to retain in the output file [default 5]}

\item{parallel}{TRUE if parallel processing is required (does fail under Windows) [default FALSE]}

\item{n.cores}{Number of cores to use if parallel processing is requested [default 16]}

\item{verbose}{-- specify the level of verbosity: 0, silent, fatal errors only; 1, flag function begin and end; 2, progress log ; 3, progress and results summary; 5, full report [default 2]}
}
\value{
An object of class glPca containing the eigenvalues, factor scores and factor loadings
}
\description{
This script takes the data on SNP genotypes for individuals and undertakes a Gower PCoa ordination using Euclidean distance and drawing upon
data in the original genlight \{adegenet\} object (entity x attribute matrix).
The script is essentially a wrapper for glPca() \{adegenet\} with default settings apart from setting parallel=FALSE and
converting the eigenvalues to percentages.
}
\examples{
pcoa <- gl.pcoa(testset.gl, nfactors=3)
}
\author{
Arthur Georges (Post to \url{https://groups.google.com/d/forum/dartr})
}
