% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.plot.heatmap.r
\name{gl.plot.heatmap}
\alias{gl.plot.heatmap}
\title{Represent a distance matrix as a heatmap}
\usage{
gl.plot.heatmap(D, ncolors = 5, labels = TRUE, labels.cex = 1,
  values = TRUE, values.cex = 1, legend = TRUE, rank = FALSE,
  verbose = NULL)
}
\arguments{
\item{D}{-- name of the distance matrix [required]}

\item{ncolors}{-- number of colors to display [default 5]}

\item{labels}{-- if TRUE, and the number of rows is <= 20, labels are added to the heatmap [default = TRUE]}

\item{labels.cex}{-- size of the labels [default = 1]}

\item{values}{-- if TRUE, and the number of rows is <= 20, distances are added to the body of the heatmap [default = TRUE]}

\item{values.cex}{-- size of the values to print inside each cell [default = 1]}

\item{legend}{-- if TRUE, a legend will be added to the plot [default = TRUE]}

\item{rank}{-- if TRUE, then the distance matrix will be reordered to group like with like, otherwise order will be displayed as given [default FALSE]}

\item{verbose}{-- verbosity: 0, silent or fatal errors; 1, begin and end; 2, progress log ; 3, progress and results summary; 5, full report [default 2]}
}
\description{
The script plots a heat map to represent the distances in the distance or dissimilarity matrix
}
\examples{
   gl <- testset.gl[1:10,]
   d <- dist(as.matrix(gl),upper=TRUE,diag=TRUE)
   gl.plot.heatmap(d)
   gl.plot.heatmap(d, ncolors=10, rank=TRUE, legend=TRUE)
}
\author{
Arthur Georges (Post to \url{https://groups.google.com/d/forum/dartr})
}
