% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.plot.r
\name{gl.plot}
\alias{gl.plot}
\title{Plotting genlight object}
\usage{
gl.plot(x, indlabels = indNames(x), col = NULL, legend = TRUE,
  posi = "bottomleft", bg = rgb(1, 1, 1, 0.5), ...)
}
\arguments{
\item{x}{-- a genlight object}

\item{indlabels}{-- labels for individuals. if not provided labels are taken from the first 8 letters from indNames.}

\item{col}{-- optional color vector (see ?glPlot)}

\item{legend}{-- a logical indicating whether a legend should be added}

\item{posi}{-- position of the legend}

\item{bg}{-- background color of the legend [default is transparent white]}

\item{...}{--- additional arguments passed to glPlot function.}
}
\description{
This function is based on the glPlot function from adegenet. It simply aims to put labels on the individuals and scales them accordingly (if there are not too many). 
For arguments please refer to the original adegenet function ?glPlot.
}
\examples{
gl.plot(foxes.gl[1:30,])
}
