% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.plot.structure.r
\name{gl.plot.structure}
\alias{gl.plot.structure}
\alias{gl.plot.structure_2}
\title{Plots a STRUCTURE analysis using a genlight object}
\usage{
gl.plot.structure_2(
  sr,
  K = NULL,
  met_clumpp = "greedyLargeK",
  iter_clumpp = 100,
  clumpak = TRUE,
  plot_theme = NULL,
  colors_clusters = NULL,
  ind_name = TRUE,
  border_ind = 0.15,
  plot.out = TRUE,
  verbose = NULL
)
}
\arguments{
\item{sr}{structure run object from \code{\link{gl.run.structure}} [required].}

\item{K}{The number for K the q matrix should be based on. Needs to
be within you simulated range of K's in your sr structure run object.}

\item{met_clumpp}{The algorithm to use to infer the correct permutations.
One of 'greedy' or 'greedyLargeK' or 'stephens' [default "greedy"].}

\item{iter_clumpp}{The number of iterations to use if running either 'greedy'
'greedyLargeK' [default 100].}

\item{clumpak}{[default TRUE].}

\item{plot_theme}{Theme for the plot. See Details for options
[default NULL].}

\item{colors_clusters}{A color palette for clusters (K) or a list with
as many colors as there are clusters (K)
[default NULL].}

\item{ind_name}{Whether to plot individual names [default TRUE].}

\item{border_ind}{The width of the border line between individuals [default 0.25].}

\item{plot.out}{Specify if plot is to be produced [default TRUE].}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2,
progress log ; 3, progress and results summary; 5, full report [default
 NULL, unless specified using gl.set.verbosity]}
}
\value{
a barplot in ggplot format
}
\description{
This function takes a structure run object (output from
 \code{\link{gl.run.structure}}) and plots the typical strcture bar
  plot that visiualise the q matrix of a structure run.
}
\details{
The function outputs a barplot which is the typical output of
 structure. This function needs the use of clumpp, which is an external
  program that needs to be installed For a evanno plot use gl.evanno.
}
\examples{
\dontrun{
#CLUMPP needs to be installed to be able to run the example
#bc <- bandicoot.gl[,1:100]
#sr <- gl.run.structure(bc, k.range = 2:5, num.k.rep = 3,
exec = './structure.exe')
#ev <- gl.evanno(sr)
#ev
#qmat <- gl.plot.structure(sr, K=3, CLUMPP='d:/structure/')
#head(qmat)
#gl.map.structure(qmat, bc, scalex=1, scaley=0.5)
}
}
\references{
\itemize{
\item Pritchard, J.K., Stephens, M., Donnelly, P. (2000) Inference of
population structure using multilocus genotype data. Genetics 155, 945-959.
\item Archer, F. I., Adams, P. E. and Schneiders, B. B. (2016) strataG: An R
package for manipulating, summarizing and analysing population genetic data.
Mol Ecol Resour. doi:10.1111/1755-0998.12559
\item Mattias Jakobsson and Noah A. Rosenberg. 2007. CLUMPP: a cluster
matching and permutation program for dealing with label switching and
multimodality in analysis of population structure. Bioinformatics
23(14):1801-1806. Available at
\href{http://web.stanford.edu/group/rosenberglab/clumppDownload.html}{clumpp}
}
}
\seealso{
\code{gl.run.structure},  \code{clumpp}, \code{gl.plot.structure}
}
\author{
Bernd Gruber (Post to \url{https://groups.google.com/d/forum/dartr})
}
