% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.read.vcf.r
\name{gl.read.vcf}
\alias{gl.read.vcf}
\title{Converts a vcf file into a genlight object}
\usage{
gl.read.vcf(vcffile, verbose = 1)
}
\arguments{
\item{vcffile}{-- a vcf file (works only for diploid data)}

\item{verbose}{set to 2}
}
\value{
A genlight object, with most slots filled.
}
\description{
This function needs package vcfR. At the time of writing (August 2020) the package was no longer available from CRAN. To install the package check their github repository. \url{https://github.com/knausb/vcfR}.
}
\author{
Bernd Gruber (Post to \url{https://groups.google.com/d/forum/dartr})
}
