% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.read.vcf.r
\name{gl.read.vcf}
\alias{gl.read.vcf}
\title{Converts a vcf file into a genlight object}
\usage{
gl.read.vcf(vcffile, verbose = 2)
}
\arguments{
\item{vcffile}{-- a vcf file (works only for diploid data)}

\item{verbose}{set to 2}
}
\value{
A genlight object.
}
\description{
This function needs package vcfR, please install it. The converted genlight object does not have individual metrics. You need to add them 'manually' to the other$ind.metrics slot.
}
\examples{
\dontrun{
test <- gl.read.vcf("your_data.vcf")
}
}
\author{
Bernd Gruber (Post to \url{https://groups.google.com/d/forum/dartr})
}
