% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.report.bases.r
\name{gl.report.bases}
\alias{gl.report.bases}
\title{Report summary of base pair frequencies}
\usage{
gl.report.bases(
  x,
  plot = TRUE,
  plot_theme = theme_dartR(),
  plot_colours = two_colors,
  verbose = NULL
)
}
\arguments{
\item{x}{Name of the genlight object containing the SNP or presence/absence (SilicoDArT) data [required]}

\item{plot}{If TRUE, histograms of base composition are produced [default TRUE]}

\item{plot_theme}{Theme for the plot. See Details for options [default theme_dartR()]}

\item{plot_colours}{Two color names for borders and fill of the plots [default two_colors]}

\item{verbose}{-- verbosity: 0, silent or fatal errors; 1, begin and end; 2, progress log ; 3, progress and results summary; 5, full report [default 2, unless specified using gl.set.verbosity]}
}
\value{
A list containing 
        [[1]] $freq -- the table of base frequencies and transition/transversion ratios;
        [[2]] $plotbases -- ggplot bargraph of base frequencies;
        [[3]] $plottstv -- ggplot bargraph of transitions and transversions.
}
\description{
This script calculates the frequencies of the four DNA nucleotide bases: adenine (A), cytosine (C), 
guanine (G) and thymine (T), and the frequency of transitions (Ts) and transversions (Tv) in a DArT genlight object.
}
\details{
The script checks first if trimmed sequences are included in the locus metadata (@other$loc.metrics$TrimmedSequence), 
and if so, tallies up the numbers of A, T, G and C bases. Only the reference state at the SNP locus is counted. Counts of 
transitions (Ts) and transversions (Tv) assume that there is no directionality, that is C->T is the same as T->C, because
the reference state is arbitrary.

For presence/absence data (SilicoDArT), it is not possible to count transversions or transitions or transversions/transitions ratio
because the SNP data is not available, only a single sequence tag.

Examples of other themes that can be used can be consulted in \itemize{
 \item \url{https://ggplot2.tidyverse.org/reference/ggtheme.html} and \item
 \url{https://yutannihilation.github.io/allYourFigureAreBelongToUs/ggthemes/}
 }
}
\examples{
# SNP data
  out <- gl.report.bases(testset.gl)
  out$freq
  out$plotbases
  out$plottvts
# Tag P/A data
  out <- gl.report.bases(testset.gs)
  out

}
\author{
Core dartR Team (Post to \url{https://groups.google.com/d/forum/dartr})
}
