% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.report.callrate.r
\name{gl.report.callrate}
\alias{gl.report.callrate}
\title{Report summary of Call Rate for loci or individuals}
\usage{
gl.report.callrate(x, method = "loc", plot = TRUE, v = 2)
}
\arguments{
\item{x}{-- name of the genlight or genind object containing the SNP data [required]}

\item{method}{specify the type of report by locus (method="loc") or individual (method="ind") [default method="loc"]}

\item{plot}{specify if a histogram of call rate is to be produced [default TRUE]}

\item{v}{-- verbosity: 0, silent or fatal errors; 1, begin and end; 2, progress log ; 3, progress and results summary; 5, full report [default 2]}
}
\value{
Mean call rate by locus (method="loc") or individual (method="ind")
}
\description{
SNP datasets generated by DArT have missing values primarily arising from failure to call a SNP because of a mutation
at one or both of the the restriction enzyme recognition sites. This script reports the number of missing values for each
of several percentiles. The script gl.filter.callrate() will filter out the loci with call rates below a specified threshold.
}
\examples{
gl.report.callrate(testset.gl)
}
\author{
Arthur Georges (glbugs@aerg.canberra.edu.au)
}
