% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.report.callrate.r
\name{gl.report.callrate}
\alias{gl.report.callrate}
\title{Report summary of Call Rate for loci or individuals}
\usage{
gl.report.callrate(
  x,
  method = "loc",
  plot_theme = theme_dartR(),
  plot_colours = two_colors,
  verbose = options()$dartR_verbose
)
}
\arguments{
\item{x}{Name of the genlight object containing the SNP or presence/absence
(SilicoDArT) data [required].}

\item{method}{Specify the type of report by locus (method='loc') or individual
(method='ind') [default method='loc'].}

\item{plot_theme}{Theme for the plot. See Details for options [default theme_dartR()].}

\item{plot_colours}{List of two color names for the borders and fill of the
plots [default two_colors].}

\item{verbose}{-- verbosity: 0, silent or fatal errors; 1, begin and end; 2, progress log ; 3, progress and results summary; 5, full report [default 2 or as specified using gl.set.verbosity]}
}
\value{
Returns an unchanged genlight object
}
\description{
SNP datasets generated by DArT have missing values primarily
arising from failure to call a SNP because of a mutation at one or both of
the the restriction enzyme recognition sites. This function reports the
number of missing values for each of several quantiles. Quantiles are
partitions of a finite set of values into q subsets of (nearly) equal sizes.
In this function q = 20. Quantiles are useful measures because they are less
 susceptible to long-tailed distributions and outliers.
}
\details{
The function \code{\link{gl.filter.callrate}} will filter out the
 loci with call rates below a specified threshold.

 Tag Presence/Absence datasets (SilicoDArT) have missing values where it is
 not possible to determine reliably if the sequence tag can be called at a
 particular locus.
 
\strong{ Function's output }

 The minimum, maximum, mean and a tabulation of call rate quantiles against
 thresholds rate are provided. Output also includes a boxplot and a
 histogram to guide in the selection of a threshold for filtering on callrate.

 Plots and table are saved to the temporal directory (tempdir) and can be accessed with the function \code{\link{gl.print.reports}} and listed with the function \code{\link{gl.list.reports}}. Note that they can be accessed only in the current R session because tempdir is cleared each time that the R session is closed.
  
 Examples of other themes that can be used can be consulted in \itemize{
 \item \url{https://ggplot2.tidyverse.org/reference/ggtheme.html} and \item
 \url{https://yutannihilation.github.io/allYourFigureAreBelongToUs/ggthemes/}
 }
}
\examples{
# SNP data
  gl.report.callrate(testset.gl)
  gl.report.callrate(testset.gl,method="ind")
# Tag P/A data
  gl.report.callrate(testset.gs)
  gl.report.callrate(testset.gs,method="ind")

}
\seealso{
\code{\link{gl.filter.callrate}}, \code{\link{gl.list.reports}},
 \code{\link{gl.print.reports}}

Other filters and filter reports: 
\code{\link{gl.report.locmetric}()},
\code{\link{gl.report.maf}()}
}
\author{
Arthur Georges (Post to \url{https://groups.google.com/d/forum/dartr})
}
\concept{filters and filter reports}
