% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.report.callrate.r
\name{gl.report.callrate}
\alias{gl.report.callrate}
\title{Report summary of Call Rate for loci or individuals}
\usage{
gl.report.callrate(gl, method = "loc")
}
\arguments{
\item{gl}{-- name of the genlight or genind object containing the SNP data [required]}

\item{method}{specify the type of report by locus (method="loc") or individual (method="ind") [default method="loc"]}
}
\value{
Mean call rate by locus (method="loc") or individual (method="ind")
}
\description{
SNP datasets generated by DArT have missing values primarily arising from failure to call a SNP because of a mutation
at one or both of the the restriction enzyme recognition sites. This script reports the number of missing values for each
of several percentiles. The script gl.filter.callrate() will filter out the loci with call rates below a specified threshold.
}
\examples{
\dontrun{
result <- gl.report.callrate(gl)
}
}
\author{
Arthur Georges (glbugs@aerg.canberra.edu.au)
}

