% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.report.callrate.r
\name{gl.report.callrate}
\alias{gl.report.callrate}
\title{Report summary of Call Rate for loci or individuals}
\usage{
gl.report.callrate(
  x,
  method = "loc",
  plot.out = TRUE,
  plot_theme = theme_dartR(),
  plot_colors = two_colors,
  save2tmp = FALSE,
  verbose = NULL
)
}
\arguments{
\item{x}{Name of the genlight object containing the SNP or presence/absence
(SilicoDArT) data [required].}

\item{method}{Specify the type of report by locus (method='loc') or individual
(method='ind') [default 'loc'].}

\item{plot.out}{Specify if plot is to be produced [default TRUE].}

\item{plot_theme}{User specified theme [default theme_dartR()].}

\item{plot_colors}{Vector with two colour names for the borders and fill 
[default two_colors].}

\item{save2tmp}{If TRUE, saves any ggplots and listings to the session 
temporary directory (tempdir) [default FALSE].}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2, 
progress log ; 3, progress and results summary; 5, full report 
[default 2, unless specified using gl.set.verbosity].}
}
\value{
Returns unaltered genlight object
}
\description{
SNP datasets generated by DArT have missing values primarily arising from 
failure to call a SNP because of a mutation at one or both of the restriction
enzyme recognition sites. P/A datasets (SilicoDArT) have missing values 
because it was not possible to call whether a sequence tag was amplified or 
not. This function tabulates the number of missing values as quantiles.
}
\details{
This function expects a genlight object, containing either SNP data or 
SilicoDArT (=presence/absence data).

Callrate is summarized by locus or by individual to allow sensible decisions 
on thresholds for filtering taking into consideration consequential loss of 
data. The summary is in the form of a tabulation and plots.

Plot themes can be obtained from \itemize{
 \item \url{https://ggplot2.tidyverse.org/reference/ggtheme.html} and \item
 \url{https://yutannihilation.github.io/allYourFigureAreBelongToUs/ggthemes/}
 }

Resultant ggplots and the tabulation are saved to the session's temporary 
directory.
}
\examples{
# SNP data
  gl.report.callrate(testset.gl)
  gl.report.callrate(testset.gl,method="ind")
# Tag P/A data
  gl.report.callrate(testset.gs)
  gl.report.callrate(testset.gs,method="ind")

}
\seealso{
\code{\link{gl.filter.callrate}}

Other filters and filter reports: 
\code{\link{gl.filter.allna}()},
\code{\link{gl.filter.monomorphs}()},
\code{\link{gl.filter.rdepth}()},
\code{\link{gl.filter.reproducibility}()},
\code{\link{gl.filter.secondaries}()},
\code{\link{gl.report.hamming}()},
\code{\link{gl.report.locmetric}()},
\code{\link{gl.report.monomorphs}()},
\code{\link{gl.report.rdepth}()},
\code{\link{gl.report.reproducibility}()},
\code{\link{gl.report.taglength}()}
}
\author{
Custodian: Arthur Georges -- Post to \url{https://groups.google.com/d/forum/dartr}
}
\concept{filters and filter reports}
