% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.report.heterozygosity.r
\name{gl.report.heterozygosity}
\alias{gl.report.heterozygosity}
\title{Reports hetrozygosity}
\usage{
gl.report.heterozygosity(gl)
}
\arguments{
\item{gl}{-- a genlight object containing the SNP genotypes [Required]}
}
\value{
a dataframe containing population labels, observed heterozygosities and sample sizes
}
\description{
Calculates the observed heterozygisities by population from a genlight object
and plots as a barchart ordered on heterozygosity.
}
\details{
#Input is a genlight {adegenet} object containing SNP genotypes (0 homozygous for reference SNP, 
#1 heterozygous, 2 homozygous for alternate SNP).
}
\examples{
gl.report.heterozygosity(testset.gl)
}
\author{
Bernd Gruber (bugs? Post to https://groups.google.com/d/forum/dartr)
}
