% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.report.heterozygosity.r
\name{gl.report.heterozygosity}
\alias{gl.report.heterozygosity}
\title{Reports observed and expected hetrozygosity by population}
\usage{
gl.report.heterozygosity(x, plot = FALSE, cex.labels = 0.7,
  verbose = 2)
}
\arguments{
\item{x}{-- a genlight object containing the SNP genotypes [Required]}

\item{plot}{-- if TRUE, plots a barchart of observed heterozygosity [default FALSE]}

\item{cex.labels}{-- sets the size of the population labels [default 0.7]}

\item{verbose}{-- verbosity: 0, silent or fatal errors; 1, begin and end; 2, progress log ; 3, progress and results summary; 5, full report [default 2]}
}
\value{
a dataframe containing population labels, heterozygosities and sample sizes
}
\description{
Calculates the observed and expected heterozygisities by population from a genlight object
and plots as an ordered barchart of observed heterozygosity. Also calculates the estimated variance
for expected heterozygosity.
}
\details{
Expected heterozygosity is calculated according to Nei, M. (1987) Molecular evolutionary genetics. New York: Columbia University Press, using the Pegas R package.
}
\author{
Bernd Gruber (Post to \url{https://groups.google.com/d/forum/dartr})
}
