% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.report.heterozygosity.r
\name{gl.report.heterozygosity}
\alias{gl.report.heterozygosity}
\title{Reports observed and expected heterozygosity by population or by individual from SNP data}
\usage{
gl.report.heterozygosity(
  x,
  method = "pop",
  n.invariant = 0,
  plot_theme = theme_dartR(),
  plot_colours_pop = discrete_palette,
  plot_colours_ind = two_colors,
  verbose = options()$dartR_verbose
)
}
\arguments{
\item{x}{Name of the genlight object containing the SNP [required].}

\item{method}{Calculate heterozygosity by population (method='pop') or by individual (method='ind') [default 'pop']}

\item{n.invariant}{An estimate of the number of invariant sequence tags used to adjust the heterozygosity rate [default 0]}

\item{plot_theme}{Theme for the plot. See Details for options [default theme_dartR()].}

\item{plot_colours_pop}{A color palette for population plots [default discrete_palette].}

\item{plot_colours_ind}{List of two color names for the borders and fill of the plot by individual [default two_colors].}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2, progress log; 
3, progress and results summary; 5, full report [default 2 or as specified using gl.set.verbosity]}
}
\value{
A dataframe containing population labels, heterozygosities and sample sizes
}
\description{
Calculates the observed and expected heterozygosities for each population
or the observed heterozygosity for each individual in a genlight object.
}
\details{
Observed heterozygosity for a population takes the proportion of heterozygous
loci for each individual then averages over the individuals in that population. 
The calculations take into account missing values.

Expected heterozygosity for a population takes the expected proportion of
heterozygotes, that is, expected under Hardy-Weinberg equilibrium, for each locus, then
averages this across the loci for an average estimate for the population.
The calculations of expected heterozygosity use the unbiassed estimates of Nei, M. (1987) 

Output for method='pop' is an ordered barchart of observed heterozygosity across 
populations together with a table of observed and expected heterozygosity by population.

Observed heterozygosity for individuals is calculated as the proportion of loci that
are heterozygous for that individual.

Finally, the loci that are invariant across all individuals in the dataset (that is,
across populations), is typically unknown. This can render estimates of heterozygosity
analysis specific, and so it is not valid to compare such estimates across species
or even across different analyses. This is a similar problem faced by microsatellites.
If you have an estimate of the number of invariant sequence tags (loci) in your data,
such as provided by gl.report.secondaries, you can specify it with the n.invariant
parameter to standardize your estimates of heterozygosity.

\strong{ Function's output }
Output for method='ind' is a histogram and a boxplot of heterozygosity across individuals.

 Plots and table are saved to the temporal directory (tempdir) and can be accessed with the function \code{\link{gl.print.reports}} and listed with the function \code{\link{gl.list.reports}}. Note that they can be accessed only in the current R session because tempdir is cleared each time that the R session is closed.

 Examples of other themes that can be used can be consulted in \itemize{
 \item \url{https://ggplot2.tidyverse.org/reference/ggtheme.html} and \item
 \url{https://yutannihilation.github.io/allYourFigureAreBelongToUs/ggthemes/}
 }
}
\examples{
df <- gl.report.heterozygosity(testset.gl)
df <- gl.report.heterozygosity(testset.gl,method='ind')

}
\references{
Nei, M. and R. K. Chesser (1983). 'Estimation of fixation indices and gene diversities.' Annals of human genetics 47(3): 253-259.
}
\seealso{
\code{\link{gl.filter.heterozygosity}},\code{\link{gl.list.reports}},
 \code{\link{gl.print.reports}}

Other reporting functions: 
\code{\link{gl.report.bases}()},
\code{\link{gl.report.diversity}()},
\code{\link{gl.report.parent.offspring}()},
\code{\link{gl.report.pa}()}
}
\author{
Bernd Gruber, Arthur Georges and Renee Catullo (Post to \url{https://groups.google.com/d/forum/dartr})
}
\concept{reporting functions}
