% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.report.hwe.r
\name{gl.report.hwe}
\alias{gl.report.hwe}
\title{Reports departure from Hardy-Weinberg Equilibrium}
\usage{
gl.report.hwe(
  x,
  subset = "each",
  plot.out = FALSE,
  multi_comp = TRUE,
  multi_comp_method = "bonferroni",
  method_sig = "Exact",
  alpha_val = 0.05,
  sample_size = 5,
  verbose = NULL
)
}
\arguments{
\item{x}{Name of the genlight object containing the SNP data [required].}

\item{subset}{Either a vector with population names, "each", "all" (see details)
[default "each"].}

\item{plot.out}{If TRUE, will produce a Ternary Plot(s) [default FALSE].}

\item{multi_comp}{Whether to adjust p-values for multiple comparisons [default TRUE].}

\item{multi_comp_method}{Method to adjust p-values for multiple comparisons: 
"holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr"
(see details) [default "bonferroni"]]}

\item{method_sig}{Method for determining statistical significance: "ChiSquare" 
or "Exact" [default "Exact"].}

\item{alpha_val}{Level of significance for testing [default 0.05].}

\item{sample_size}{= 5}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2, 
progress log ; 3, progress and results summary; 5, full report 
[default NULL, unless specified using gl.set.verbosity]}
}
\value{
returns a dataframe containing loci, counts of reference SNP homozygotes,
heterozygotes and alternate SNP homozygotes; probability of departure from 
H-W equilibrium, and per locus significance with and without Bonferroni correction.
}
\description{
Calculates the probabilities of agreement with H-W equilibrium based on observed
frequencies of reference homozygotes, heterozygotes and alternate homozygotes.
}
\details{
Factors that can cause deviations from Hardy-Weinberg proportions include: 
mutation, finite population size, selection, population structure, age 
structure, assortative mating, sex linkage, nonrandom sampling and genotyping
 errors. Testing for Hardy-Weinberg proportions should be a process that involves
 a careful evaluation of the results.  
 
This function applies HWE tests to each locus in three ways:
\itemize{
\item across all populations together using: subset="all"
\item within each population separately using: subset="each"
\item within selected populations using for example: subset=c("pop1","pop2")
}

This function tests for deviations from Hardy Weinberg proportions using two 
different methods:
\itemize{
\item "ChiSquare"   \code{\link[HardyWeinberg]{HWChisqStats}}
\item "Exact"  Uses the exact calculations contained in the function 
\code{\link[HardyWeinberg]{HWExactStats}} as developed by Wigginton 
et al. (2005). The C++ code was generously shared by Christopher Chang, and the same code is 
used in the program PLINK (2.0).
}

This function uses correction for multiple tests using the following methods
\code{\link[stats]{p.adjust}}:
\itemize{
\item "holm", 
\item "hochberg", 
\item "hommel", 
\item "bonferroni", 
\item "BH", 
\item "BY", 
\item "fdr"
}
 
Tests for HWE are only valid if there is no population substructure (assuming
random mating), and the tests have sufficient power only when there is 
sufficient sample size (say, n individuals > 20). Note also that correction 
for multiple comparisons is probably required if you wish to place particular
importance on one or a few significant departures.

A Ternary Plot is optionally produced (plot.out = TRUE) -- see Graffelman et al.(2008) for 
further details. Implementation of the Ternary Plot is via package {HardyWeinberg} 
(Graffelman (2015). The plot labels loci that depart significantly from HWE 
as red, and those not showing significant departure as green. 
HWTernaryPlot is a routine that draws a ternary plot for three-way genotypic compositions (AA,AB,BB), and represents the acceptance region for different tests for Hardy-Weinberg equilibrium (HWE) in the plot.

say the monomorphic loci are filtered
populations with less than 5 loci cannot be assessed tests for this in all cases

p.adjust {stats}

p.adjust.methods
colour the marker points automatically according to the result of a 
significance test (green markers non-significant, red markers significant).
#' Uses the exact calculations contained in function \code{\link{utils.prob.hwe}}
as developed by Wigginton et al. (2005).
}
\examples{
gl.report.hwe(x = bandicoot.gl,plot.out = FALSE)
}
\references{
\itemize{
\item Wigginton, J.E., Cutler, D.J., & Abecasis, G.R. (2005). A Note on Exact 
Tests of Hardy-Weinberg Equilibrium. American Journal of Human Genetics 76:887-893.
\item Graffelman, J. & Morales-Camarena, J. (2008). Graphical tests for 
Hardy-Weinberg equilibrium based on the ternary plot. Human Heredity 65:77-84.
\item Graffelman, J. (2015). Exploring Diallelic Genetic Markers: The Hardy 
Weinberg Package. Journal of Statistical Software 64:1-23.
\item Waples, R. S. (2015). Testing for Hardy–Weinberg proportions: have we 
lost the plot?. Journal of heredity, 106(1), 1-19.
}
}
\seealso{
\code{\link{gl.filter.hwe}}

Other filters/filter reports: 
\code{\link{gl.filter.hwe}()}
}
\author{
Arthur Georges -- Post to \url{https://groups.google.com/d/forum/dartr}
}
\concept{filters/filter reports}
