% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.report.hwe.r
\name{gl.report.hwe}
\alias{gl.report.hwe}
\title{Reports departure from Hardy-Weinberg Equilibrium}
\usage{
gl.report.hwe(gl, p = 0.05, pop = "all")
}
\arguments{
\item{gl}{-- a genlight object containing the SNP genotypes [Required]}

\item{p}{-- level of significance (per locus) [Default 0.05]}

\item{pop}{-- populations to combine in the analysis [Default "all"]}
}
\value{
a dataframe containing loci, counts of reference SNP homozygotes, heterozygotes
and alternate SNP homozygotes; probability of departure from H-W equilibrium,
and per locus significance.
}
\description{
Calculates the probabilities of agreement with H-W equilibrium based on observed
frequencies of reference homozygotes, heterozygotes and alternate homozygotes. 
Uses the exact calculations contained in function prob.hwe() as developed by
Wigginton, JE, Cutler, DJ, and Abecasis, GR.
}
\details{
#Input is a genlight {adegenet} object containing SNP genotypes (0 homozygous for reference SNP, 
#1 heterozygous, 2 homozygous for alternate SNP).
}
\examples{
result <- gl.report.hwe(testset.gl,0.05, pop=c("EmmacMaclGeor", "EmmacCoopCully"))
}
\author{
Arthur Georges (glbugs@aerg.canberra.edu.au)
}

