% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.report.maf.r
\name{gl.report.maf}
\alias{gl.report.maf}
\title{Report minor allele frequency (MAF) for each locus in a genlight {adegenet} object}
\usage{
gl.report.maf(x, maf.limit = 0.5, ind.limit = 5, loc.limit = 30,
  v = 2)
}
\arguments{
\item{x}{-- name of the genlight object containing the SNP data [required]}

\item{maf.limit}{-- show histograms maf range <= maf.limit [default 0.5]}

\item{ind.limit}{-- show histograms only for populations of size greater than ind.limit [default 5]}

\item{loc.limit}{-- show histograms only for populations with more than loc.limit polymorphic loci [default 30]}

\item{v}{-- verbosity: 0, silent or fatal errors; 1, begin and end; 2, progress log ; 3, progress and results summary; 5, full report [default 2]}
}
\description{
Summary of minor allele frequencies across loci is reported as histograms.
}
\examples{
f <- gl.report.maf(testset.gl)
}
\author{
Arthur Georges (glbugs@aerg.canberra.edu.au)
}
