% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.report.monomorphs.r
\name{gl.report.monomorphs}
\alias{gl.report.monomorphs}
\title{Report monomorphic loci, including those with all NAs}
\usage{
gl.report.monomorphs(gl)
}
\arguments{
\item{gl}{-- name of the input genlight object [required]}
}
\value{
A report on loci, polymorphic, monomorphic, all NAs
}
\description{
This script reports the number of monomorphic loci from a genlight \{adegenet\} object
}
\details{
A DArT dataset will not have monomorphic loci, but they can arise when populations or individuals are deleted.
Retaining monomorphic loci unnecessarily increases the size of the dataset.
}
\examples{
gl2 <- gl.report.monomorphs(testset.gl)
}
\author{
Arthur Georges (bugs? Post to \url{https://groups.google.com/d/forum/dartr})
}
