% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.report.monomorphs.r
\name{gl.report.monomorphs}
\alias{gl.report.monomorphs}
\title{Report monomorphic loci}
\usage{
gl.report.monomorphs(x, verbose = NULL)
}
\arguments{
\item{x}{Name of the input genlight object [required].}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2,
progress log; 3, progress and results summary; 5, full report
[default NULL, unless specified using gl.set.verbosity].}
}
\value{
An unaltered genlight object
}
\description{
This script reports the number of monomorphic loci and those with all NAs in
a genlight \{adegenet\} object
}
\details{
A DArT dataset will not have monomorphic loci, but they can arise, along with
loci that are scored all NA, when populations or individuals are deleted.
Retaining monomorphic loci unnecessarily increases the size of the dataset
and will affect some calculations.

Note that for SNP data, NAs likely represent null alleles; in tag
presence/absence data, NAs represent missing values (presence/absence could
not be reliably scored)
}
\examples{
# SNP data
  gl.report.monomorphs(testset.gl)
# SilicoDArT data
  gl.report.monomorphs(testset.gs)
}
\seealso{
\code{\link{gl.filter.monomorphs}}

Other filters and filter reports: 
\code{\link{gl.filter.allna}()},
\code{\link{gl.filter.monomorphs}()},
\code{\link{gl.filter.rdepth}()},
\code{\link{gl.filter.reproducibility}()},
\code{\link{gl.filter.secondaries}()},
\code{\link{gl.report.callrate}()},
\code{\link{gl.report.hamming}()},
\code{\link{gl.report.locmetric}()},
\code{\link{gl.report.rdepth}()},
\code{\link{gl.report.reproducibility}()},
\code{\link{gl.report.taglength}()}
}
\author{
Custodian: Arthur Georges -- Post to
 \url{https://groups.google.com/d/forum/dartr}
}
\concept{filters and filter reports}
