% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.report.rdepth.r
\name{gl.report.rdepth}
\alias{gl.report.rdepth}
\title{Report summary of Read Depth for each locus}
\usage{
gl.report.rdepth(
  x,
  plot_theme = theme_dartR(),
  plot_colours = two_colors,
  verbose = NULL
)
}
\arguments{
\item{x}{Name of the genlight object containing the SNP or presence/absence
(SilicoDArT) data [required].}

\item{plot_theme}{Theme for the plot. See Details for options [default theme_dartR()].}

\item{plot_colours}{List of two color names for the borders and fill of the
plots [default two_colors].}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2, progress log ; 3, progress and results summary; 5, full report [default NULL, unless specified using gl.set.verbosity]}
}
\value{
An unaltered genlight object
}
\description{
SNP datasets generated by DArT report AvgCountRef and AvgCountSnp as counts 
of sequence tags for the reference and alternate alleles respectively.
These can be used to back calculate Read Depth. Fragment presence/absence 
datasets as provided by DArT (SilicoDArT) provide Average Read Depth and 
Standard Deviation of Read Depth as standard columns in their report. This 
function reports the read depth by locus for each of several quantiles.
}
\details{
Filtering on Read Depth using the companion function \code{\link{gl.filter.rdepth}}
 can be on the basis of loci with exceptionally low counts, 
or loci with exceptionally high counts.

 Quantiles are
partitions of a finite set of values into q subsets of (nearly) equal sizes.
In this function q = 20. Quantiles are useful measures because they are less
 susceptible to long-tailed distributions and outliers.
 
\strong{ Function's output }

 The minimum, maximum, mean and a tabulation of read depth quantiles against
 thresholds rate are provided. Output also includes a boxplot and a
 histogram to guide in the selection of a threshold for filtering on callrate.

 Plots and table are saved to the temporal directory (tempdir) and can be 
 accessed with the function \code{\link{gl.print.reports}} and listed with 
 the function \code{\link{gl.list.reports}}. Note that they can be accessed 
 only in the current R session because tempdir is cleared each time that the
  R session is closed.
  
 Examples of other themes that can be used can be consulted in \itemize{
 \item \url{https://ggplot2.tidyverse.org/reference/ggtheme.html} and \item
 \url{https://yutannihilation.github.io/allYourFigureAreBelongToUs/ggthemes/}
 }
}
\examples{
# SNP data
df <- gl.report.rdepth(testset.gl)

}
\seealso{
\code{\link{gl.filter.rdepth}}, \code{\link{gl.list.reports}},
 \code{\link{gl.print.reports}}

Other filters and filter reports: 
\code{\link{gl.filter.callrate}()},
\code{\link{gl.report.callrate}()},
\code{\link{gl.report.hamming}()},
\code{\link{gl.report.locmetric}()},
\code{\link{gl.report.maf}()},
\code{\link{gl.report.monomorphs}()},
\code{\link{gl.report.overshoot}()},
\code{\link{gl.report.reproducibility}()},
\code{\link{gl.report.secondaries}()},
\code{\link{gl.report.taglength}()}
}
\author{
Arthur Georges (Post to \url{https://groups.google.com/d/forum/dartr})
}
\concept{filters and filter reports}
