% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.report.repavg.r
\name{gl.report.repavg}
\alias{gl.report.repavg}
\title{Report summary of RepAvg, repeatability averaged over both alleles for each locus in a genlight {adegenet} object}
\usage{
gl.report.repavg(x, plot = FALSE, smearplot = FALSE)
}
\arguments{
\item{x}{-- name of the genlight object containing the SNP data [required]}

\item{plot}{if TRUE, will produce a histogram of call rate [default FALSE]}

\item{smearplot}{if TRUE, will produce a smearplot of individuals against loci [default FALSE]}
}
\value{
-- Tabulation of repAvg against prospective Thresholds
}
\description{
SNP datasets generated by DArT have in index, RepAvg, generated by reproducing the data independently for 30% of loci.
RepAvg is the proportion of alleles that give a repeatable result, averaged over both alleles for each locus.
}
\details{
A histogram and or a smearplot can be requested. Note that the smearplot is computationally intensive, and will take time to 
execute for large datasets.
}
\examples{
gl.report.repavg(testset.gl)
}
\author{
Arthur Georges (Post to \url{https://groups.google.com/d/forum/dartr})
}
