% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.report.repavg.r
\name{gl.report.repavg}
\alias{gl.report.repavg}
\title{Report summary of RepAvg, reproducibility averaged over both alleles for each locus in a genlight {adegenet} object}
\usage{
gl.report.repavg(gl)
}
\arguments{
\item{gl}{-- name of the genlight object containing the SNP data [required]}
}
\value{
-- the mean call rate
}
\description{
SNP datasets generated by DArT have in index, RepAvg, generated by reproducing the data independently for 30% of loci.
RepAvg is the proportion of alleles that give a reproducible result, averaged over both alleles for each locus.
}
\examples{
gl.report.repavg(testset.gl)
}
\author{
Arthur Georges (bugs? Post to https://groups.google.com/d/forum/dartr)
}
