% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.report.reproducibility.r
\name{gl.report.reproducibility}
\alias{gl.report.reproducibility}
\title{Report summary of RepAvg (repeatability averaged over both alleles for 
each locus) or reproducibility (repeatability of the scores for fragment presence/absence)}
\usage{
gl.report.reproducibility(
  x,
  plot_theme = theme_dartR(),
  plot_colours = two_colors,
  verbose = options()$dartR_verbose
)
}
\arguments{
\item{x}{Name of the genlight object containing the SNP or presence/absence
(SilicoDArT) data [required].}

\item{plot_theme}{Theme for the plot. See Details for options [default theme_dartR()].}

\item{plot_colours}{List of two color names for the borders and fill of the
plots [default two_colors].}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2, progress log ; 3, progress and results summary; 5, full report [default 2 or as specified using gl.set.verbosity].}
}
\value{
An unaltered genlight object
}
\description{
SNP datasets generated by DArT have an index, RepAvg, generated by reproducing 
the data independently for 30% of loci. RepAvg is the proportion of alleles that 
give a repeatable result, averaged over both alleles for each locus.

In the case of fragment presence/absence data (SilicoDArT), repeatability is
 the percentage of scores that are repeated in the technical replicate dataset.
}
\details{
The function \code{\link{gl.filter.reproducibility}} will filter out the
 loci with repeatability below a specified threshold.
 
 Quantiles are
partitions of a finite set of values into q subsets of (nearly) equal sizes.
In this function q = 20. Quantiles are useful measures because they are less
 susceptible to long-tailed distributions and outliers.
 
\strong{ Function's output }

 The minimum, maximum, mean and a tabulation of repeatability quantiles against
 thresholds are provided. Output also includes a boxplot and a
 histogram to guide in the selection of a threshold for filtering on repeatability.

 Plots and table are saved to the temporal directory (tempdir) and can be accessed with the function \code{\link{gl.print.reports}} and listed with the function \code{\link{gl.list.reports}}. Note that they can be accessed only in the current R session because tempdir is cleared each time that the R session is closed.
  
 Examples of other themes that can be used can be consulted in \itemize{
 \item \url{https://ggplot2.tidyverse.org/reference/ggtheme.html} and \item
 \url{https://yutannihilation.github.io/allYourFigureAreBelongToUs/ggthemes/}
 }
}
\examples{
# SNP data
  out <- gl.report.reproducibility(testset.gl)
# Tag P/A data
  out <- gl.report.reproducibility(testset.gs)

}
\seealso{
\code{\link{gl.filter.reproducibility}}, \code{\link{gl.list.reports}},
 \code{\link{gl.print.reports}}

Other filters and filter reports: 
\code{\link{gl.report.callrate}()},
\code{\link{gl.report.hamming}()},
\code{\link{gl.report.locmetric}()},
\code{\link{gl.report.maf}()},
\code{\link{gl.report.monomorphs}()},
\code{\link{gl.report.overshoot}()},
\code{\link{gl.report.rdepth}()},
\code{\link{gl.report.secondaries}()}
}
\author{
Arthur Georges (Post to \url{https://groups.google.com/d/forum/dartr})
}
\concept{filters and filter reports}
