% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.report.secondaries.r
\name{gl.report.secondaries}
\alias{gl.report.secondaries}
\title{Report loci containing secondary SNPs in sequence tags}
\usage{
gl.report.secondaries(
  x,
  plot_theme = theme_dartR(),
  plot_colours = two_colors,
  verbose = NULL
)
}
\arguments{
\item{x}{Name of the genlight object containing the SNP data [required].}

\item{plot_theme}{Theme for the plot. See Details for options [default theme_dartR()].}

\item{plot_colours}{List of two color names for the borders and fill of the
plots [default two_colors].}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2, progress log ; 3, progress and results summary; 5, full report [default NULL, unless specified using gl.set.verbosity]}
}
\value{
A genlight object containing only those loci with secondaries
}
\description{
SNP datasets generated by DArT include fragments with more than one SNP (that
 is, with secondaries) and record them separately with the same CloneID (=AlleleID).
These multiple SNP loci within a fragment are likely to be linked, and so you may 
wish to remove secondaries.

This function reports statistics associated with secondaries, and the consequences 
of filtering them out, and provides three plots. The first is a boxplot, the second is a barplot of the frequency 
of secondaries per sequence tag, and the third is the Poisson expectation for those
frequencies including an estimate of the zero class (no. of sequence tags with 
no SNP scored).
}
\details{
The function \code{\link{gl.filter.secondaries}} will filter out the
 loci with secondaries.
 
Heterozygosity as estimated by the function \code{\link{gl.report.heterozygosity}} 
is in a sense relative, because it is calculated
against a background of only those loci that are polymorphic somewhere in the dataset.
To allow intercomparability across studies and species, any measure of heterozygosity
needs to accommodate loci that are invariant. However, the number of invariant loci
are unknown given the SNPs are detected as single point mutational variants and 
invariant sequences are discarded, and because of
the particular additional filtering pre-analysis. Modelling the counts
of SNPs per sequence tag as a Poisson distribution in this script allows estimate of the zero class,
that is, the number of invariant loci. This is reported, and the veracity of the 
estimate can be assessed by the correspondence of the observed frequencies against
those under Poisson expectation in the associated graphs. The number of invariant loci can then be optionally
provided to the function \code{\link{gl.report.heterozygosity}} via the parameter n.invariants.
 
\strong{ Function's output }

 Plots are saved to the temporal directory (tempdir) and can be accessed with the function \code{\link{gl.print.reports}} and listed with the function \code{\link{gl.list.reports}}. Note that they can be accessed only in the current R session because tempdir is cleared each time that the R session is closed.
  
 Examples of other themes that can be used can be consulted in \itemize{
 \item \url{https://ggplot2.tidyverse.org/reference/ggtheme.html} and \item
 \url{https://yutannihilation.github.io/allYourFigureAreBelongToUs/ggthemes/}
 }
}
\examples{
out <- gl.report.secondaries(bandicoot.gl)

}
\seealso{
\code{\link{gl.filter.secondaries}}, \code{\link{gl.list.reports}},
 \code{\link{gl.print.reports}},\code{\link{gl.report.heterozygosity}}

Other filters and filter reports: 
\code{\link{gl.filter.callrate}()},
\code{\link{gl.report.callrate}()},
\code{\link{gl.report.hamming}()},
\code{\link{gl.report.locmetric}()},
\code{\link{gl.report.maf}()},
\code{\link{gl.report.monomorphs}()},
\code{\link{gl.report.overshoot}()},
\code{\link{gl.report.rdepth}()},
\code{\link{gl.report.reproducibility}()},
\code{\link{gl.report.taglength}()}
}
\author{
Arthur Georges (Post to \url{https://groups.google.com/d/forum/dartr})
}
\concept{filters and filter reports}
