% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.report.secondaries.r
\name{gl.report.secondaries}
\alias{gl.report.secondaries}
\title{Report loci containing secondary SNPs in a genlight \{adegenet\} object}
\usage{
gl.report.secondaries(gl)
}
\arguments{
\item{gl}{-- name of the genlight object containing the SNP data [required]}
}
\value{
1
}
\description{
SNP datasets generated by DArT include fragments with more than one SNP (that is, with secondaries) and record them separately with the same CloneID (=AlleleID).
These multiple SNP loci within a fragment are likely to be linked, and so you may wish to remove secondaries.
This script reports duplicate loci.
}
\examples{
gl.report.secondaries(testset.gl)
}
\author{
Arthur Georges (Post to \url{https://groups.google.com/d/forum/dartr})
}
