% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.sample.r
\name{gl.sample}
\alias{gl.sample}
\title{Samples individuals from populations}
\usage{
gl.sample(x, nsample = min(table(pop(x))), replace = TRUE, verbose = NULL)
}
\arguments{
\item{x}{genlight object containing SNP/silicodart genotypes}

\item{nsample}{the number of individuals that should be sampled}

\item{replace}{a switch to sample by replacement (default).}
}
\value{
returns a genlight object with nsample samples from each populations.
}
\description{
This is a convenience function to prepare a bootstrap approach in dartR. For a bootstrap approach it is often desirable to sample a defined number of individuals for each of the populations in a genlight object and then calculate a certain quantity for that subset (redo a 1000 times)
}
\details{
This is convenience function to facilitate a bootstrap approach
}
\seealso{
Other base dartR: 
\code{\link{gl.sort}()}
}
\author{
Bernd Gruber (Post to \url{https://groups.google.com/d/forum/dartr})
}
\concept{base dartR}
