% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl2fasta.r
\name{gl2fasta}
\alias{gl2fasta}
\title{Concatenates DArT trimmed sequences and outputs a fastA file.}
\usage{
gl2fasta(gl, method = 1, outfile = "output.fasta")
}
\arguments{
\item{gl}{-- name of the DArT genlight object [required]}

\item{method}{-- 1 | 2 | 3 | 4. Type method=0 for a list of options  [method=1]}

\item{outfile}{-- name of the output file (fasta format) [output.fasta]}
}
\value{
A new gl object with all loci rendered homozygous
}
\description{
Heterozygous loci are resolved by either assigning ambiguity codes or by random allele assignment.
}
\details{
Method 1 -- heterozygous positions are replaced by
the standard ambiguity codes. The resultant sequence fragments are concatenated across loci to generate
a single combined sequence to be used in subsequent ML phylogenetic analyses.

Method=2 -- the heterozyous state is resolved by randomly assigning one or the other SNP variant
to the individual. The resultant sequence fragments are concatenated across loci to generate
a single composite haplotype to be used in subsequent ML phylogenetic analyses.

Method 3 -- heterozygous positions are replaced by
the standard ambiguity codes. The resultant SNP bases are concatenated across loci to generate
a single combined sequence to be used in subsequent MP phylogenetic analyses.

Method=4 -- the heterozyous state is resolved by randomly assigning one or the other SNP variant
to the individual. The resultant SNP bases are concatenated across loci to generate
a single composite haplotype to be used in subsequent MP phylogenetic analyses.

The script writes out the composite haplotypes for each individual as a fastA file. Requires
'TrimmedSequence' and 'SNP' (position and type of transition/transversion of a locus) to be among the locus metrics (\code{@other$loc.metrics}) headers.
}
\examples{
\dontrun{
gl2fasta(gl, method=1, outfile="test.fasta")
}
}
\author{
Bernd Gruber and Arthur Georges (glbugs@aerg.canberra.edu.au)
}

