% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl2gds.r
\name{gl2gds}
\alias{gl2gds}
\title{Convert a genlight object to nexus format PAUP SVDquartets}
\usage{
gl2gds(gl, outfile = "gl2gds.gds")
}
\arguments{
\item{gl}{-- name of the genlight object containing the SNP data [required]}

\item{outfile}{-- file name of the output file (including extension).}
}
\description{
Package SNPRelate relies on a bit-level representation of a SNP dataset that competes with \{adegenet\} genlight
objects and associated files. This function saves a genlight object to a gds format file.
}
\examples{
gl2gds(testset.gl)
}
\author{
Arthur Georges (glbugs@aerg.canberra.edu.au)
}
