% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl2gi.r
\name{gl2gi}
\alias{gl2gi}
\title{Converts a genlight object to genind object}
\usage{
gl2gi(gl, probar = TRUE)
}
\arguments{
\item{gl}{-- a genlight object}

\item{probar}{-- switch on/off progress bar}
}
\value{
A genind object, with all slots filled.
}
\description{
Converts a genlight object to genind object
}
\details{
this function uses a faster version of df2genind (from the adgegenet package)
}
\author{
Bernd Gruber (glbugs@aerg.canberra.edu.au)
}

