% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl2plink.r
\name{gl2plink}
\alias{gl2plink}
\title{Converts a genlight object to plink file format}
\usage{
gl2plink(x, outfile = "plink.csv", outpath = tempdir())
}
\arguments{
\item{x}{-- genlight object}

\item{outfile}{-- name (path) of the output plink file [default plink.csv]}

\item{outpath}{-- path of the output file. Default is to tempdir(). If to be saved in the current working directory change to "."}
}
\description{
This function exports a genlight object into plink format and save it into a file
}
\examples{
gl2plink(testset.gl)
}
\author{
Bernd Guber (glbugs@aerg.canberra.edu.au)
}
