% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl2plink.r
\name{gl2plink}
\alias{gl2plink}
\title{Converts a genlight object into PLINK format}
\usage{
gl2plink(
  x,
  plink_path = getwd(),
  bed_file = FALSE,
  outfile = "gl_plink",
  outpath = tempdir(),
  chr_format = "character",
  pos_cM = "0",
  ID_dad = "0",
  ID_mom = "0",
  sex_code = "unknown",
  phen_value = "0",
  verbose = NULL
)
}
\arguments{
\item{x}{Name of the genlight object containing the SNP data [required].}

\item{plink_path}{Path of PLINK binary file [default getwd()].}

\item{bed_file}{Whether create PLINK files .bed, .bim and .fam
[default FALSE].}

\item{outfile}{File name of the output file [default 'gl_plink'].}

\item{outpath}{Path where to save the output file
[default tempdir(), mandated by CRAN]. Use outpath=getwd() or outpath='.'
 when calling this function to direct output files to your working directory.}

\item{chr_format}{Whether chromosome information is stored as 'numeric' or as
'character', see details [default 'character'].}

\item{pos_cM}{A vector, with as many elements as there are loci, containing
the SNP position in morgans or centimorgans [default '0'].}

\item{ID_dad}{A vector, with as many elements as there are individuals,
containing the ID of the father, '0' if father isn't in dataset [default '0'].}

\item{ID_mom}{A vector, with as many elements as there are individuals,
containing the ID of the mother, '0' if mother isn't in dataset [default '0'].}

\item{sex_code}{A vector, with as many elements as there are individuals,
containing the sex code ('male', 'female', 'unknown'). Sex information needs 
just to start with an "F" or "f" for females, with an "M" or "m" for males 
and with a "U", "u" or being empty if the sex is unknown [default 'unknown'].}

\item{phen_value}{A vector, with as many elements as there are individuals,
containing the phenotype value. '1' = control, '2' = case, '0' = unknown
[default '0'].}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2,
progress log; 3, progress and results summary; 5, full report
[default 2 or as specified using gl.set.verbosity].}
}
\description{
This function exports a genlight object into PLINK format and save it into a
file.
This function produces the following PLINK files: bed, bim, fam, ped and map.
}
\details{
To create PLINK files .bed, .bim and .fam (bed_file = TRUE), it is necessary
to download the binary file of PLINK 1.9 and provide its path (plink_path).
The binary file can be downloaded from:
\url{https://www.cog-genomics.org/plink/}

The chromosome of each SNP can be a character or numeric. The chromosome
information for unmapped SNPS is coded as 0.
Family ID is taken from  x$pop.
Within-family ID (cannot be '0') is taken from indNames(x).
Variant identifier is taken from locNames(x).
SNP position is taken from the accessor x$position.
Chromosome name is taken from the accessor x$chromosome
Note that if names of populations or individuals contain spaces, they are 
replaced by an underscore "_".
}
\examples{
\donttest{
test <- platypus.gl
test$position <- test$other$loc.metrics$ChromPos_Platypus_Chrom_NCBIv1
test$chromosome <- test$other$loc.metrics$Chrom_Platypus_Chrom_NCBIv1
gl2plink(test)
}
}
\references{
Purcell, Shaun, et al. 'PLINK: a tool set for whole-genome association and
population-based linkage analyses.' The American journal of human genetics
81.3 (2007): 559-575.
}
\author{
Custodian: Luis Mijangos (Post to
 \url{https://groups.google.com/d/forum/dartr})
}
