% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl2related.r
\name{gl2related}
\alias{gl2related}
\title{Convert a genlight object to format suitable to be run with Coancestry}
\usage{
gl2related(x, outfile = "related.txt", outpath = tempdir())
}
\arguments{
\item{x}{-- name of the genlight object containing the SNP data [required]}

\item{outfile}{-- file name of the output file (including extension) [default 'related.txt']}

\item{outpath}{-- path where to save the output file [default tempdir()]}
}
\value{
a data.frame that can be used to run with the related package
}
\description{
The output txt file contains the snp data and an additional column with the names if the individual. The file then can be used and loaded into coancestry or - if installed - run with the related package. Be aware the related package was crashing in previous versions, but in general is using the same code as coancestry and therefore should have identical results. Also running coancestry with thousands of SNPs via the GUI seems to be not reliable and therefore for comparisons between coancestry and related we suggest to use the command line version of coancestry.
}
\examples{
#only a small number of loci
gtd <- gl.gl2related(bandicoot.gl[1:10,1:20])
\dontrun{
running with the related package
install.packages("related", repos="http://R-Forge.R-project.org")
library(related)
coan <- coancestry(gtd, wang=1) 
head(coan$relatedness)
#check ?coancestry for information how to use the function.
}
}
\references{
Jack Pew, Jinliang Wang, Paul Muir and Tim Frasier (2014). 
related: related: an R package for analyzing pairwise relatedness
 data based on codominant molecular markers. 
 R package version 0.8/r2.
  \url{https://R-Forge.R-project.org/projects/related/}
}
\author{
Bernd Gruber (bugs? Post to \url{https://groups.google.com/d/forum/dartr})
}
