% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.sims.r
\name{interactive_reference}
\alias{interactive_reference}
\title{Shiny app for the input of the reference table for the simulations}
\usage{
interactive_reference()
}
\description{
Shiny app for the input of the reference table for the simulations
}
\author{
Custodian: Luis Mijangos -- Post to
\url{https://groups.google.com/d/forum/dartr}
}
