% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{testset.gl}
\alias{testset.gl}
\title{A genlight object created via the read.dart functions}
\format{genlight object}
\usage{
testset.gl
}
\description{
A genlight object created via the read.dart functions
}
\author{
Arthur Georges \email{(glbugs@aerg.canberra.edu.au}
}
\keyword{datasets}

