% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{testset_metadata}
\alias{testset_metadata}
\title{Metadata file. Can be integrated via the dart2genlight function.}
\format{csv}
\description{
Metadata file. Can be integrated via the dart2genlight function.
}
\author{
Arthur Georges \email{(glbugs@aerg.canberra.edu.au}
}
\keyword{datasets}
