% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{testset_pop_recode}
\alias{testset_pop_recode}
\title{Recode file to be used with the function.}
\format{csv}
\description{
this test data set is provided to show a typical DArT file format. Can be used to create a genlight object using the read.dart function.
}
\author{
Arthur Georges \email{(glbugs@aerg.canberra.edu.au}
}
\keyword{datasets}

