% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{testset_SNPs_2Row}
\alias{testset_SNPs_2Row}
\title{Testfile in DArT format (as provided by DArT)}
\format{csv}
\description{
This test data set is provided to show a typical DArT file format. Can be used to create a genlight object using the read.dart function.
}
\author{
Arthur Georges (bugs? Post to \url{https://groups.google.com/d/forum/dartr}
}
\keyword{datasets}
