% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.basic.stats.r
\name{utils.basic.stats}
\alias{utils.basic.stats}
\title{Calculates mean observed heterozygosity, mean expected heterozygosity and Fis
 per population and Fst across all populations}
\usage{
utils.basic.stats(x)
}
\arguments{
\item{x}{A genlight object containing the SNP genotypes [required].}
}
\value{
A list with with the statistics for each population
}
\description{
Calculates mean observed heterozygosity, mean expected heterozygosity and Fis
 per population and Fst across all populations
}
\examples{
out <- utils.basic.stats(platypus.gl)
}
\author{
Luis Mijangos (bugs? Post to
\url{https://groups.google.com/d/forum/dartr})
}
