% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.check.datatype.R
\name{utils.check.datatype}
\alias{utils.check.datatype}
\title{Utility function to check that the object passed to the function is a genlight object.}
\usage{
utils.check.datatype(x, verbose = NULL)
}
\arguments{
\item{x}{Name of the genlight object containing the SNP data or tag presence/absence data (SilicoDArT) [required]}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2, progress log ; 3, progress and results summary; 5, full report [default NULL, unless specified using gl.set.verbosity]}
}
\value{
datatype, "SNP" for SNP data or  "SilicoDArT" for P/A data, or class(x)[1]
}
\description{
Most functions require access to a genlight object, and this function checks that a genlight object has been passed,
whether it is a SNP dataset or a SilicoDArT object, and reports back if verbosity is >=2
}
\details{
This function checks whether the input object is a genlight object and sets the datatype to "SNP" or "SilicoDArT" as appropriate.
}
\examples{
utils.check.datatype(testset.gl)

}
