% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.check.gl.r
\name{utils.check.gl}
\alias{utils.check.gl}
\title{A utility function to check that the object passed to the function is a genlight object.}
\usage{
utils.check.gl(x = NULL, verbose = NULL)
}
\arguments{
\item{x}{-- name of the genlight object containing the SNP data or tag presence/absence data (SilicoDArT) [required]}

\item{verbose}{-- verbosity: 0, silent or fatal errors; 1, begin and end; 2, progress log ; 3, progress and results summary; 5, full report [default 2 or as specified using gl.set.verbosity]}
}
\value{
datatype, "SNP" for SNP data or  "SilicoDArT" for P/A data
}
\description{
Most functions require access to a genlight object, and this function checks that a genlight object has been passed,
whether it is a SNP dataset or a SilicoDArT object, and reports back if verbosity is >=2
}
\examples{
utils.check.gl(testset.gs)
}
