% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.recalc.avgpic.r
\name{utils.recalc.avgpic}
\alias{utils.recalc.avgpic}
\title{A utility script to recalculate the OneRatioRef, OneRatioSnp, PICRef,	PICSnp, and	AvgPIC by locus after some 
populations have been deleted.}
\usage{
utils.recalc.avgpic(gl, v = 1)
}
\arguments{
\item{gl}{-- name of the genlight object containing the SNP data [required]}

\item{v}{-- v=0, silent; v=1, low verbosity; v=2, high verbosity [default 1]}
}
\value{
The modified genlight object
}
\description{
The locus metadata supplied by DArT has OneRatioRef, OneRatioSnp, PICRef,	PICSnp, and	AvgPIC included,
but the allelec composition will change when some individuals are removed from the dataset and so the initial statistics will
no longer apply. This script recalculates these statistics and places the recalculated values in the appropriate place 
in the genlight object.
}
\examples{
result <- dartR:::utils.recalc.avgpic(testset.gl)
}
\author{
Arthur Georges (glbugs@aerg.canberra.edu.au)
}
